### 2019-12-02 #1929 @whedon
[PRE REVIEW]: ROSS - Rotordynamic Open Source Software
**Submitting author:** @raphaeltimbo (<a href="http://orcid.org/0000-0001-7689-5486">Raphael Silva</a>)
**Repository:** <a href="https://github.com/ross-rotordynamics/ross" target ="_blank">https://github.com/ross-rotordynamics/ross</a>
**Version:** v0.2.0
