### 2019-12-02 #1929 @danielskatz
👋 @raphaeltimbo - note that whedon found 2 DOIs that should be added to your paper. Also, please check the references to make sure the cases (upper vs lower) are correct, and add {}s in the .bib file where needed to protect cases. Once you have made changes, enter a new comment in this issue with ```@whedon generate pdf``` to regenerate the pdf to make sure it looks right.
