### 2019-12-02 #1929 @danielskatz
I also see a number of spelling mistakes in the paper - please check this as well and make needed fixes.  e.g. softares, comercial, ...
