### 2019-12-03 4f36887
Correct spelling issues
