### 2019-12-03 7a51866
Remove paragraph duplicated when merging
