### 2019-12-03 51f820f
Add doi and correct cases (lower upper)
