+++ paper/paper.bib
@@ -2,29 +2,37 @@
   title={Dynamics of rotating machines},
   author={Friswell, Michael and Penny, John and Garvey, Seamus and Lees, Arthur},
   year={2010},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  url={https://doi.org/10.1017/CBO9780511780509},
+  doi={10.1017/CBO9780511780509},
 }
 
 @article{van2011numpy,
-  title={The NumPy array: a structure for efficient numerical computation},
-  author={Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael},
-  journal={Computing in Science \& Engineering},
-  volume={13},
-  number={2},
-  pages={22},
-  year={2011},
-  publisher={IEEE Computer Society}
+  author    = {St{\'{e}}fan van der Walt and
+               S. Chris Colbert and
+               Ga{\"{e}}l Varoquaux},
+  title     = {The {NumPy} Array: {A} Structure for Efficient Numerical Computation},
+  journal   = {Computing in Science and Engineering},
+  volume    = {13},
+  number    = {2},
+  pages     = {22--30},
+  year      = {2011},
+  url       = {https://doi.org/10.1109/MCSE.2011.37},
+  doi       = {10.1109/MCSE.2011.37},
+  timestamp = {Wed, 14 Nov 2018 10:48:31 +0100},
+  biburl    = {https://dblp.org/rec/bib/journals/cse/WaltCV11},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
 @article{jones2001scipy,
-  title={SciPy: Open source scientific tools for Python},
+  title={{SciPy}: Open source scientific tools for {Python}},
   author={Jones, Eric and Oliphant, Travis and Peterson, Pearu},
   year={2001},
   publisher={https://www.scipy.org}
 }
 
 @Manual{bokeh2019,
-title = {Bokeh: Python library for interactive visualization},
+title = {{Bokeh}: {Python} library for interactive visualization},
 author = {{Bokeh Development Team}},
 year = {2019},
 url = {https://bokeh.org/},
@@ -34,7 +42,8 @@ url = {https://bokeh.org/},
   title={Machinery vibration and rotordynamics},
   author={Vance, John M and Zeidan, Fouad Y and Murphy, Brian G},
   year={2010},
-  publisher={John Wiley \& Sons}
+  publisher={John Wiley \& Sons},
+  doi={10.1002/9780470903704},
 }
 
 @book{childs1993turbomachinery,
@@ -48,7 +57,8 @@ url = {https://bokeh.org/},
   title={Linear and nonlinear rotordynamics},
   author={Ishida, Yukio and Yamamoto, Toshio},
   year={2012},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/9783527651894},
 }
 
 @book{lalanne1998rotordynamics,
@@ -63,18 +73,20 @@ url = {https://bokeh.org/},
   title={Rotordynamik},
   author={Gasch, Robert and Nordmann, Rainer and Pf{\"u}tzner, Herbert},
   year={2006},
-  publisher={Springer}
+  publisher={Springer},
+  url={https://doi.org/10.1007/978-3-662-09786-1},
+  doi={10.1007/978-3-662-09786-1},
 }
 
 @misc{comsol,
-  title={COMSOL - Rotordynamics Module},
+  title={COMSOL - {Rotordynamics Module}},
   year = {2019},
   url = {https://www.comsol.com/blogs/tag/rotordynamics-module},
   note = {Acessed: 2019-12-02},
 }
 
 @misc{ansys,
[…]
