### 2019-12-03 #1929 @raphaeltimbo
@danielskatz, thanks for the comments!
I have changed the paper here ross-rotordynamics/ross#398 to fix the issues.
