### 2019-12-09 7c311e6
Assemble the 3D graphic of the bearing pressure surface.
