### 2019-12-10 02d27b2
Add boolean parameter to immediately calculate pressure matrix
