### 2019-12-10 b6ead45
Keep examples from immediately calculating pressure matrix
