### 2019-12-10 faa6402
Fix docstring
Close #400
