### 2019-12-10 f40209d
Fix docstring
