### 2019-12-10 021b643
Fix missing comma in class docstring
