### 2019-12-10 40ea6ad
Change docstring expected result to prevent error
