+++ ross/fluid_flow/fluid_flow_coefficients.py
@@ -199,7 +199,7 @@ def calculate_stiffness_matrix(fluid_flow_object, oil_film_force='numerical'):
     >>> from ross.fluid_flow.fluid_flow import fluid_flow_example
     >>> my_fluid_flow = fluid_flow_example()
     >>> calculate_stiffness_matrix(my_fluid_flow)  # doctest: +ELLIPSIS
-    [-0.003...
+    [-0.0...
     """
     [radial_force, tangential_force, force_x, force_y] = \
         calculate_oil_film_force(fluid_flow_object, force_type=oil_film_force)
