### 2019-12-11 338d67f
Timoshenko "Kappa" factor automated selection. Addition extra Kappa formulation based on Hutchinson2001
