### 2019-12-11 a38fbe1
Shaft "M" matrix update to 6gdl
