### 2019-12-12 dca479b
Correct boundary condition
One of the boundary conditions for calculating the pressure matrix\
was being calculated with the wrong parameter.\
This committ corrects that.
