### 2019-12-13 a394179
Remove ShaftElement and replace with the ShaftTaperedElement
