+++ ross/shaft_element.py
@@ -12,8 +12,7 @@ from ross.element import Element
 from ross.materials import Material, steel
 from ross.utils import read_table_file
 
-__all__ = ["ShaftElement", "ShaftTaperedElement"]
-
+__all__ = ["ShaftElement"]
 bokeh_colors = bp.RdGy[11]
 
 
@@ -34,881 +33,16 @@ class ShaftElement(Element):
     ----------
     L : float
         Element length.
-    i_d : float
-        Inner diameter of the element.
-    o_d : float
-        Outer diameter of the element.
-    material : ross.material
-        Shaft material.
-    n : int, optional
-        Element number (coincident with it's first node).
-        If not given, it will be set when the rotor is assembled
-        according to the element's position in the list supplied to
-        the rotor constructor.
-    axial_force : float, optional
-        Axial force.
-    torque : float, optional
-        Torque.
-    shear_effects : bool, optional
-        Determine if shear effects are taken into account.
-        Default is True.
-    rotary_inertia : bool, optional
-        Determine if rotary_inertia effects are taken into account.
-        Default is True.
-    gyroscopic : bool, optional
-        Determine if gyroscopic effects are taken into account.
-        Default is True.
-    shear_method_calc : string, optional
-        Determines which shear calculation method the user will adopt
-        Default is 'cowper'
-
-    Returns
-    -------
-
-    Attributes
-    ----------
-    Poisson : float
-        Poisson coefficient for the element.
-    A : float
-        Element section area.
-    Ie : float
-        Ie is the second moment of area of the cross section about
-        the neutral plane :math:`Ie = \pi r^2/4`
-    phi : float
-        Constant that is used according to :cite:`friswell2010dynamics` to
-        consider rotary inertia and shear effects. If these are not considered
-        :math:`\phi=0`.
-
-    References
-    ----------
-
-    .. bibliography:: ../../../docs/refs.bib
-
-    Examples
-    --------
-    >>> from ross.materials import steel
-    >>> le = 0.25
-    >>> i_d = 0
-    >>> o_d = 0.05
-    >>> Euler_Bernoulli_Element = ShaftElement(le, i_d, o_d, steel,
-    ...                                        shear_effects=False, rotary_inertia=False)
-    >>> Euler_Bernoulli_Element.phi
-    0
-    >>> Timoshenko_Element = ShaftElement(le, i_d, o_d, steel,
-    ...                                   rotary_inertia=True,
-    ...                                   shear_effects=True)
-    >>> Timoshenko_Element.phi
-    0.08795566502463055
-    """
-
-    def __init__(
-        self,
-        L,
-        i_d,
-        o_d,
-        material,
-        n=None,
-        axial_force=0,
-        torque=0,
-        shear_effects=True,
-        rotary_inertia=True,
-        gyroscopic=True,
-        shear_method_calc="cowper",
-        tag=None,
-    ):
-
-        if type(material) is str:
-            os.chdir(Path(os.path.dirname(ross.__file__)))
-            self.material = Material.use_material(material)
-        else:
-            self.material = material
-
-        self.shear_effects = shear_effects
-        self.rotary_inertia = rotary_inertia
-        self.gyroscopic = gyroscopic
-        self.axial_force = axial_force
-        self.torque = torque
-        self._n = n
-        self.n_l = n
-        self.n_r = None
-        if n is not None:
-            self.n_r = n + 1
-
-        self.shear_method_calc = shear_method_calc
-        self.tag = tag
-
-        self.L = float(L)
-
-        # diameters
-        self.i_d = float(i_d)
-        self.o_d = float(o_d)
-        self.i_d_l = float(i_d)
[…]
