### 2019-12-13 2f34de2
Remove ShaftTaperedElement tests
* Remove tests for ShaftTaperedElement deprecated class
* Convert ShaftElement tests according new API.
