+++ ross/tests/test_shaft_element.py
@@ -5,7 +5,7 @@ import pytest
 from numpy.testing import assert_almost_equal, assert_allclose
 
 from ross.materials import steel
-from ross.shaft_element import ShaftElement, ShaftTaperedElement
+from ross.shaft_element import ShaftElement
 
 test_dir = os.path.dirname(__file__)
 
@@ -14,10 +14,12 @@ test_dir = os.path.dirname(__file__)
 def eb():
     #  Euler-Bernoulli element
     le_ = 0.25
-    i_d_ = 0
-    o_d_ = 0.05
+    i_d_l = 0
+    o_d_l = 0.05
+    i_d_r = 0
+    o_d_r = 0.05
     return ShaftElement(
-        le_, i_d_, o_d_, steel, shear_effects=False, rotary_inertia=False, n=3
+        le_, i_d_l, o_d_l, i_d_r, o_d_r, steel, shear_effects=False, rotary_inertia=False, n=3
     )
 
 
@@ -35,8 +37,10 @@ def test_index(eb):
 def test_parameters_eb(eb):
     assert eb.phi == 0
     assert eb.L == 0.25
-    assert eb.i_d == 0
-    assert eb.o_d == 0.05
+    assert eb.i_d_l == 0
+    assert eb.o_d_l == 0.05
+    assert eb.i_d_r == 0
+    assert eb.o_d_r == 0.05
     assert eb.material.E == 211e9
     assert eb.material.G_s == 81.2e9
     assert eb.material.rho == 7810
@@ -78,9 +82,11 @@ def tim():
     #  Timoshenko element
     z_ = 0
     le_ = 0.25
-    i_d_ = 0
-    o_d_ = 0.05
-    return ShaftElement(le_, i_d_, o_d_, steel, rotary_inertia=True, shear_effects=True)
+    i_d_l = 0
+    o_d_l = 0.05
+    i_d_r = 0
+    o_d_r = 0.05
+    return ShaftElement(le_, i_d_l, o_d_l, i_d_r, o_d_r, steel, rotary_inertia=True, shear_effects=True)
 
 
 def test_parameters_tim(tim):
@@ -167,7 +173,7 @@ def tap_tim():
     o_d_l = 0.25
     o_d_r = 0.10
 
-    return ShaftTaperedElement(
+    return ShaftElement(
         steel,
         L,
         i_d_l,
@@ -189,7 +195,6 @@ def test_tapered_index(tap_tim):
     assert tap_tim.dof_local_index().y_1 == 5
     assert tap_tim.dof_local_index().alpha_1 == 6
     assert tap_tim.dof_local_index().beta_1 == 7
-    print(tap_tim.dof_global_index())
     assert tap_tim.dof_global_index().x_3 == 12
     assert tap_tim.dof_global_index().y_3 == 13
     assert tap_tim.dof_global_index().alpha_3 == 14
@@ -271,7 +276,7 @@ def tap2():
     o_d_l = 0.25
     o_d_r = 0.25
 
-    return ShaftTaperedElement(
+    return ShaftElement(
         steel, L, i_d_l, o_d_l, i_d_r, o_d_r, shear_effects=True, rotary_inertia=True
     )
 
@@ -282,7 +287,7 @@ def tim2():
     le_ = 0.4
     i_d_ = 0
     o_d_ = 0.25
-    return ShaftElement(le_, i_d_, o_d_, steel, rotary_inertia=True, shear_effects=True)
+    return ShaftElement(le_, i_d_, o_d_, material=steel, rotary_inertia=True, shear_effects=True)
 
 
 def test_match_mass_matrix(tap2, tim2):
