### 2019-12-13 80226dd
Adapt test_rotor_assembly.py to ShaftElement new API
