### 2019-12-13 67dcb66
Adapt rotor_assembly.py to new ShaftElement API
