+++ ross/rotor_assembly.py
@@ -111,11 +111,13 @@ class Rotor(object):
     >>> le = 0.25
     >>> i_d = 0
     >>> o_d = 0.05
-    >>> tim0 = rs.ShaftElement(le, i_d, o_d, steel,
+    >>> tim0 = rs.ShaftElement(le, i_d, o_d,
+    ...                        material=steel,
     ...                        shear_effects=True,
     ...                        rotary_inertia=True,
     ...                        gyroscopic=True)
-    >>> tim1 = rs.ShaftElement(le, i_d, o_d, steel,
+    >>> tim1 = rs.ShaftElement(le, i_d, o_d,
+    ...                        material=steel,
     ...                        shear_effects=True,
     ...                        rotary_inertia=True,
     ...                        gyroscopic=True)
@@ -239,8 +241,8 @@ class Rotor(object):
             "y_pos",
             "i_d",
             "o_d",
-            "i_d_r",
-            "o_d_r",
+            "idr",
+            "odr",
             "material",
             "rho",
             "volume",
@@ -414,7 +416,7 @@ class Rotor(object):
             if z_pos == df_shaft["nodes_pos_l"].iloc[0]:
                 y_pos = (
                     max(
-                        df_shaft["o_d_l"][
+                        df_shaft["odl"][
                             df_shaft.n_l == int(dfb_z_pos.iloc[0]["n_l"])
                         ].values
                     )
@@ -423,7 +425,7 @@ class Rotor(object):
             elif z_pos == df_shaft["nodes_pos_r"].iloc[-1]:
                 y_pos = (
                     max(
-                        df_shaft["o_d_r"][
+                        df_shaft["odr"][
                             df_shaft.n_r == int(dfb_z_pos.iloc[0]["n_r"])
                         ].values
                     )
@@ -434,12 +436,12 @@ class Rotor(object):
                     max(
                         [
                             max(
-                                df_shaft["o_d_l"][
+                                df_shaft["odl"][
                                     df_shaft._n == int(dfb_z_pos.iloc[0]["n_l"])
                                 ].values
                             ),
                             max(
-                                df_shaft["o_d_r"][
+                                df_shaft["odr"][
                                     df_shaft._n == int(dfb_z_pos.iloc[0]["n_l"]) - 1
                                 ].values
                             ),
@@ -569,7 +571,7 @@ class Rotor(object):
         >>> n = 6
         >>> L = [0.25 for _ in range(n)]
         ...
-        >>> shaft_elem = [rs.ShaftElement(l, i_d, o_d, steel,
+        >>> shaft_elem = [rs.ShaftElement(l, i_d, o_d, material=steel,
         ... shear_effects=True, rotary_inertia=True, gyroscopic=True) for l in L]
         >>> disk0 = DiskElement.from_geometry(2, steel, 0.07, 0.05, 0.28)
         >>> disk1 = DiskElement.from_geometry(4, steel, 0.07, 0.05, 0.35)
@@ -590,7 +592,7 @@ class Rotor(object):
         eigv_arr = np.append(eigv_arr, modal.wn[n_eigval])
 
         # this value is up to start the loop while
-        error = 1
+        error = 1.0e10
         nel_r = 2
 
         while error > err_max:
@@ -600,25 +602,25 @@ class Rotor(object):
 
             for shaft in self.shaft_elements:
                 le = shaft.L / nel_r
-                odl = shaft.o_d_l
-                odr = shaft.o_d_r
-                idl = shaft.i_d_l
-                idr = shaft.i_d_r
+                odl = shaft.odl
+                odr = shaft.odr
+                idl = shaft.idl
+                idr = shaft.idr
 
[…]
