### 2019-12-13 0a02377
Remove ShaftTaperedElement from imports
