### 2019-12-16 fd13fa0
Fix syntax for some pytest.fixtures
* tap_tim() - correct attribute position.
* tap2() - - correct attribute position.
