+++ ross/tests/test_shaft_element.py
@@ -174,12 +174,12 @@ def tap_tim():
     o_d_r = 0.10
 
     return ShaftElement(
-        steel,
         L,
         i_d_l,
         o_d_l,
         i_d_r,
         o_d_r,
+        steel,
         shear_effects=True,
         rotary_inertia=True,
         n=3,
@@ -277,7 +277,7 @@ def tap2():
     o_d_r = 0.25
 
     return ShaftElement(
-        steel, L, i_d_l, o_d_l, i_d_r, o_d_r, shear_effects=True, rotary_inertia=True
+        L, i_d_l, o_d_l, i_d_r, o_d_r, steel, shear_effects=True, rotary_inertia=True
     )
 
 
