### 2019-12-16 6a557a8
Fix rotor.from_section() docsting
