### 2019-12-16 97f9de9
Fix .test_parameters_eb() test sintax
