### 2019-12-16 8d18ebd
Fix typo in docstring
