### 2019-12-16 f74f28a
Fix .test_parameters_tap_tim() attributes sintax
