### 2019-12-16 1f97a44
Adapt test_from_section() test to new API
