### 2019-12-16 468165e
Gerenal updates to Rotor.from_section()
* include material as attribute - single material for all shaft or for each region.
* adapt the method to accept tapered elements.
* include new error messages to guide the user.
