+++ ross/rotor_assembly.py
@@ -2322,8 +2322,11 @@ class Rotor(object):
     def from_section(
         cls,
         leng_data,
-        o_ds_data,
-        i_ds_data,
+        idl_data,
+        odl_data,
+        idr_data=None,
+        odr_data=None,
+        material_data=None,
         disk_data=None,
         brg_seal_data=None,
         sparse=True,
@@ -2332,6 +2335,7 @@ class Rotor(object):
         rated_w=None,
         n_eigen=12,
         nel_r=1,
+        tag=None,
     ):
         """This class is an alternative to build rotors from separated
         sections. Each section has the same number (n) of shaft elements.
@@ -2340,10 +2344,19 @@ class Rotor(object):
         ----------
         leng_data : list
             List with the lengths of rotor regions.
-        o_d_data : list
-            List with the outer diameters of rotor regions.
-        i_d_data : list
-            List with the inner diameters of rotor regions.
+        idl_data : list
+            List with the inner diameters of rotor regions (Left Station).
+        odl_data : list
+            List with the outer diameters of rotor regions (Left Station).
+        idr_data : list, optional
+            List with the inner diameters of rotor regions (Right Station).
+            Default is equal to idl_data (cylindrical element).
+        odr_data : list, optional
+            List with the outer diameters of rotor regions (Right Station).
+            Default is equal to odl_data (cylindrical element).
+        material_data : ross.material or list of ross.material
+            Defines a single material for all sections or each section can be
+            defined by a material individually.
         disk_data : dict, optional
             Dict holding disks datas.
             Example : disk_data=DiskElement.from_geometry(n=2,
@@ -2359,14 +2372,14 @@ class Rotor(object):
                                                    kyy=1e6, cyy=0, kxy=0,
                                                    cxy=0, kyx=0, cyx=0)
             ***See 'bearing_seal_element.py' docstring for more information***
-        w : float, optional
-            Rotor speed.
         nel_r : int, optional
             Number or elements per shaft region.
             Default is 1.
         n_eigen : int, optional
             Number of eigenvalues calculated by arpack.
             Default is 12.
+        tag : str
+            A tag for the rotor
 
         Returns
         -------
@@ -2374,10 +2387,11 @@ class Rotor(object):
 
         Example
         -------
-
+        >>> from ross.materials import steel
         >>> rotor = Rotor.from_section(leng_data=[0.5,0.5,0.5],
-        ...             o_ds_data=[0.05,0.05,0.05],
-        ...             i_ds_data=[0,0,0],
+        ...             odl_data=[0.05,0.05,0.05],
+        ...             idl_data=[0,0,0],
+        ...             material_data=steel,
         ...             disk_data=[DiskElement.from_geometry(n=1, material=steel, width=0.07, i_d=0, o_d=0.28),
         ...                        DiskElement.from_geometry(n=2, material=steel, width=0.07, i_d=0, o_d=0.35)],
         ...             brg_seal_data=[BearingElement(n=0, kxx=1e6, cxx=0, kyy=1e6, cyy=0, kxy=0, cxy=0, kyx=0, cyx=0),
@@ -2387,8 +2401,20 @@ class Rotor(object):
         >>> modal.wn.round(4)
         array([ 85.7634,  85.7634, 271.9326, 271.9326, 718.58  , 718.58  ])
         """
-        if len(leng_data) != len(o_ds_data) or len(leng_data) != len(i_ds_data):
-            raise ValueError("The matrices lenght do not match.")
+
+        if len(leng_data) != len(odl_data) or len(leng_data) != len(idl_data):
+            raise ValueError("The lists size do not match (leng_data, odl_data and idl_data).")
+
+        if material_data is None:
[…]
