### 2019-12-16 e26c04d
Update tests for Rotor.from_section()
* Adapt tests for from_section() new API
* Add more assertions covering the error messages
