### 2019-12-16 d8c971e
Allow Rotor.from_section() to create tapered elements
