### 2019-12-16 c34381a
Fix typo in test_from_section()
