### 2019-12-16 97666aa
replace pytest.raises(ValueError) by pytest.raises(AttributeError)
* test - test_from_section() - replace pytest.raises(ValueError) by pytest.raises(AttributeError)
for better representation.
