### 2019-12-16 bd06b6d
Using IndexError in the last test from test_from_section()
* To match the error raised in Rotor.from_section()
