### 2019-12-16 5c9e2a1
Fix referenced long bearing
