### 2019-12-18 212e052
Add repetition for Disks and Bearing "n_l" and "n_r"
* In the loop that copies disk and bearing elements, it's necessary to change de "n_l" and "n_r" attributes.
The lack of these lines was inducing a bug while plotting the rotor when instantiating Rotor.from_section().
