### 2019-12-18 4c8ce3b
Remove .dof_global_index() from element.py
* We were having problems to define de global indexes for the elements using this method intenally to a element class.
* It makes more sense to use .dof_global_index() when build the rotor model.
