### 2019-12-18 904701b
Add routine to calc. dof global indexes
* The routine to calculate the elements' global indexes have been moved to rotor_assembly (.__init__() method).
* Now, each element has an attribute that keeps those index, instead of calling a function to this job.
