### 2019-12-18 f49eff4
Add dof_global_index attribute to bearing_seal_element.py
* This attribute will store information about the global index of the element.
* Add n_link info to .__repr__().
