### 2019-12-18 3c84b4d
Add dof_global_index attribute to point_mass.py
* This attribute will store information about the global index of the element.
