### 2019-12-18 d18f27c
Replace tests for n_link and global index
* Replace test_rotor_link() by test_global_index().
* This new test has a greater coverage, considering 2 point masses (2 n_links) and assert the global index for every type
of element presented in the model.
