### 2019-12-18 97b96b7
Adapt run_static() to dof_global_index attribute
