### 2019-12-19 81e3989
Set masses as float as default
Without it, `.__repr__()` returns an error when defining the masses as integers.
