### 2020-01-08 #1929 @raphaeltimbo
If you need suggestions, I have filtered the joss-reviewers list for python and mechanical engineering:
- kyleniemeyer
- moorepants
- trallard
- nnadeau
- ctdegroot
- srmnitc
