### 2020-01-13 #1929 @meg-simula
Hi @raphaeltimbo, thank you very much for this submission and the suggested reviewers - my apologies for the delay in editorial handling. Before we proceed, could I ask you to briefly elaborate on the (potential) use of ROSS as *research* software?
