### 2020-01-14 #1929 @raphaeltimbo
Hi @meg-simula!
I have some examples where we can use models that ROSS generates in research:
- Evaluate the impact that components such as bearings and seals have on the rotor. Particularly, we have some studies on how the uncertainties on the dynamic coefficients for these components can impact the rotor dynamic;
- Generate data to test learning algorithms that should diagnose machine failure. We can, for example, use ROSS to generate a rotor with some problem such as an unbalance or misalignment, and then use the output data from this model to test the learning algorithm and see how good it is in diagnosing the problem;
- We also have an interest in model updating. Here we could use vibrations measurements from a rotor to update a model generated by ROSS.

Let me know if we need to update the paper and also if you need more clarification.

Thanks!
