### 2020-01-14 #1929 @meg-simula
@moorepants, @trallard, @nnadeau, @ctdegroot, @srmnitc: Would either of you be willing and available to review this submission for JOSS?
