### 2020-01-14 #1929 @srmnitc
@meg-simula I could review this submission for JOSS. However, I will be away for couple of days, but I will be able to start the review next week.
