### 2020-01-18 e6b5480
Group "n" and "n_link" attributes for the .__repr__
