### 2020-01-18 6a901c5
Fix doctests based on BearingElement __repr__
