### 2020-01-21 4de0c5b
Adds a preview to how API_report pdf file will look like.
