### 2020-01-21 bba4c8d
Adds a preview to how API_report pdf file will look like.
