### 2020-01-21 1403797
Shaft "K" Matrix adaptation to 6 gdls
