+++ ross/shaft_element.py
@@ -547,23 +547,85 @@ class ShaftElement(Element):
                [ 0.        , -5.71205534,  0.97294287,  0.        ],
                [ 5.71205534,  0.        ,  0.        ,  0.97294287]])
         """
-        phi = self.phi
+
+        # temporary material and geometrical constants
         L = self.L
-        # fmt: off
-        K = np.array([
-            [12,     0,            0,          6*L,  -12,     0,            0,          6*L],
-            [0,     12,         -6*L,            0,    0,   -12,         -6*L,            0],
-            [0,   -6*L, (4+phi)*L**2,            0,    0,   6*L, (2-phi)*L**2,            0],
-            [6*L,    0,            0, (4+phi)*L**2, -6*L,     0,            0, (2-phi)*L**2],
-            [-12,    0,            0,         -6*L,   12,     0,            0,         -6*L],
-            [0,    -12,          6*L,            0,    0,    12,          6*L,            0],
-            [0,   -6*L, (2-phi)*L**2,            0,    0,   6*L, (4+phi)*L**2,            0],
-            [6*L,    0,            0, (2-phi)*L**2, -6*L,     0,            0, (4+phi)*L**2]
-        ])
-        # fmt: on
-        K = self.material.E * self.Ie * K / ((1 + phi) * L ** 3)
+        tempG = self.material.E / (2 * (1 + self.n))
+        tempS = np.pi * ((self.o_d / 2) ** 2 - (self.i_d / 2) ** 2)
+        tempI = np.pi / 4 * ((self.o_d / 2) ** 4 - (self.i_d / 2) ** 4)
+        tempJ = np.pi / 2 * ((self.o_d / 2) ** 4 - (self.i_d / 2) ** 4)
+        tempMM = (self.i_d / 2) / (self.o_d / 2)
+
+        # temporary variables dependent on kappa
+        tempA = (
+            12
+            * self.material.E
+            * tempI
+            / (self.material.G_s * self.kappa * tempS * L ** 2)
+        )
+
+        # element level matrix declaration
+
+        aux1 = self.material.E * tempI / ((1 + tempA) * L**3)
+        aux2 = self.material.G * tempJ / L
+        aux3 = self.material.E * tempS / L
+
+        KcEst = aux1*np.array([12   ,0          ,0   ,0         ,0          ,-6*L      ,-12 ,0          ,0    ,0         ,0          ,-6*L     ]  
+            [0    ,aux3/aux1  ,0   ,0         ,0          ,0         ,0   ,-aux3/aux1 ,0    ,0         ,0          ,0        ]
+            [0    ,0          ,12  ,6*L       ,0          ,0         ,0   ,0          ,-12  ,6*L       ,0          ,0        ]  
[…]
