### 2020-01-21 25fa000
Shaft "G" Matrix adaptation to 6 gdls
