### 2020-01-22 cabc074
Replace checking type by "isinstance()" when setting global indexes
