### 2020-01-30 91f8708
Shorter code to global indexes and add them to the DataFrame
