+++ ross/rotor_assembly.py
@@ -380,24 +380,19 @@ class Rotor(object):
             if elm.n <= n_last + 1:
                 for k, v in global_dof_mapping.items():
                     global_dof_mapping[k] = 4 * elm.n + v
+                if hasattr(elm, "n_link") and elm.n_link is not None:
+                    global_dof_mapping[f"x_{elm.n_link}"] = 4 * elm.n_link
+                    global_dof_mapping[f"y_{elm.n_link}"] = 4 * elm.n_link + 1
             else:
                 for k, v in global_dof_mapping.items():
                     global_dof_mapping[k] = 2 * n_last + 2 * elm.n + 4 + v
+                if hasattr(elm, "n_link") and elm.n_link is not None:
+                    global_dof_mapping[f"x_{elm.n_link}"] = 2 * n_last + 2 * elm.n_link + 4 
+                    global_dof_mapping[f"y_{elm.n_link}"] = 2 * n_last + 2 * elm.n_link + 5
 
+            dof_tuple = namedtuple("GlobalIndex", global_dof_mapping)
             elm.dof_global_index = dof_tuple(**global_dof_mapping)
-
-            if isinstance(elm, BearingElement):
-                if elm.n_link is not None:
-                    dof_global_index = elm.dof_global_index._asdict()
-                    if elm.n_link <= n_last + 1:
-                        dof_global_index[f"x_{elm.n_link}"] = 4 * elm.n_link
-                        dof_global_index[f"y_{elm.n_link}"] = 4 * elm.n_link + 1
-                    else:
-                        dof_global_index[f"x_{elm.n_link}"] = 2 * n_last + 2 * elm.n_link + 4 
-                        dof_global_index[f"y_{elm.n_link}"] = 2 * n_last + 2 * elm.n_link + 5
-
-                    dof_tuple = namedtuple("GlobalIndex", dof_global_index)
-                    elm.dof_global_index = dof_tuple(**dof_global_index)
+            df.at[df.loc[df.tag == elm.tag].index[0], "dof_global_index"] = elm.dof_global_index
 
         #  values for static analysis will be calculated by def static
         self.Vx = None
