### 2020-01-30 7f3735b
shorter code to global indexes - check n_link attr.
There was an error in the last commit that the attr. n_link was not being checked.
This should fix this problem.
