### 2020-02-04 8b83665
Updating the code to new changes and fixing doctests.
