+++ ross/rotor_assembly.py
@@ -20,6 +20,7 @@ from bokeh.models import ColumnDataSource
 from bokeh.models.glyphs import Text
 from bokeh.plotting import figure, output_file
 from cycler import cycler
+from ross.utils import convert
 
 import ross
 from ross.bearing_seal_element import BearingElement
@@ -74,7 +75,7 @@ class Rotor(object):
         List with the shaft elements
     disk_elements : list
         List with the disk elements
-    bearing_seal_elements : list
+    bearing_elements : list
         List with the bearing elements
     point_mass_elements: list
         List with the point mass elements
@@ -136,7 +137,7 @@ class Rotor(object):
         self,
         shaft_elements,
         disk_elements=None,
-        bearing_seal_elements=None,
+        bearing_elements=None,
         point_mass_elements=None,
         sparse=True,
         n_eigen=12,
@@ -191,8 +192,8 @@ class Rotor(object):
 
         if disk_elements is None:
             disk_elements = []
-        if bearing_seal_elements is None:
-            bearing_seal_elements = []
+        if bearing_elements is None:
+            bearing_elements = []
         if point_mass_elements is None:
             point_mass_elements = []
 
@@ -200,7 +201,7 @@ class Rotor(object):
             if disk.tag is None:
                 disk.tag = "Disk " + str(i)
 
-        for i, brg in enumerate(bearing_seal_elements):
+        for i, brg in enumerate(bearing_elements):
             if brg.__class__.__name__ == "BearingElement" and brg.tag is None:
                 brg.tag = "Bearing " + str(i)
             if brg.__class__.__name__ == "SealElement" and brg.tag is None:
@@ -211,7 +212,7 @@ class Rotor(object):
                 p_mass.tag = "Point Mass " + str(i)
 
         self.shaft_elements = sorted(shaft_elements, key=lambda el: el.n)
-        self.bearing_seal_elements = bearing_seal_elements
+        self.bearing_elements = bearing_elements
         self.disk_elements = disk_elements
         self.point_mass_elements = point_mass_elements
         self.elements = [
@@ -220,7 +221,7 @@ class Rotor(object):
                 [
                     self.shaft_elements,
                     self.disk_elements,
-                    self.bearing_seal_elements,
+                    self.bearing_elements,
                     self.point_mass_elements,
                 ]
             )
@@ -255,14 +256,14 @@ class Rotor(object):
         df_bearings = pd.DataFrame(
             [
                 el.summary()
-                for el in self.bearing_seal_elements
+                for el in self.bearing_elements
                 if (el.__class__.__name__ == "BearingElement")
             ]
         )
         df_seals = pd.DataFrame(
             [
                 el.summary()
-                for el in self.bearing_seal_elements
+                for el in self.bearing_elements
                 if (el.__class__.__name__ == "SealElement")
             ]
         )
@@ -382,11 +383,11 @@ class Rotor(object):
         # check if there are bearings without location
         bearings_no_zloc = {
             b
-            for b in bearing_seal_elements
+            for b in bearing_elements
             if pd.isna(df.loc[df.tag == b.tag, "nodes_pos_l"]).all()
         }
         # cycle while there are bearings without a z location
-        for b in cycle(bearing_seal_elements):
+        for b in cycle(bearing_elements):
             if bearings_no_zloc:
                 if b in bearings_no_zloc:
                     # first check if b.n is on list, if not, check for n_link
@@ -632,7 +633,7 @@ class Rotor(object):
[…]
