### 2020-02-04 5a1a2fc
Changing docttest trying to make it work in travis-ci and appveyor.
