### 2020-02-04 698ad0c
Changing docttest trying to make it work in travis-ci and appveyor.
