### 2020-02-04 4a41a5f
Changing docttest trying to make it work in travis-ci and appveyor.
