### 2020-02-04 562c6b3
Removing os.chdir() method from save_material method.
