### 2020-02-04 bf2d3c5
Changing load_data() methods to get_data() for better understanding.
