### 2020-02-04 76301dd
Minor fixes to materials __repr__ and __eq__ method.
