### 2020-02-04 ebd184c
Removing rotors folder
