+++ ross/rotors/Rotor_format/elements/disk_elements.toml
@@ -1,5 +1,0 @@
-[disk_element.48]
-n = 48
-m = 51.5252611115262
-Id = 0.4235805840543383
-Ip = 0.8050822048675969
