+++ ross/rotors/Benchmarks/elements/BearingElement.toml
@@ -1,3189 +1,0 @@
-[BearingElement.0]
-n = 0
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.1]
-n = 1
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.2]
-n = 2
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.3]
-n = 3
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.4]
-n = 4
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.5]
-n = 5
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.6]
-n = 6
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.7]
-n = 7
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.8]
-n = 8
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.9]
-n = 9
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.10]
-n = 10
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.11]
-n = 11
-kxx = [ 123,]
-cxx = [ 1233,]
-kyy = [ 123,]
-kxy = [ 123,]
-kyx = [ 123123,]
-cyy = [ 1231,]
-cxy = [ 123123,]
-cyx = [ 123123,]
-
-[BearingElement.12]
[…]
