### 2020-02-04 853e3a4
Changing remove method
