### 2020-02-04 1a5ac5f
Removing available_rotors() method.
