### 2020-02-04 c66bc40
Fixing doctests for new code format.
