+++ ross/materials.py
@@ -70,7 +70,28 @@ class Material:
             self.Poisson = (self.E / (2 * self.G_s)) - 1
 
     def __eq__(self, other):
-        """Material is considered equal if properties are equal."""
+        """Function used to compare two Materials.
+        Parameters
+        ----------
+
+        self : Material
+
+        other: Material
+
+        Returns
+        ----------
+
+        bool
+            True if all the Materials properties are equivalent.
+
+        Examples
+        ----------
+        >>> import ross as rs
+        >>> steel = rs.steel
+        >>> AISI4140 = rs.Material.use_material('AISI4140')
+        >>> steel == AISI4140
+        False
+        """
         self_list = [v for v in self.__dict__.values() if isinstance(v, (float, int))]
         other_list = [v for v in other.__dict__.values() if isinstance(v, (float, int))]
 
@@ -80,6 +101,25 @@ class Material:
             return False
 
     def __repr__(self):
+        """Function used to give a representation of a Material element, when called.
+
+        Parameters
+        ----------
+
+        self : Material
+
+        Returns
+        ----------
+
+        string : Representation of the given Material.
+
+        Examples
+        ----------
+        >>> import ross as rs
+        >>> steel = rs.steel
+        >>> steel # doctest: +ELLIPSIS
+        Material(name="Steel", rho=7.810e+03, G_s=8.120e+10, E=2.110e+11, Poisson=2.993e-01, color='#525252')
+        """
         selfE = "{:.5e}".format(self.E)
         selfPoisson = "{:.5e}".format(self.Poisson)
         selfrho = "{:.5e}".format(self.rho)
@@ -92,6 +132,30 @@ class Material:
         )
 
     def __str__(self):
+        """Function used to set what is shown when a Material is printed.
+
+        Parameters
+        ----------
+
+        self : Material
+
+        Returns
+        ----------
+
+        str
+            Containing all the Materials properties organized in a table.
+
+        Examples
+        ----------
+        >>> import ross as rs
+        >>> print(rs.steel)
+        Steel
+        -----------------------------------
+        Density         (N/m**3): 7810.0
+        Young`s modulus (N/m**2): 2.11e+11
+        Shear modulus   (N/m**2): 8.12e+10
+        Poisson coefficient     : 0.29926108
+        """
         return (
             f"{self.name}"
             f'\n{35*"-"}'
@@ -103,11 +167,34 @@ class Material:
 
     @staticmethod
     def dump_data(data):
+        """Auxiliary function to save the materials properties in the save method.
+
+        Parameters
+        ----------
+
+        data : dict
+            Dictionary containing all data needed to instantiate the Object.
+
+        Returns
+        ----------
+        """
         with open(Path(os.path.dirname(rs.__file__))/"available_materials.toml", "w") as f:
             toml.dump(data, f)
 
     @staticmethod
     def get_data():
+        """Auxiliary function to load all saved materials properties in the use_material method.
+
+        Parameters
+        ----------
+
+        Returns
+        ----------
+
+        data : dict
+            Containing all data needed to instantiate a Material Object.
+
+        """
         try:
             with open(Path(os.path.dirname(rs.__file__))/"available_materials.toml", "r") as f:
                 data = toml.load(f)
@@ -118,6 +205,27 @@ class Material:
 
     @staticmethod
     def use_material(name):
+        """Function to load the materials properties and instantiate a Material Object.
+
+        Parameters
+        ----------
+
+        name : string
+            Material's name.
+
+        Returns
+        ----------
+
+        Material : Material Object
+
+        Examples
+        ----------
+        >>> import ross as rs
+        >>> AISI4140 = rs.Material.use_material('AISI4140')
+        >>> AISI4140
[…]
