### 2020-02-04 fb148aa
substituting bearing_seal_elements for bearing_elements
