### 2020-02-04 5dc175a
Updating materials documentation
