### 2020-02-10 599b280
Organizes the function that calculates the stiffness matrix
