### 2020-02-10 28b0940
Organizes the function that calculates the stiffness matrix
