+++ ross/fluid_flow/fluid_flow_coefficients.py
@@ -4,50 +4,6 @@ from math import isnan
 from ross.fluid_flow.fluid_flow_geometry import move_rotor_center
 
 
-def calculate_analytical_stiffness_matrix(load, eccentricity_ratio, radial_clearance):
-    """Returns the stiffness matrix calculated analytically.
-    Suitable only for short bearings.
-    Parameters
-    -------
-    load: float
-        Load applied to the rotor (N).
-    eccentricity_ratio: float
-        The ratio between the journal displacement, called just eccentricity, and
-        the radial clearance.
-    radial_clearance: float
-        Difference between both stator and rotor radius, regardless of eccentricity.
-    Returns
-    -------
-    list of floats
-        A list of length four including stiffness floats in this order: kxx, kxy, kyx, kyy
-    Examples
-    --------
-    >>> from ross.fluid_flow.fluid_flow import fluid_flow_example
-    >>> my_fluid_flow = fluid_flow_example()
-    >>> load = my_fluid_flow.load
-    >>> eccentricity_ratio = my_fluid_flow.eccentricity_ratio
-    >>> radial_clearance = my_fluid_flow.radial_clearance
-    >>> calculate_analytical_stiffness_matrix(load, eccentricity_ratio, radial_clearance) # doctest: +ELLIPSIS
-    [...
-    """
-    # fmt: off
-    h0 = 1.0 / (((np.pi ** 2) * (1 - eccentricity_ratio ** 2) + 16 * eccentricity_ratio ** 2) ** 1.5)
-    a = load / radial_clearance
-    kxx = a * h0 * 4 * ((np.pi ** 2) * (2 - eccentricity_ratio ** 2) + 16 * eccentricity_ratio ** 2)
-    kxy = (a * h0 * np.pi * ((np.pi ** 2) * (1 - eccentricity_ratio ** 2) ** 2 -
-                             16 * eccentricity_ratio ** 4) /
-           (eccentricity_ratio * np.sqrt(1 - eccentricity_ratio ** 2)))
-    kyx = (-a * h0 * np.pi * ((np.pi ** 2) * (1 - eccentricity_ratio ** 2) *
-                              (1 + 2 * eccentricity_ratio ** 2) +
-                              (32 * eccentricity_ratio ** 2) * (1 + eccentricity_ratio ** 2)) /
-           (eccentricity_ratio * np.sqrt(1 - eccentricity_ratio ** 2)))
-    kyy = (a * h0 * 4 * ((np.pi ** 2) * (1 + 2 * eccentricity_ratio ** 2) +
-                         ((32 * eccentricity_ratio ** 2) *
-                          (1 + eccentricity_ratio ** 2)) / (1 - eccentricity_ratio ** 2)))
-    # fmt: on
-    return [kxx, kxy, kyx, kyy]
-
-
 def calculate_analytical_damping_matrix(load, eccentricity_ratio, radial_clearance, omega):
     """Returns the damping matrix calculated analytically.
     Suitable only for short bearings.
@@ -182,14 +138,17 @@ def calculate_oil_film_force(fluid_flow_object, force_type=None):
     return radial_force, tangential_force, force_x, force_y
 
 
-def calculate_stiffness_matrix(fluid_flow_object, oil_film_force='numerical'):
+def calculate_stiffness_matrix(fluid_flow_object, force_type=None, oil_film_force='numerical'):
     """This function calculates the bearing stiffness matrix numerically.
     Parameters
     ----------
     fluid_flow_object: A FluidFlow object.
     oil_film_force: str
-        If set, calculates the oil film force analytically considering the chosen type: 'short' or 'long'.
[…]
