### 2020-02-10 dbc1b55
Organizes the function that calculates the damping matrix
