### 2020-02-10 6a9192d
Organizes the function that calculates the damping matrix
