### 2020-02-11 d6649b7
Update bearing_seal_element.py
