+++ ross/api_report.py
@@ -807,7 +807,7 @@ class Report:
         bearing_list = [
             copy(b)
             for b in self.rotor.bearing_elements
-            if not isinstance(b, SealElement)
+            if not isinstance(b, rs.SealElement)
             ]
 
         # Applying cross-coupling on rotor mid-span
