### 2020-02-11 2dad4b7
Update bearing_seal_element.py
