### 2020-02-11 f3e23aa
Correct docstrings.
