### 2020-02-11 98d1517
Add check_units module
If we use the check_units decorator in a function the arguments are checked, and if they are in the units
dictionary, they are converted to the 'default' unit given in the dictionary.

For example:
>>> units = {
... "L": "meter",
... }

>>> @check_units
... def foo(L=None):
...     print(L)
...

If we call the function with the argument as a float:
>>> foo(L=0.5)
0.5 meter

the L argument will be available inside the function as a pint.Quantity object
(with the magnitude and units attributes).

If we call the function with a pint.Quantity object the value is automaticall
[…]
