### 2020-02-11 9f75eab
Add check_units to ShaftElement
This add the check_units decorator to the ShaftElement's __init__
method. The arguments with units for now are L, idl, idr, odl and odr.
