### 2020-02-12 46329fc
Move assignment of magnitude to names
This moves the assignment of the pint.Quantity magnitude to names
to after checking for None in idr or odr. This way we won't get a
attribute error trying to get .m on a None object.
