### 2020-02-12 acefa9c
Add tests to units
We basically test the consistency to the default units and the
conversion to the default units.
