### 2020-02-13 ddf20f7
Correct docstrings.
