### 2020-02-13 6f7deca
Correct docstrings
Changing the fluid_flow_example it was necessary to correct the docstrings of the functions that use this
example
