### 2020-02-13 774cc93
Correct docstrings
Changing the fluid_flow_example it was necessary to correct the docstrings of the functions that use this
example
