+++ ross/fluid_flow/fluid_flow.py
@@ -506,7 +506,7 @@ def fluid_flow_example():
     --------
     >>> my_fluid_flow = fluid_flow_example()
     >>> my_fluid_flow.eccentricity
-    0.001
+    0.0001
     """
     my_pressure_matrix = FluidFlow(nz=16, ntheta=132, nradius=11, length=0.04, omega=100. * 2 * np.pi / 60,
                                    p_in=0., p_out=0., radius_rotor=0.2, radius_stator=0.2002,
