### 2020-02-16 a3908d4
Fix grid to decrease testing time
