### 2020-02-18 af3d8bc
Reduces test_find_equilibrium_position function time
