+++ ross/tests/test_fluid_flow.py
@@ -320,7 +320,7 @@ def test_matplotlib_plots():
 def test_find_equilibrium_position():
     bearing = flow.fluid_flow_example2()
     find_equilibrium_position(bearing, print_along=False, tolerance=0.1, increment_factor=0.01,
-                              max_iterations=5, increment_reduction_limit=1e-03)
+                              max_iterations=2, increment_reduction_limit=1e-03)
     assert_allclose(bearing.eccentricity, (bearing.radius_stator - bearing.radius_rotor)*0.2663, atol=0.001)
 
 
