### 2020-02-18 8859244
Initial branch commint, including a requirements txt file and an update to gitignore for virtual environments
