### 2020-02-18 f42ce58
Basic class definitions for 6 DoF tapered shaft element.
