### 2020-02-18 9ed9163
Basic error testing functions for tapered 6 DoF shaft elements.
