### 2020-02-18 3792691
Addition of Lalanne rotordynamics book to the bibliography file.
