+++ docs/refs.bib
@@ -27,6 +27,13 @@
   publisher={Cambridge university press}
 }
 
+@book{lalanne2001,
+  title={Rotordynamics Prediction in Engineering},
+  author={M Lalanne and G Ferraris},
+  year={1998},
+  publisher={John Wiley and Sons
+}
+
 @article{zang2005review,
   title={A review of robust optimal design and its application in dynamics},
   author={Zang, C and Friswell, MI and Mottershead, JE},
