### 2020-02-18 81ea73a
Introduction of 6 DoF shaft element initial documentation into class header.
