### 2020-02-19 e94b994
Bibliography update
