+++ docs/refs.bib
@@ -634,3 +634,10 @@ year={2011}
 	year={2012}
 }
 
+@article{Hutchingson2001,
+	title={Shear Coefficients for Timoshenko Beam Theory},
+	author={Hutchingson, J. R.},
+	journal={J. Appl. Mech.},
+	volume={68},
+	year={2001}
+}
