### 2020-02-19 d12dabc
Shaft 6DoF initial unit check and description update
