+++ ross/shaft_element.py
@@ -1226,45 +1226,44 @@ class ShaftElement6DoF(Element):
     L : float, pint.Quantity
         Element length.
     idl : float, pint.Quantity
-        Inner diameter of the element at the left position..
+        Inner diameter of the element at the left node.
     odl : float, pint.Quantity
-        Outer diameter of the element at the left position.
+        Outer diameter of the element at the left node.
     idr : float, pint.Quantity, optional
-        Inner diameter of the element at the right position
-        Default is equal to idl value (cylindrical element)
+        Inner diameter of the element at the right node;
+        Default is equal to idl value for cylindrical element.
     odr : float, pint.Quantity, optional
-        Outer diameter of the element at the right position.
-        Default is equal to odl value (cylindrical element)
+        Outer diameter of the element at the right node;
+        Default is equal to odl value for cylindrical element.
     material : ross.material
         Shaft material.
     n : int, optional
-        Element number (coincident with it's first node).
+        Element number, coincident with it's first node.
         If not given, it will be set when the rotor is assembled
         according to the element's position in the list supplied to
         the rotor constructor.
     axial_force : float, optional
-        Axial force.
+        Axial force;
+        Default is zero.
     torque : float, optional
-        Torque.
+        Torque moment;
+        Default is zero.
     shear_effects : bool, optional
-        Determine if shear effects are taken into account.
+        Determine if shear effects are taken into account;
         Default is True.
     rotary_inertia : bool, optional
-        Determine if rotary_inertia effects are taken into account.
+        Determine if rotary_inertia effects are taken into account;
         Default is True.
     gyroscopic : bool, optional
-        Determine if gyroscopic effects are taken into account.
+        Determine if gyroscopic effects are taken into account;
         Default is True.
-    shear_method_calc : string, optional
-        Determines which shear calculation method the user will adopt
-        Default is 'cowper'
     tag : str, optional
-        Element tag.
+        Element tag;
         Default is None.
 
     Returns
     -------
-    A 6 degrees of freedom shaft element
+    A 6 degrees of freedom shaft element, with available gyroscopic, shear and rotary inertia effects.
 
     Attributes
     ----------
@@ -1290,7 +1289,7 @@ class ShaftElement6DoF(Element):
         consider rotary inertia and shear effects. If these are not considered
         :math:`\phi=0`.
     kappa : float
-        Shear coefficient for the element.
+        Shear coefficient for the element, determined from :cite:`Hutchingson2001` formulation.
     alpha : float
         Proportional damping coefficient, associated to the element Mass matrix
     beta : float
@@ -1298,7 +1297,6 @@ class ShaftElement6DoF(Element):
 
     References
     ----------
-
     .. bibliography:: ../../../docs/refs.bib
 
     Examples
@@ -1315,13 +1313,92 @@ class ShaftElement6DoF(Element):
     """
 
     @check_units
-    def __init__(self):
-        pass
+    def __init__(
+        self,
+        L,
+        idl,
+        odl,
+        idr,
+        odr,
+        material=None,
+        n=None,
+        axial_force=0,
+        torque=0,
+        shear_effects=True,
+        rotary_inertia=True,
+        gyroscopic=True,
+        shear_method_calc="cowper",
+        tag=None,
+    ):
+
+        if idr is None:
+            idr = idl
+        if odr is None:
+            odr = odl
+
+        # Changing units to only their magnitude
+        L = L.m
+        idl = idl.m
+        odl = odl.m
+        idr = idr.m
+        odr = odr.m
+
+        if material is None:
[…]
