### 2020-02-19 4e3d82a
Update .run_modal() docstring
Update docstring to match the modifications done to ModalResults class in results.py
