### 2020-02-20 dd03e0b
Fixing error in remove_material()
