### 2020-02-20 71784a4
Introduction of Timoshenko factor determination, eq def and hash def
