### 2020-02-20 4f22c59
Introduction of Alpha and Beta material damping coefficients as self.alpha and self.beta. Introduction of Save and Load functionalities
