### 2020-02-20 ca143b6
Classmethod from_table introduction and correction, removing shear calculation method and adding alpha and beta
