### 2020-02-20 3c11005
defs for shaft string representation. Added 6 DoF variables and both alpha and beta damping factors
