### 2020-02-20 a52d0cb
defs for patches and bokeh plots, adding alpha and beta to the tooltips
