+++ ross/shaft_element.py
@@ -1716,6 +1716,7 @@ class ShaftElement6DoF(Element):
             f"n={self.n})"
         )
 
+
     def __str__(self):
         """
         Method to convert object into string
@@ -1740,23 +1741,178 @@ class ShaftElement6DoF(Element):
             f"\n"
         )
 
+
     def M(self):
         pass
+
 
     def K(self):
         pass
 
+
     def C(self):
         pass
+
 
     def G(self):
         pass
 
+
     def patch(self, position, check_sld, ax):
-        pass
+        """Shaft element patch.
+        Patch that will be used to draw the shaft element.
+        Parameters
+        ----------
+        position : float
+            Position in which the patch will be drawn.
+        check_sld : bool
+            If True, color the elements in yellow if slenderness ratio < 1.6
+        ax : matplotlib axes, optional
+            Axes in which the plot will be drawn.
+        Returns
+        -------
+        """
+        if check_sld is True and self.slenderness_ratio < 1.6:
+            mpl_color = "yellow"
+            legend = "Shaft - Slenderness Ratio < 1.6"
+        else:
+            mpl_color = self.color
+            legend = "Shaft"
+
+        shaft_points_u = [
+            [position, self.idl / 2],
+            [position, self.odl / 2],
+            [position + self.L, self.odr / 2],
+            [position + self.L, self.idr / 2],
+            [position, self.idl / 2],
+        ]
+        shaft_points_l = [
+            [position, -self.idl / 2],
+            [position, -self.odl / 2],
+            [position + self.L, -self.odr / 2],
+            [position + self.L, -self.idr / 2],
+            [position, -self.idl / 2],
+        ]
+
+        # matplotlib - plot the upper half of the shaft
+        ax.add_patch(
+            mpatches.Polygon(
+                shaft_points_u,
+                facecolor=mpl_color,
+                linestyle="--",
+                linewidth=0.5,
+                ec="k",
+                alpha=0.8,
+                label=legend,
+            )
+        )
+        # matplotlib - plot the lower half of the shaft
+        ax.add_patch(
+            mpatches.Polygon(
+                shaft_points_l,
+                facecolor=mpl_color,
+                linestyle="--",
+                linewidth=0.5,
+                ec="k",
+                alpha=0.8,
+            )
+        )
+
 
     def bokeh_patch(self, position, check_sld, bk_ax):
-        pass
+        """Shaft element patch.
+        Patch that will be used to draw the shaft element.
+        Parameters
+        ----------
+        position : float
+            Position in which the patch will be drawn.
+        check_sld : bool
+            If True, HoverTool displays only the slenderness ratio and color
+            the elements in yellow if slenderness ratio < 1.6
+        bk_ax : bokeh plotting axes, optional
+            Axes in which the plot will be drawn.
+        Returns
+        -------
+        hover : Bokeh HoverTool
+            Bokeh HoverTool axes
+        """
+        if check_sld is True and self.slenderness_ratio < 1.6:
+            bk_color = "yellow"
+            legend = "Shaft - Slenderness Ratio < 1.6"
+        else:
+            bk_color = self.material.color
+            legend = "Shaft"
+
+        # bokeh plot - plot the shaft
+        z_upper = [position, position, position + self.L, position + self.L]
+        y_upper = [self.idl / 2, self.odl / 2, self.odr / 2, self.idr / 2]
+        z_lower = [position, position, position + self.L, position + self.L]
+        y_lower = [-self.idl / 2, -self.odl / 2, -self.odr / 2, -self.idr / 2]
+
+        source = ColumnDataSource(
+            dict(
+                z_l=[z_lower],
+                y_l=[y_lower],
+                z_u=[z_upper],
+                y_u=[y_upper],
[…]
