### 2020-02-20 d891b87
Adapt Bearings glyphs to CoAxialRotor class
* Better variables to plot bearings: it's all based on the width (`icon_w`) and height (`icon_h`).
* Remove `patch()` and `bokeh_patch()` from all classes but `BearingElement` because it's inherited from this one.
