+++ ross/bearing_seal_element.py
@@ -538,16 +538,14 @@ class BearingElement(Element):
 
         return G
 
-    def patch(self, position, mean, ax, **kwargs):
+    def patch(self, position, ax, **kwargs):
         """Bearing element patch.
         Patch that will be used to draw the bearing element.
 
         Parameters
         ----------
         position : tuple
-            Position (z, y) in which the patch will be drawn.
-        mean : float
-            Mean value of shaft element outer diameters
+            Position (z, y_low, y_upp) in which the patch will be drawn.
         ax : matplotlib axes, optional
             Axes in which the plot will be drawn.
 
@@ -559,15 +557,17 @@ class BearingElement(Element):
             kwargs.setdefault(k, v)
 
         # geometric factors
-        mean = self.scale_factor * mean
-        zpos, ypos = position
+        zpos, ypos, ypos_s = position
+
+        icon_h = ypos_s - ypos  # bearing icon height
+        icon_w = icon_h / 2.0  # bearing icon width
         coils = 6  # number of points to generate spring
         n = 5  # number of ground lines
-        step = mean / (coils + 1)  # spring step
+        step = icon_w / (coils + 1)  # spring step
 
-        zs0 = zpos - (mean / 2.0)
-        zs1 = zpos + (mean / 2.0)
-        ys0 = ypos + 0.5 * mean
+        zs0 = zpos - (icon_w / 2.0)
+        zs1 = zpos + (icon_w / 2.0)
+        ys0 = ypos + 0.25 * icon_h
 
         # plot bottom base
         x_bot = [zpos, zpos, zs0, zs1]
@@ -579,10 +579,10 @@ class BearingElement(Element):
         # plot top base
         x_top = [zpos, zpos, zs0, zs1]
         yl_top = [
-            ypos + 2 * mean,
-            ypos + 1.5 * mean,
-            ypos + 1.5 * mean,
-            ypos + 1.5 * mean,
+            ypos + icon_h,
+            ypos + 0.75 * icon_h,
+            ypos + 0.75 * icon_h,
+            ypos + 0.75 * icon_h,
         ]
         yu_top = [-y for y in yl_top]
         ax.add_line(mlines.Line2D(x_top, yl_top, **kwargs))
@@ -606,7 +606,7 @@ class BearingElement(Element):
 
         # plot spring
         z_spring = np.array([zs0, zs0, zs0, zs0])
-        yl_spring = np.array([ys0, ys0 + step, ys0 + mean - step, ys0 + mean])
+        yl_spring = np.array([ys0, ys0 + step, ys0 + icon_w - step, ys0 + icon_w])
 
         for i in range(coils):
             z_spring = np.insert(z_spring, i + 2, zs0 - (-1) ** i * step)
@@ -634,22 +634,20 @@ class BearingElement(Element):
 
         # plot damper - top
         z_damper3 = [z_damper2[0], z_damper2[2], zs1, zs1]
-        yl_damper3 = [ys0 + 4 * step, ys0 + 4 * step, ys0 + 4 * step, ypos + 1.5 * mean]
+        yl_damper3 = [ys0 + 4 * step, ys0 + 4 * step, ys0 + 4 * step, ypos + 1.5 * icon_w]
         yu_damper3 = [-y for y in yl_damper3]
 
         ax.add_line(mlines.Line2D(z_damper3, yl_damper3, **kwargs))
         ax.add_line(mlines.Line2D(z_damper3, yu_damper3, **kwargs))
 
-    def bokeh_patch(self, position, mean, bk_ax, **kwargs):
+    def bokeh_patch(self, position, bk_ax, **kwargs):
         """Bearing element patch.
[…]
