### 2020-02-20 700dfbb
Adapt Disk glyph to CoAxialRotor class
