### 2020-02-20 f718e61
Add CoAxialRotor class
* Add CoAxialRotor class
* Add new example to run CoAxialRotor class
* Adapt `.plot_rotor()` method to plot co axial rotor system.
* Add ["shaft_number"] column to the dataframe to identify each shaft in the rotor system.
* Run black to format code.
