+++ ross/rotor_assembly.py
@@ -1,9 +1,9 @@
 import os
 import shutil
 import warnings
-from collections import Iterable, namedtuple
+from collections import Iterable, namedtuple, Counter
 from copy import copy, deepcopy
-from itertools import cycle
+from itertools import cycle, chain
 from pathlib import Path
 
 import bokeh.palettes as bp
@@ -18,12 +18,12 @@ import scipy.sparse.linalg as las
 import toml
 from bokeh.models import ColumnDataSource
 from bokeh.models.glyphs import Text
-from bokeh.plotting import figure, output_file
+from bokeh.plotting import figure, output_file, show
 from cycler import cycler
 from ross.utils import convert
 
 import ross
-from ross.bearing_seal_element import BearingElement
+from ross.bearing_seal_element import BearingElement, SealElement
 from ross.disk_element import DiskElement
 from ross.materials import steel
 from ross.results import (
@@ -35,11 +35,11 @@ from ross.results import (
     StaticResults,
     SummaryResults,
     TimeResponseResults,
-    OrbitResponseResults
+    OrbitResponseResults,
 )
 from ross.shaft_element import ShaftElement
 
-__all__ = ["Rotor", "rotor_example"]
+__all__ = ["Rotor", "CoAxialRotor", "rotor_example", "coaxrotor_example"]
 
 # set style and colors
 plt.style.use("seaborn-white")
@@ -212,7 +212,6 @@ class Rotor(object):
                 p_mass.tag = "Point Mass " + str(i)
 
         self.shaft_elements = sorted(shaft_elements, key=lambda el: el.n)
-
         self.bearing_elements = sorted(bearing_elements, key=lambda el: el.n)
         self.disk_elements = disk_elements
         self.point_mass_elements = point_mass_elements
@@ -300,6 +299,13 @@ class Rotor(object):
         )
         df = df.sort_values(by="n_l")
         df = df.reset_index(drop=True)
+        df["shaft_number"] = np.zeros(len(df))
+
+        df_shaft["shaft_number"] = np.zeros(len(df_shaft))
+        df_disks["shaft_number"] = np.zeros(len(df_disks))
+        df_bearings["shaft_number"] = np.zeros(len(df_bearings))
+        df_seals["shaft_number"] = np.zeros(len(df_seals))
+        df_point_mass["shaft_number"] = np.zeros(len(df_point_mass))
 
         self.df_disks = df_disks
         self.df_bearings = df_bearings
@@ -384,19 +390,25 @@ class Rotor(object):
                     global_dof_mapping[k] = 4 * elm.n + v
             else:
                 for k, v in global_dof_mapping.items():
-                    global_dof_mapping[k] = 2 * n_last + 2 * elm.n + 4 + v                
+                    global_dof_mapping[k] = 2 * n_last + 2 * elm.n + 4 + v
 
             if hasattr(elm, "n_link") and elm.n_link is not None:
-                if elm.n_link <= n_last + 1:                
+                if elm.n_link <= n_last + 1:
                     global_dof_mapping[f"x_{elm.n_link}"] = 4 * elm.n_link
                     global_dof_mapping[f"y_{elm.n_link}"] = 4 * elm.n_link + 1
                 else:
-                    global_dof_mapping[f"x_{elm.n_link}"] = 2 * n_last + 2 * elm.n_link + 4
-                    global_dof_mapping[f"y_{elm.n_link}"] = 2 * n_last + 2 * elm.n_link + 5
+                    global_dof_mapping[f"x_{elm.n_link}"] = (
+                        2 * n_last + 2 * elm.n_link + 4
+                    )
+                    global_dof_mapping[f"y_{elm.n_link}"] = (
[…]
