### 2020-02-20 ecdf601
Fix typo in CoAxialRotor docstring
