### 2020-02-20 b16adbf
Add initial tests for CoAxialRotor class
* These tests for CoAxialRotor verify some assembling aspects.
* It's replacing a static analysis test. It already has 2 tests.
