+++ ross/tests/test_rotor_assembly.py
@@ -851,91 +851,72 @@ def test_static_analysis_rotor5(rotor5):
 
 @pytest.fixture
 def rotor6():
-    #  Overhung rotor without damping with 10 shaft elements
-    #  2 disks and 2 bearings
+    #  Co-axial rotor system with 2 shafts, 4 disks and
+    #  4 bearings (3 to ground and 1 to body) 
     i_d = 0
     o_d = 0.05
     n = 10
     L = [0.25 for _ in range(n)]
 
-    shaft_elem = [
-        ShaftElement(
-            l, i_d, o_d, material=steel, shear_effects=True, rotary_inertia=True, gyroscopic=True
-        )
-        for l in L
-    ]
+    axial_shaft = [ShaftElement(l, i_d, o_d, material=steel) for l in L]
 
-    disk0 = DiskElement.from_geometry(5, steel, 0.07, 0.05, 0.28)
-    disk1 = DiskElement.from_geometry(10, steel, 0.07, 0.05, 0.35)
+    i_d = 0.25
+    o_d = 0.30
+    n = 6
+    L = [0.25 for _ in range(n)]
+
+    coaxial_shaft = [ShaftElement(l, i_d, o_d, material=steel) for l in L]
+
+    disk0 = DiskElement.from_geometry(
+        n=1, material=steel, width=0.07, i_d=0.05, o_d=0.28
+    )
+    disk1 = DiskElement.from_geometry(
+        n=9, material=steel, width=0.07, i_d=0.05, o_d=0.28
+    )
+    disk2 = DiskElement.from_geometry(
+        n=13, material=steel, width=0.07, i_d=0.20, o_d=0.48
+    )
+    disk3 = DiskElement.from_geometry(
+        n=15, material=steel, width=0.07, i_d=0.20, o_d=0.48
+    )
+
+    shaft = [axial_shaft, coaxial_shaft]
+    disks = [disk0, disk1, disk2, disk3]
 
     stfx = 1e6
     stfy = 1e6
-    bearing0 = BearingElement(2, kxx=stfx, kyy=stfy, cxx=0)
-    bearing1 = BearingElement(8, kxx=stfx, kyy=stfy, cxx=0)
+    bearing0 = BearingElement(0, kxx=stfx, kyy=stfy, cxx=0)
+    bearing1 = BearingElement(10, kxx=stfx, kyy=stfy, cxx=0)
+    bearing2 = BearingElement(11, kxx=stfx, kyy=stfy, cxx=0)
+    bearing3 = BearingElement(8, n_link=17, kxx=stfx, kyy=stfy, cxx=0)
+    bearings = [bearing0, bearing1, bearing2, bearing3]
 
-    return Rotor(shaft_elem, [disk0, disk1], [bearing0, bearing1])
+    return CoAxialRotor(shaft, disks, bearings)
 
 
-def test_static_analysis_rotor6(rotor6):
-    rotor6.run_static()
-
-    assert_almost_equal(
-        rotor6.disp_y,
-        (
-            [
-                -1.03951876e-04,
-                -4.93624668e-05,
-                -1.79345202e-12,
-                3.74213098e-05,
-                7.66066703e-05,
-                1.29084322e-04,
-                1.85016673e-04,
-                1.72933811e-04,
-                -1.02148266e-11,
-                -3.96409257e-04,
-                -9.20006704e-04,
-            ]
-        ),
-        decimal=6,
-    )
-    assert_almost_equal(
-        rotor6.Vx,
-        [
-            0,
-            37.5954,
[…]
