### 2020-02-20 f13a538
It was added the stochastic part of the model
