### 2020-02-21 91a17ad
Mass matrix auxiliary commands
