### 2020-02-21 b6cecdf
Add back the test removed on the last commit
