### 2020-02-21 b778879
Remove ross import and add path as constant
This removes ross import from materials.py and ross path and
available materials path as constants so we can use it throughout the
code.
