### 2020-02-21 89f367d
Change how E, G_s and Poisson are handled
This fix the case were the user could create a material with
inconsistent values for E, G_s and Poisson by providing the tree
arguments.
