### 2020-02-21 fb7a4b5
Use available_materials path constant
