### 2020-02-21 a67c6f5
Remove Poisson from repr and from load
This way we can guarantee that the material will be reloaded by using
only E and G_s as arguments.
