+++ ross/tests/test_materials.py
@@ -1,26 +1,19 @@
-import os
-
 import pytest
 from numpy.testing import assert_allclose
 
-import ross
 from ross.materials import *
 
-os.chdir(os.path.dirname(ross.__file__))
 
-AISI4140 = Material(
-    name="AISI4140",
-    rho=7850,
-    E=203200000000.0,
-    Poisson=0.27,
-    G_s=80000000000.0,
-    color="#525252",
-)
+@pytest.fixture
+def AISI4140():
+    return Material(
+        name="AISI4140", rho=7850, E=203200000000.0, G_s=80000000000.0, color="#525252",
+    )
 
 
 def test_raise_name_material():
-    with pytest.raises(AssertionError) as excinfo:
-        mat = Material("with space", rho=7850, G_s=80e9, Poisson=0.27)
+    with pytest.raises(ValueError) as excinfo:
+        Material("with space", rho=7850, G_s=80e9, Poisson=0.27)
     assert "Spaces are not allowed" in str(excinfo.value)
 
 
@@ -46,13 +39,13 @@ def test_Poisson():
 
 
 def test_E_G_s_Poisson():
-    mat = Material(name="test", rho=7850, E=203.2e9, G_s=80e9, Poisson=0.27)
+    mat = Material(name="test", rho=7850, E=203.2e9, G_s=80e9)
     assert_allclose(mat.E, 203.2e9)
     assert_allclose(mat.G_s, 80e9)
     assert_allclose(mat.Poisson, 0.27)
 
 
-def test_specific_material():
+def test_specific_material(AISI4140):
     assert_allclose(AISI4140.rho, 7850)
     assert_allclose(AISI4140.E, 203.2e9)
     assert_allclose(AISI4140.G_s, 80e9)
@@ -67,7 +60,7 @@ def test_error_rho():
 
 
 def test_error_E_G_s_Poisson():
-    with pytest.raises(AssertionError) as ex:
+    with pytest.raises(ValueError) as ex:
         Material(name="test", rho=785, E=203.2e9)
     assert "At least 2 arguments from E" in str(ex.value)
 
