### 2020-02-21 19433d7
Get magnitude within the check_units function
Since we are not using pint on all functions, we could have problems
when interacting pint Quantities with standard floats. This requires
us to aways get the magnitude attribute. Before we would do this in the
decorated function, but it is simpler (and avoids repetition) to do it
within the check_units function.
