### 2020-02-21 e36fc68
Improve error message
One arguments it not sufficient and three arguments would be
inconsistent, so the error message asks for exactly two arguments.
