### 2020-02-21 e761d49
Change test to check for wrong number of arguments
