### 2020-02-25 b984372
Adapt run_static() for multiple shafts
Some improvements to `.run_static()`:
* Supports CoAxialRotor class (multiple shafts)
* More information when returning disks and bearing forces. The user can choose to return a dict with nodal position for each force or with a element tag.
* Remove some repetitive calculus on code.
