+++ ross/rotor_assembly.py
@@ -2178,141 +2178,248 @@ class Rotor(object):
         -------
         >>> rotor = rotor_example()
         >>> static = rotor.run_static()
-        >>> rotor.bearing_reaction_forces
-        [432.4, 432.4]
+        >>> rotor.bearing_forces_nodal
+        {'node_0': 432.4, 'node_6': 432.4}
+        >>> rotor.bearing_forces_tag
+        {'Bearing 0': 432.4, 'Bearing 1': 432.4}
         """
-        if len(self.df_bearings) == 0:
+        if not len(self.df_bearings):
             raise ValueError("Rotor has no bearings")
 
+        aux_brg = []
+        for elm in self.bearing_elements:
+            if elm.n not in self.nodes:
+                pass
+            elif elm.n_link in self.nodes:
+                aux_brg.append(
+                    BearingElement(n=elm.n, n_link=elm.n_link, kxx=1e14, cxx=0)
+                )
+            else:
+                aux_brg.append(
+                    BearingElement(n=elm.n, kxx=1e14, cxx=0)
+                )
+ 
+        if isinstance(self, CoAxialRotor):
+            aux_rotor = CoAxialRotor(
+                    self.shafts,
+                    self.disk_elements,
+                    aux_brg,
+            )
         else:
-            # gravity aceleration vector
-            grav = np.zeros((len(self.M()), 1))
+            aux_rotor = Rotor(
+                    self.shaft_elements,
+                    self.disk_elements,
+                    aux_brg,
+            )
+        aux_K = aux_rotor.K(0)
+        aux_M = aux_rotor.M()
 
-            # place gravity effect on shaft and disks nodes
-            for node_y in range(int(len(self.M()) / 4)):
-                grav[4 * node_y + 1] = -9.8065
+        for elm in aux_rotor.bearing_elements:
+            if isinstance(elm, SealElement):
+                dofs = elm.dof_global_index
+                try:
+                    aux_K[np.ix_(dofs, dofs)] -= elm.K(0)
+                except TypeError:
+                    aux_K[np.ix_(dofs, dofs)] -= elm.K()
 
-            aux_brg = []
-            for n in self.df_bearings["n"]:
-                aux_brg.append(BearingElement(n=n, kxx=1e14, cxx=0))
+        df_num = aux_rotor.df["shaft_number"].values
+        sh_num = [int(item) for item, count in Counter(df_num).items() if count > 1]
 
-            aux_rotor = Rotor(self.shaft_elements, self.disk_elements, aux_brg)
-            aux_K = np.zeros((aux_rotor.ndof, aux_rotor.ndof))
+        # gravity aceleration vector
+        g = 9.8065
+        grav = np.zeros(len(aux_K))
 
-            for elm in aux_rotor.elements:
-                if elm.__class__.__name__ != "SealElement":
-                    dofs = elm.dof_global_index
-                    n0 = dofs[0]
-                    n1 = dofs[-1] + 1  # +1 to include this dof in the slice
-                    try:
-                        aux_K[n0:n1, n0:n1] += elm.K(0)
-                    except TypeError:
-                        aux_K[n0:n1, n0:n1] += elm.K()
+        # place gravity effect on shaft and disks nodes
+        for node_y in range(int(len(aux_K) / 4)):
+            grav[4 * node_y + 1] = - g
 
-            # calculates x, for [K]*(x) = [M]*(g)
-            disp = (la.solve(aux_K, self.M() @ grav)).flatten()
+        # calculates x, for [K]*(x) = [M]*(g)
[…]
