### 2020-02-25 123899d
Adapt sintaxe to `.run_static()` tests
