+++ ross/rotor_assembly.py
@@ -2195,22 +2195,12 @@ class Rotor(object):
                     BearingElement(n=elm.n, n_link=elm.n_link, kxx=1e14, cxx=0)
                 )
             else:
-                aux_brg.append(
-                    BearingElement(n=elm.n, kxx=1e14, cxx=0)
-                )
- 
+                aux_brg.append(BearingElement(n=elm.n, kxx=1e14, cxx=0))
+
         if isinstance(self, CoAxialRotor):
-            aux_rotor = CoAxialRotor(
-                    self.shafts,
-                    self.disk_elements,
-                    aux_brg,
-            )
+            aux_rotor = CoAxialRotor(self.shafts, self.disk_elements, aux_brg)
         else:
-            aux_rotor = Rotor(
-                    self.shaft_elements,
-                    self.disk_elements,
-                    aux_brg,
-            )
+            aux_rotor = Rotor(self.shaft_elements, self.disk_elements, aux_brg)
         aux_K = aux_rotor.K(0)
         aux_M = aux_rotor.M()
 
@@ -2231,7 +2221,7 @@ class Rotor(object):
 
         # place gravity effect on shaft and disks nodes
         for node_y in range(int(len(aux_K) / 4)):
-            grav[4 * node_y + 1] = - g
+            grav[4 * node_y + 1] = -g
 
         # calculates x, for [K]*(x) = [M]*(g)
         disp = (la.solve(aux_K, aux_M @ grav)).flatten()
@@ -2256,44 +2246,56 @@ class Rotor(object):
             if not pd.isna(aux_rotor.df_bearings.loc[i, "n_link"]):
                 BRG[node] = (
                     BRG[node]
-                    + disp_y[node] * self.df_bearings.loc[
-                        self.df_bearings.tag == aux_rotor.df_bearings.tag, "kyy"][0].coefficient[0]
+                    + disp_y[node]
+                    * self.df_bearings.loc[
+                        self.df_bearings.tag == aux_rotor.df_bearings.tag, "kyy"
+                    ][0].coefficient[0]
                 )
                 BrgForce_nodal["node_" + str(node)] = np.around(
                     BrgForce_nodal["node_" + str(node)]
-                    + disp_y[node] * self.df_bearings.loc[
-                        self.df_bearings.tag == aux_rotor.df_bearings.tag, "kyy"][0].coefficient[0],
+                    + disp_y[node]
+                    * self.df_bearings.loc[
+                        self.df_bearings.tag == aux_rotor.df_bearings.tag, "kyy"
+                    ][0].coefficient[0],
                     decimals=1,
                 )
-                BrgForce_tag[aux_rotor.df_bearings.loc[i, "tag"]] = BrgForce_nodal["node_" + str(node)]
+                BrgForce_tag[aux_rotor.df_bearings.loc[i, "tag"]] = BrgForce_nodal[
+                    "node_" + str(node)
+                ]
 
                 node = int(aux_rotor.df_bearings.loc[i, "n_link"])
                 BRG[node] = (
                     BRG[node]
                     - disp_y[node]
-                    * self.df_bearings.loc[self.df_bearings.n_link == node, "kyy"].values[0].coefficient[0]
+                    * self.df_bearings.loc[self.df_bearings.n_link == node, "kyy"]
+                    .values[0]
+                    .coefficient[0]
                 )
                 BrgForce_nodal["node_" + str(node)] = np.around(
                     BrgForce_nodal["node_" + str(node)]
                     - disp_y[node]
[…]
