### 2020-02-25 21eb1e0
Adapt `StaticResults` to new sintaxe for `run_static()`
* `StaticResults` now supports plotting results for multiple shafts.
* Multiple shafts are plotted in the same figure (`.plot_free_body_diagram()` will display multiple figures).
* New: the user can ommit a result by clicking on the respective legend (not valid for `.plot_free_body_diagram()`).
