+++ ross/results.py
@@ -1697,15 +1697,15 @@ class StaticResults:
     """
 
     def __init__(
-        self, disp_y, Vx, Bm, df_shaft, df_disks, df_bearings, nodes, nodes_pos, Vx_axis
+        self, disp_y, Vx, Bm, w_shaft, disk_forces, bearing_forces, nodes, nodes_pos, Vx_axis
     ):
 
         self.disp_y = disp_y
         self.Vx = Vx
         self.Bm = Bm
-        self.df_shaft = df_shaft
-        self.df_disks = df_disks
-        self.df_bearings = df_bearings
+        self.w_shaft = w_shaft
+        self.disk_forces = disk_forces
+        self.bearing_forces = bearing_forces
         self.nodes = nodes
         self.nodes_pos = nodes_pos
         self.Vx_axis = Vx_axis
@@ -1724,76 +1724,70 @@ class StaticResults:
         fig : bokeh figure
             Bokeh figure with static deformation plot
         """
-        source = ColumnDataSource(
-            data=dict(x=self.nodes_pos, y0=self.disp_y, y1=[0] * len(self.nodes_pos))
-        )
-
-        TOOLTIPS = [
-            ("Shaft lenght:", "@x"),
-            ("Underformed:", "@y1"),
-            ("Displacement:", "@y0"),
-        ]
-
         # create displacement plot
         fig = figure(
-            tools="pan, wheel_zoom, box_zoom, reset, save, box_select, hover",
-            tooltips=TOOLTIPS,
+            tools="pan, wheel_zoom, box_zoom, reset, save, box_select",
             width=640,
             height=480,
-            title="Static Analysis",
+            title="Deformation",
             x_axis_label="Shaft lenght",
             y_axis_label="Lateral displacement",
         )
-        fig.xaxis.axis_label_text_font_size = "20pt"
-        fig.yaxis.axis_label_text_font_size = "20pt"
+        fig.xaxis.axis_label_text_font_size = "16pt"
+        fig.yaxis.axis_label_text_font_size = "16pt"
         fig.axis.major_label_text_font_size = "16pt"
         fig.title.text_font_size = "14pt"
 
-        interpolated = interpolate.interp1d(
-            source.data["x"], source.data["y0"], kind="cubic"
-        )
-        xnew = np.linspace(
-            source.data["x"][0],
-            source.data["x"][-1],
-            num=len(self.nodes_pos) * 20,
-            endpoint=True,
-        )
+        count = 0
+        for disp_y, Vx, Bm, nodes, nodes_pos, Vx_axis in zip(
+                self.disp_y, self.Vx, self.Bm, self.nodes, self.nodes_pos, self.Vx_axis
+        ):
+            source = ColumnDataSource(
+                data=dict(x=nodes_pos, y=disp_y)
+            )
 
-        ynew = interpolated(xnew)
-        auxsource = ColumnDataSource(data=dict(x=xnew, y0=ynew, y1=[0] * len(xnew)))
+            interpolated = interpolate.interp1d(
+                source.data["x"], source.data["y"], kind="cubic"
+            )
+            xnew = np.linspace(
+                source.data["x"][0],
+                source.data["x"][-1],
+                num=len(nodes_pos) * 20,
+                endpoint=True,
+            )
 
-        fig.line(
-            "x",
-            "y0",
-            source=auxsource,
-            legend="Deformed shaft",
-            line_width=3,
-            line_color=bokeh_colors[9],
-        )
-        fig.circle(
-            "x",
-            "y0",
-            source=source,
-            legend="Deformed shaft",
-            size=8,
-            fill_color=bokeh_colors[9],
-        )
-        fig.line(
-            "x",
-            "y1",
-            source=source,
[…]
