### 2020-02-25 2a883ee
Adapt `.summary()` to `.run_static` sintaxe
