### 2020-02-25 8796043
Adapt `SummaryResults` to changes in `.summary()`
