### 2020-02-25 #2120 @whedon
[REVIEW]: ROSS - Rotordynamic Open Source Software
**Submitting author:** @raphaeltimbo (<a href="http://orcid.org/0000-0001-7689-5486">Raphael Silva</a>)
**Repository:** <a href="https://github.com/ross-rotordynamics/ross" target ="_blank">https://github.com/ross-rotordynamics/ross</a>
**Version:** v0.3.2
