### 2020-02-26 6ecbc19
Fix error in SummaryResults
* Fix sintaxe error
* Remove units from labels
* Add shaft number to the table columns
