### 2020-02-27 da7e7e2
Reduce decimal parameters on run_static() tests
