+++ ross/tests/test_rotor_assembly.py
@@ -741,17 +741,17 @@ def test_static_analysis_rotor3(rotor3):
         rotor3.Vx[0],
         np.array(
             [
-                -3.58824082e-12,
-                -4.94274533e02,
-                -4.56679111e02,
-                -4.19083689e02,
-                -9.94925251e01,
-                -6.18971030e01,
-                -2.43016808e01,
-                4.80980792e02,
-                5.18576214e02,
-                5.56171636e02,
-                -3.58824082e-12,
+                0,
+                -494.2745,
+                -456.6791,
+                -419.0836,
+                -99.4925,
+                -61.8971,
+                -24.3016,
+                480.9807,
+                518.5762,
+                556.1716,
+                0,
             ]
         ),
         decimal=3,
@@ -759,15 +759,7 @@ def test_static_analysis_rotor3(rotor3):
     assert_almost_equal(
         rotor3.Bm[0],
         np.array(
-            [
-                0.00000000e00,
-                -1.18869206e02,
-                -2.28339556e02,
-                -2.48513259e02,
-                -2.59288107e02,
-                -1.34343481e02,
-                -5.54223334e-12,
-            ]
+            [0, -118.8692, -228.3395, -248.5132, -259.2881, -134.3434, 0],
         ),
         decimal=3,
     )
@@ -825,21 +817,21 @@ def test_static_analysis_rotor5(rotor5):
         rotor5.Vx[0],
         np.array(
             [
-                0.00000000e00,
-                3.75954221e01,
-                7.51908442e01,
-                -4.94274533e02,
-                -4.56679111e02,
-                -4.19083689e02,
-                -9.94925251e01,
-                -6.18971030e01,
-                -2.43016808e01,
-                4.80980792e02,
-                5.18576214e02,
-                5.56171636e02,
-                -7.51908442e01,
-                -3.75954221e01,
-                -3.14059889e-12,
+                0,
+                37.5954,
+                75.1908,
+                -494.2745,
+                -456.6791,
+                -419.08368,
+                -99.4925,
+                -61.8971,
+                -24.3016,
+                480.9807,
+                518.5762,
+                556.1716,
+                -75.1908,
+                -37.5954,
+                0,
             ]
         ),
         decimal=3,
@@ -848,17 +840,17 @@ def test_static_analysis_rotor5(rotor5):
         rotor5.Bm[0],
         np.array(
             [
-                0.00000000e00,
-                4.69942776e00,
-                1.87977111e01,
-                -1.00071495e02,
-                -2.09541845e02,
-                -2.29715548e02,
-                -2.40490396e02,
-                -1.15545770e02,
-                1.87977111e01,
-                4.69942776e00,
-                -5.57065505e-12,
+                0,
+                4.6994,
+                18.7977,
+                -100.0714,
+                -209.5418,
+                -229.7155,
+                -240.4903,
+                -115.5457,
+                18.7977,
+                4.6994,
+                0,
             ]
         ),
         decimal=3,
@@ -918,21 +910,21 @@ def test_static_analysis_rotor6(rotor6):
         rotor6.Vx[0],
         np.array(
             [
-                0.00000000e+00,
-                3.75954221e+01,
-                7.51908442e+01,
-                -1.04154357e+02,
-                -6.65589353e+01,
-                -2.89635132e+01,
-                8.63190892e+00,
-                3.28223073e+02,
-                3.65818495e+02,
-                4.03413917e+02,
-                4.41009339e+02,
-                -5.80473317e+02,
-                -5.42877895e+02,
-                -5.05282473e+02,
-                2.20268248e-13
+                0,
+                37.5954,
+                75.1908,
+                -104.1543,
+                -66.5589,
+                -28.9635,
+                8.6319,
[…]
