### 2020-02-28 f20cb7d
Stiffness matrix documentation and accessory codes for the 6 DoF element
