### 2020-03-01 b1fbb0f
Adapting code to fit the ross.Report.generate_pdf() method
