### 2020-03-01 aa128fa
Organizing jinja2 HTML template.
