+++ ross/api_report.py
@@ -17,11 +17,15 @@ import ross as rs
 import bokeh.palettes as bp
 from bokeh.plotting import figure
 from bokeh.layouts import gridplot, widgetbox
-from bokeh.models.widgets import (
-    DataTable, NumberFormatter, TableColumn, Panel, Tabs
-)
+from bokeh.models.widgets import DataTable, NumberFormatter, TableColumn, Panel, Tabs
 from bokeh.models import (
-    ColumnDataSource, HoverTool, Span, Label, LinearAxis, Range1d, DataRange1d,
+    ColumnDataSource,
+    HoverTool,
+    Span,
+    Label,
+    LinearAxis,
+    Range1d,
+    DataRange1d,
 )
 
 # set bokeh palette of colors
@@ -48,17 +52,22 @@ class Report:
         ----------
         rotor: object
             A rotor built from rotor_assembly.
+
         maxspeed: float
             Maximum operation speed.
+
         minspeed: float
             Minimum operation speed.
+
         machine_type: str
             Machine type analyzed. Options: compressor, turbine or axial_flow.
             If other option is given, it will be treated as a compressor
             Default is compressor
+
         speed_units: str
             String defining the unit for rotor speed.
             Default is "rpm".
+
         tag: str
             String to name the rotor model
             Default is the Rotor.tag attribute
@@ -67,6 +76,7 @@ class Report:
         ----------
         rotor_type: str
             Defines if the rotor is between bearings or overhung
+
         disk_nodes: list
             List of disk between bearings or overhung (depending on the
             rotor type)
@@ -222,31 +232,39 @@ class Report:
             return "output_path should be a valid path"
 
         if output_filename is None:
-            output_filename = self.rotor.tag + '.pdf'
+            output_filename = self.rotor.tag + ".pdf"
 
         rotor = self.rotor
 
         template = "report.html"
-        root = Path(os.path.dirname(rs.__file__))/"API_Report"
+        root = Path(os.path.dirname(rs.__file__)) / "API_Report"
 
-        env = Environment(loader=FileSystemLoader(str(root/"templates")))
+        env = Environment(loader=FileSystemLoader(str(root / "templates")))
         template = env.get_template(template)
-        css = root/"static"/"css"
+        css = root / "static" / "css"
 
+        # Plotting rotor
         bokeh_fig = rotor.plot_rotor()
         bokeh_fig.plot_width = 500
         bokeh_fig.plot_height = 400
-        bokeh_fig.output_backend = 'svg'
-        bokeh.io.export_svgs(bokeh_fig, filename=Path(root/"assets") / 'plot_rotor.svg')
+        bokeh_fig.output_backend = "svg"
+        bokeh.io.export_svgs(
+            bokeh_fig, filename=Path(root / "assets") / "plot_rotor.svg"
+        )
 
-        template_vars = {"ASSETS_DIR": root/"assets", 'ROTOR_NAME': rotor.tag}
+        # Setting variables to jinja2 template
+        template_vars = {
+            "assets_dir": str(root / "assets"),
+            "machine_type": self.machine_type,
+            "tag": rotor.tag,
+        }
 
         # Used to render variables in template and generate PDF.
         rendered_string = template.render(template_vars)
         html = weasyprint.HTML(string=rendered_string)
         report = os.path.join(output_dir, output_filename)
-        html.write_pdf(report, stylesheets=[str(css/"report.css")])
-        print(f'ross generated in {output_dir}')
[…]
