### 2020-03-02 09f76e4
Correction according to commit 19433d7 and the change in the check_units decorator
