### 2020-03-02 2ad1fcf
Correction to the calculation of the cross-section geometrical properties
