### 2020-03-02 6dfb9e0
Proportional damping matrix implementation, with alpha and beta damping coefficients
