### 2020-03-02 0b2fe32
Gyroscopic matrix auxiliary formulation and documentation
