### 2020-03-03 603b3ba
Mass matrices for the 6 DoF formulation
