### 2020-03-03 f7efb5c
Stiffness and dynamic stiffness matrices for the 6 DoF formulation
