### 2020-03-03 a0c1983
Internal damping matrix for the 6 DoF formulation
