### 2020-03-03 a5785f1
Gyroscopic matrix for the 6 DoF formulation
