### 2020-03-03 e15273f
Additional corrections and minor adaptations
