+++ ross/shaft_element.py
@@ -1590,8 +1590,6 @@ class ShaftElement6DoF(Element):
                         n=parameters["n"][i],
                         axial_force=parameters["axial_force"][i],
                         torque=parameters["torque"][i],
-                        shear_effects=parameters["shear_effects"][i],
-                        rotary_inertia=parameters["rotary_inertia"][i],
                         gyroscopic=parameters["gyroscopic"][i],
                     )
                 )
@@ -1608,8 +1606,6 @@ class ShaftElement6DoF(Element):
                         n=parameters["n"][i],
                         axial_force=parameters["axial_force"][i],
                         torque=parameters["torque"][i],
-                        shear_effects=parameters["shear_effects"][i],
-                        rotary_inertia=parameters["rotary_inertia"][i],
                         gyroscopic=parameters["gyroscopic"][i],
                     )
                 )
@@ -1869,13 +1865,9 @@ class ShaftElement6DoF(Element):
         array(12x12)
         """
         
-        # Axial force applied to the element. 
-        # Units in Newtons, value defaulted to zero.
-        Fa = 0
-
-        # Torque force applied to the element. 
-        # Units in Newtons/m, value defaulted to zero.
-        T = 0
+        # Axial force and torque applied to the element. 
+        Fa = self.axial_force
+        T = self.torque
 
         # temporary material and geometrical constants, determined as mean values 
         # from the left and right radii of the taperad shaft
@@ -2052,25 +2044,26 @@ class ShaftElement6DoF(Element):
         array(12x12)
         """
         
-        # temporary material and geometrical constants, determined as mean values 
-        # from the left and right radii of the tapered shaft
-        L = self.L
-        tempI = np.pi / 4 * ( (self.odr/2)**4 - (self.odl/2)**4 - (self.idr/2)**4 + (self.idl/2)**4 )
+        if self.gyroscopic:
+            # temporary material and geometrical constants, determined as mean values 
+            # from the left and right radii of the tapered shaft
+            L = self.L
+            tempI = np.pi / 4 * ( (self.odr/2)**4 - (self.odl/2)**4 - (self.idr/2)**4 + (self.idl/2)**4 )
 
-        G = (self.material.rho * tempI / (15 * L)) * np.array[
-            [   0, -36, 0,  -3*L,      0, 0,    0,   36, 0,  -3*L,      0, 0]
-            [  36,   0, 0,     0,   -3*L, 0,  -36,    0, 0,     0,   -3*L, 0]
-            [   0,   0, 0,     0,      0, 0,    0,    0, 0,     0,      0, 0]
-            [ 3*L,   0, 0,     0, -4*L^2, 0, -3*L,    0, 0,     0,    L^2, 0]
-            [   0, 3*L, 0, 4*L^2,      0, 0,    0, -3*L, 0,  -L^2,      0, 0]
-            [   0,   0, 0,     0,      0, 0,    0,    0, 0,     0,      0, 0]
-            [   0,  36, 0,   3*L,      0, 0,    0,  -36, 0,   3*L,      0, 0]
-            [ -36,   0, 0,     0,    3*L, 0,   36,    0, 0,     0,    3*L, 0]
-            [   0,   0, 0,     0,      0, 0,    0,    0, 0,     0,      0, 0]
-            [ 3*L,   0, 0,     0,    L^2, 0, -3*L,    0, 0,     0, -4*L^2, 0]
-            [   0, 3*L, 0,  -L^2,      0, 0,    0, -3*L, 0, 4*L^2,      0, 0]
-            [   0,   0, 0,     0,      0, 0,    0,    0, 0,     0,      0, 0]
-        ]
+            G = (self.material.rho * tempI / (15 * L)) * np.array[
[…]
