### 2020-03-03 1c4efe2
Black formatting applied to code
