+++ ross/shaft_element.py
@@ -13,7 +13,7 @@ from ross.materials import Material, steel
 from ross.utils import read_table_file
 from ross.units import check_units
 
-__all__ = ["ShaftElement","ShaftElement6DoF"]
+__all__ = ["ShaftElement", "ShaftElement6DoF"]
 bokeh_colors = bp.RdGy[11]
 
 
@@ -1203,9 +1203,6 @@ class ShaftElement(Element):
         return elements
 
 
-
-
-
 class ShaftElement6DoF(Element):
 
     r"""A 6 Degrees of Freedom shaft element.
@@ -1343,11 +1340,11 @@ class ShaftElement6DoF(Element):
             odr = odl
 
         ## Changing units to only their magnitude
-        #L = L.m
-        #idl = idl.m
-        #odl = odl.m
-        #idr = idr.m
-        #odr = odr.m
+        # L = L.m
+        # idl = idl.m
+        # odl = odl.m
+        # idr = idr.m
+        # odr = odr.m
 
         if material is None:
             raise AttributeError("Material is not defined.")
@@ -1384,21 +1381,21 @@ class ShaftElement6DoF(Element):
         self.alpha = float(alpha)
         self.beta = float(beta)
 
-       # # A_l = cross section area from the left side of the element
-       # # A_r = cross section area from the right side of the element
-       # A_l = np.pi * (odl ** 2 - idl ** 2) / 4
-       # A_r = np.pi * (odr ** 2 - idr ** 2) / 4
-       # self.A_l = A_l
-       # self.A_r = A_r
+        # # A_l = cross section area from the left side of the element
+        # # A_r = cross section area from the right side of the element
+        # A_l = np.pi * (odl ** 2 - idl ** 2) / 4
+        # A_r = np.pi * (odr ** 2 - idr ** 2) / 4
+        # self.A_l = A_l
+        # self.A_r = A_r
 
-       # # Second moment of area of the cross section from the left side
-       # # of the element
-       # Ie_l = np.pi * (odl ** 4 - idl ** 4) / 64
+        # # Second moment of area of the cross section from the left side
+        # # of the element
+        # Ie_l = np.pi * (odl ** 4 - idl ** 4) / 64
 
-       # outer = self.odl ** 2 + self.odl * self.odr + self.odr ** 2
-       # inner = self.idl ** 2 + self.idl * self.idr + self.idr ** 2
-       # self.volume = np.pi * (self.L / 12) * (outer - inner)
-       # self.m = self.material.rho * self.volume
+        # outer = self.odl ** 2 + self.odl * self.odr + self.odr ** 2
+        # inner = self.idl ** 2 + self.idl * self.idr + self.idr ** 2
+        # self.volume = np.pi * (self.L / 12) * (outer - inner)
+        # self.m = self.material.rho * self.volume
 
         # Timoshenko kappa factor determination, based on the diameters relation
         if self.__is_circular():
@@ -1421,15 +1418,12 @@ class ShaftElement6DoF(Element):
 
         self.kappa = kappa
 
-
     def __is_circular(self):
         return self.shear_effects and self.i_d == 0
-
 
     def __is_thickwall(self):
         p = (self.o_d - self.i_d) / self.o_d
         return self.shear_effects and p >= 0.2
-
 
     def __eq__(self, other):
         """
@@ -1502,7 +1496,6 @@ class ShaftElement6DoF(Element):
             "gyroscopic": self.gyroscopic,
         }
         self.dump_data(data, file_name)
-        
 
     @staticmethod
     def load(file_name="ShaftElement"):
[…]
