### 2020-03-04 917da41
Improvements to api_report.py
* New arguments to class Report;
* New methods: .run() and .rotor_instance()
* .unbalance_response() returns a dictionary with info about separation margin, amplification factor, etc
* .unbalance_response() returns the phase vs frequency plot to complement the magnitude plot.
* Add report_example() to simplify docstrings
* Fix a bug in .stability_level_2() where the dataframe would receive repetitive data
