+++ ross/api_report.py
@@ -28,34 +28,44 @@ class Report:
     def __init__(
         self,
         rotor,
-        minspeed,
-        maxspeed,
+        speed_range,
+        tripspeed,
+        bearing_stiffness_range=None,
+        bearing_clearance_lists=None,
         machine_type="compressor",
         speed_units="rpm",
         tag=None,
     ):
         """Report according to standard analysis.
 
+        - Perform unbalance response
         - Perform Stability_level1 analysis
         - Apply Level 1 Screening Criteria
         - Perform Stability_level2 analysis
 
         Parameters
         ----------
-        rotor: object
+        rotor : object
             A rotor built from rotor_assembly.
-        maxspeed: float
-            Maximum operation speed.
-        minspeed: float
-            Minimum operation speed.
-        machine_type: str
+        speed_range : tuple
+            Tuple with (min, max) for speed range.
+        tripspeed : float
+            Machine trip speed.
+        bearing_stiffness_range : tuple, optional
+            Tuple with (start, end) bearing stiffness range.
+            Argument to calculate the Undamped Critical Speed Map.
+        bearing_clearance_lists : list of lists, optional
+            List with two bearing elements lists:
+                The first bearing list is set for minimum clearance.
+                The second bearing list it set for maximum clearance.
+        machine_type : str
             Machine type analyzed. Options: compressor, turbine or axial_flow.
             If other option is given, it will be treated as a compressor
             Default is compressor
-        speed_units: str
+        speed_units : str
             String defining the unit for rotor speed.
             Default is "rpm".
-        tag: str
+        tag : str
             String to name the rotor model
             Default is the Rotor.tag attribute
 
@@ -74,22 +84,47 @@ class Report:
         Examples
         --------
         >>> rotor = rotor_example()
+        >>> 
+        >>> # coefficients for minimum clearance
+        >>> stfx = [0.7e7, 0.8e7, 0.9e7, 1.0e7]
+        >>> dampx = [2.0e3, 1.9e3, 1.8e3, 1.7e3]
+        >>> freq = [400, 800, 1200, 1600]
+        >>> bearing0 = BearingElement(0, kxx=stfx, cxx=dampx, frequency=freq)
+        >>> bearing1 = BearingElement(6, kxx=stfx, cxx=dampx, frequency=freq)
+        >>> min_clearance_brg = [bearing0, bearing1]
+        >>> 
+        >>> # coefficients for maximum clearance
+        >>> stfx = [0.4e7, 0.5e7, 0.6e7, 0.7e7]
+        >>> dampx = [2.8e3, 2.7e3, 2.6e3, 2.5e3]
+        >>> freq = [400, 800, 1200, 1600]
+        >>> bearing0 = BearingElement(0, kxx=stfx, cxx=dampx, frequency=freq)
+        >>> bearing1 = BearingElement(6, kxx=stfx, cxx=dampx, frequency=freq)
+        >>> max_clearance_brg = [bearing0, bearing1]
+        >>>
+        >>> bearings = [min_clearance_brg, max_clearance_brg]
         >>> report = Report(rotor=rotor,
-        ...                 minspeed=400,
-        ...                 maxspeed=1000,
+        ...                 speed_range=(400, 1000),
+        ...                 tripspeed=1200,
+        ...                 bearing_stiffness_range=(5,8),
+        ...                 bearing_clearance_lists=bearings,
         ...                 speed_units="rad/s")
         >>> report.rotor_type
         'between_bearings'
         """
         self.rotor = rotor
         self.speed_units = speed_units
+        self.speed_range = speed_range
 
         if speed_units == "rpm":
-            minspeed = minspeed * np.pi / 30
[…]
