### 2020-03-04 0f40293
adapt test to new class Report sintaxe
