+++ ross/api_report.py
@@ -13,11 +13,15 @@ from copy import deepcopy
 import bokeh.palettes as bp
 from bokeh.plotting import figure
 from bokeh.layouts import gridplot, widgetbox
-from bokeh.models.widgets import (
-    DataTable, NumberFormatter, TableColumn, Panel, Tabs
-)
+from bokeh.models.widgets import DataTable, NumberFormatter, TableColumn, Panel, Tabs
 from bokeh.models import (
-    ColumnDataSource, HoverTool, Span, Label, LinearAxis, Range1d, DataRange1d,
+    ColumnDataSource,
+    HoverTool,
+    Span,
+    Label,
+    LinearAxis,
+    Range1d,
+    DataRange1d,
 )
 
 # set bokeh palette of colors
@@ -197,15 +201,15 @@ class Report:
 
     @classmethod
     def from_saved_rotors(
-            cls,
-            path,
-            speed_range,
-            tripspeed,
-            bearing_stiffness_range=None,
-            bearing_clearance_lists=None,
-            machine_type="compressor",
-            speed_units="rpm",
-            tag=None,
+        cls,
+        path,
+        speed_range,
+        tripspeed,
+        bearing_stiffness_range=None,
+        bearing_clearance_lists=None,
+        machine_type="compressor",
+        speed_units="rpm",
+        tag=None,
     ):
         """Instantiate a rotor from a previously saved rotor model
 
@@ -244,7 +248,7 @@ class Report:
             bearing_clearance_lists,
             machine_type,
             speed_units,
-            tag
+            tag,
         )
 
     def rotor_instance(self, rotor, bearing_list):
@@ -275,7 +279,7 @@ class Report:
         >>> report = report_example(rotor)
         >>> aux_rotor = report.rotor_instance(rotor, bearings)
         """
-        
+
         sh_elm = rotor.shaft_elements
         dk_elm = rotor.disk_elements
         pm_elm = rotor.point_mass_elements
@@ -286,10 +290,20 @@ class Report:
         rated_w = rotor.rated_w
         tag = rotor.tag
 
-        aux_rotor = Rotor(sh_elm, dk_elm, bearing_list, pm_elm, sparse, n_eigen, min_w, max_w, rated_w, tag)
+        aux_rotor = Rotor(
+            sh_elm,
+            dk_elm,
+            bearing_list,
+            pm_elm,
+            sparse,
+            n_eigen,
+            min_w,
+            max_w,
+            rated_w,
+            tag,
+        )
 
         return aux_rotor
-
 
     def run(self, D, H, HP, oper_speed, RHO_ratio, RHOs, RHOd, unit="m"):
         """Run API report.
@@ -399,7 +413,16 @@ class Report:
         self.rotor = rotor0
         print(rotor0.bearing_elements)
 
-        return fig_ucs, fig_mode_shape, fig_unbalance, df_unbalance, fig_a_lvl1, fig_b_lvl1, df_lvl2, summaries
+        return (
+            fig_ucs,
+            fig_mode_shape,
+            fig_unbalance,
+            df_unbalance,
+            fig_a_lvl1,
+            fig_b_lvl1,
+            df_lvl2,
+            summaries,
+        )
 
     def plot_ucs(self, stiffness_range=None, num=20):
         """Plot undamped critical speed map.
@@ -448,7 +471,10 @@ class Report:
         for i, k in enumerate(stiffness_log):
             bearings = [BearingElement(b.n, kxx=k, cxx=0) for b in bearings_elements]
             rotor = self.rotor.__class__(
-                self.rotor.shaft_elements, self.rotor.disk_elements, bearings, n_eigen=16
+                self.rotor.shaft_elements,
+                self.rotor.disk_elements,
+                bearings,
+                n_eigen=16,
             )
[…]
