### 2020-03-04 d92d057
Fix `.mode_shape()` and `.summary()` docstrings
