### 2020-03-04 47c3d0b
Initial testing procedures and function declarations for 6DoF shaft element
