### 2020-03-04 d647944
Change from E-nu to E-G as default material properties for 6DoF shaft elements
