### 2020-03-04 64f4487
Timoshenko coefficient determination update to tapered shaft variables
