### 2020-03-04 ff61073
6DoF shaft test matrices addition
