+++ ross/tests/test_shaft_element.py
@@ -324,19 +324,7 @@ def s6_eb():
     beta = 4.8 * 10 ** (-6)
 
     return ShaftElement6DoF(
-        L,
-        idl,
-        odl,
-        idr,
-        odr,
-        steel,
-        axial_force,
-        torque,
-        # shear_effects=True,
-        # rotary_inertia=True,
-        # gyroscopic=True,
-        alpha,
-        beta,
+        L, idl, odl, idr, odr, steel, axial_force, torque, alpha, beta,
     )
 
 
@@ -345,7 +333,6 @@ def s6_test_index(eb6):
 
 
 def s6_test_parameters_eb(eb6):
-    assert eb6.phi == 0
     assert eb6.L == 0.1
     assert eb6.idl == 25.4 * 1 / 2
     assert eb6.odl == 25.4 * 3 / 4
@@ -354,23 +341,84 @@ def s6_test_parameters_eb(eb6):
     assert eb6.material.E == 211e9
     assert eb6.material.G_s == 81.2e9
     assert eb6.material.rho == 7810
-    assert_almost_equal(eb6.material.Poisson, 0.29926108)
-    assert_almost_equal(eb6.A, 0.00196349)
-    assert_almost_equal(eb6.Ie * 1e7, 3.06796157)
+    assert_almost_equal(eb6.a1, 8.9289e012)
+    assert_almost_equal(eb6.a2, 8.4249e015)
+    assert_almost_equal(eb6.a3, 3.3253e014)
 
 
 def s6_test_M_matrix(eb6):
-    pass
+    # fmt: off
+    M_mat_6DoF = np.array(
+        [[  4.8624e+08,            0,            0,            0,  -4.0523e+06,            0,  -4.8618e+08,            0,            0,            0,  -4.0512e+06,            0],
+         [           0,   4.8624e+08,            0,   4.0523e+06,            0,            0,            0,  -4.8618e+08,            0,   4.0512e+06,            0,            0],
+         [           0,            0,   4.1223e+04,            0,            0,            0,            0,            0,   2.0611e+04,            0,            0,            0],
+         [           0,   4.0523e+06,            0,   5.4023e+05,            0,            0,            0,  -4.0512e+06,            0,  -1.3506e+05,            0,            0],
+         [ -4.0523e+06,            0,            0,            0,   5.4023e+05,            0,   4.0512e+06,            0,            0,            0,  -1.3506e+05,            0],
+         [           0,            0,            0,            0,            0,   1.3505e+06,            0,            0,            0,            0,            0,   6.7527e+05],
+         [ -4.8618e+08,            0,            0,            0,   4.0512e+06,            0,   4.8624e+08,            0,            0,            0,   4.0523e+06,            0],
+         [           0,  -4.8618e+08,            0,  -4.0512e+06,            0,            0,            0,   4.8624e+08,            0,  -4.0523e+06,            0,            0],
+         [           0,            0,   2.0611e+04,            0,            0,            0,            0,            0,   4.1223e+04,            0,            0,            0],
+         [           0,   4.0512e+06,            0,  -1.3506e+05,            0,            0,            0,  -4.0523e+06,            0,   5.4023e+05,            0,            0],
+         [ -4.0512e+06,            0,            0,            0,  -1.3506e+05,            0,   4.0523e+06,            0,            0,            0,   5.4023e+05,            0],
+         [           0,            0,            0,            0,            0,   6.7527e+05,            0,            0,            0,            0,            0,   1.3505e+06]])
+    # fmt: on
+    assert_allclose(eb.M(), M_mat_6DoF, rtol=1e-3)
 
 
 def s6_test_G_matrix(eb6):
-    pass
+    # fmt: off
+    G_mat_6DoF = np.array(
[…]
