### 2020-03-04 8514c17
Modifications to report_example()
* Improve report_example() to simplify the doctests
* Fix a bug in `.stability_level_2()` where the dataframe would repeat a row if a rotor had only 1 disk or only 1 seal.
* Comment doctest from `.run()` because it would take some time to run.
