+++ ross/api_report.py
@@ -6,7 +6,9 @@ from scipy.interpolate import interp1d
 from scipy.signal import argrelextrema
 from pathlib import Path
 from ross.rotor_assembly import Rotor, rotor_example
-from ross.bearing_seal_element import BearingElement
+from ross.bearing_seal_element import BearingElement, SealElement
+from materials import steel
+
 import ross as rs
 
 from copy import deepcopy
@@ -120,13 +122,14 @@ class Report:
         self.speed_range = speed_range
 
         if speed_units == "rpm":
-            minspeed = speed_range[0] * np.pi / 30
-            maxspeed = speed_range[1] * np.pi / 30
-            tripspeed = tripspeed * np.pi / 30
+            self.minspeed = speed_range[0] * np.pi / 30
+            self.maxspeed = speed_range[1] * np.pi / 30
+            self.tripspeed = tripspeed * np.pi / 30
+        if speed_units == "rad/s":
+            self.minspeed = speed_range[0]
+            self.maxspeed = speed_range[1]
+            self.tripspeed = tripspeed
 
-        self.minspeed = speed_range[0]
-        self.maxspeed = speed_range[1]
-        self.tripspeed = tripspeed
         self.bearing_stiffness_range = bearing_stiffness_range
         self.bearing_clearance_lists = bearing_clearance_lists
 
@@ -276,7 +279,7 @@ class Report:
         >>> bearing1 = BearingElement(6, kxx=stfx, cxx=dampx, frequency=freq)
         >>> bearings = [bearing0, bearing1]
         >>> rotor = rotor_example()
-        >>> report = report_example(rotor)
+        >>> report = report_example()
         >>> aux_rotor = report.rotor_instance(rotor, bearings)
         """
 
@@ -356,8 +359,7 @@ class Report:
 
         Example
         -------
-        >>> rotor = rotor_example()
-        >>> report = report_example(rotor)
+        >>> report = report_example()
         >>> D = [0.35, 0.35]
         >>> H = [0.08, 0.08]
         >>> HP = [10000, 10000]
@@ -365,8 +367,8 @@ class Report:
         >>> RHOd = 30.45
         >>> RHOs = 37.65
         >>> oper_speed = 1000.0
-        >>> report.run(D, H, HP, oper_speed, RHO_ratio, RHOs, RHOd) # doctest: +ELLIPSIS
-        ([Figure...
+        >>> # to run the API report analysis, use:
+        >>> # report.run(D, H, HP, oper_speed, RHO_ratio, RHOs, RHOd)
         """
         fig_ucs = []
         fig_mode_shape = []
@@ -442,8 +444,7 @@ class Report:
 
         Example
         -------
-        >>> rotor = rotor_example()
-        >>> report = report_example(rotor)
+        >>> report = report_example()
         >>> report.plot_ucs(stiffness_range=(5, 8)) # doctest: +ELLIPSIS
         Figure...
         """
@@ -461,7 +462,7 @@ class Report:
 
         bearings_elements = []  # exclude the seals
         for bearing in self.rotor.bearing_elements:
-            if type(bearing) == BearingElement:
+            if not isinstance(bearing, SealElement):
                 bearings_elements.append(bearing)
 
         for i, k in enumerate(stiffness_log):
@@ -564,8 +565,7 @@ class Report:
 
         Example
         -------
-        >>> rotor = rotor_example()
-        >>> report = report_example(rotor)
+        >>> report = report_example()
         >>> report.static_forces()
         array([44.09320349, 44.09320349])
         """
@@ -603,8 +603,7 @@ class Report:
 
         Example
         -------
-        >>> rotor = rotor_example()
[…]
