### 2020-03-04 d4e6b1c
Fix tests for api_report.py
