+++ ross/tests/test_api_report.py
@@ -32,11 +32,10 @@ def report0():
         n=35, material=steel, width=0.07, i_d=0.05, o_d=0.35
     )
 
-    stfx = [0.4e7, 0.5e7, 0.6e7, 0.7e7]
-    stfy = [0.8e7, 0.9e7, 1.0e7, 1.1e7]
-    freq = [400, 800, 1200, 1600]
-    bearing0 = BearingElement(0, kxx=stfx, kyy=stfy, cxx=2e3, frequency=freq)
-    bearing1 = BearingElement(50, kxx=stfx, kyy=stfy, cxx=2e3, frequency=freq)
+    stfx = 1e6
+    stfy = 0.8e6
+    bearing0 = BearingElement(0, kxx=stfx, kyy=stfy, cxx=1000)
+    bearing1 = BearingElement(50, kxx=stfx, kyy=stfy, cxx=1000)
 
     rotor = Rotor(shaft_elem, [disk0, disk1], [bearing0, bearing1])
 
@@ -94,11 +93,10 @@ def report1():
         n=0, material=steel, width=0.07, i_d=0.05, o_d=0.28
     )
 
-    stfx = [0.4e7, 0.5e7, 0.6e7, 0.7e7]
-    stfy = [0.8e7, 0.9e7, 1.0e7, 1.1e7]
-    freq = [400, 800, 1200, 1600]
-    bearing0 = BearingElement(15, kxx=stfx, kyy=stfy, cxx=2e3, frequency=freq)
-    bearing1 = BearingElement(50, kxx=stfx, kyy=stfy, cxx=2e3, frequency=freq)
+    stfx = 1e6
+    stfy = 0.8e6
+    bearing0 = BearingElement(15, kxx=stfx, kyy=stfy, cxx=1000)
+    bearing1 = BearingElement(50, kxx=stfx, kyy=stfy, cxx=1000)
 
     rotor = Rotor(shaft_elem, [disk0], [bearing0, bearing1])
 
@@ -159,11 +157,10 @@ def report2():
         n=50, material=steel, width=0.07, i_d=0.05, o_d=0.35
     )
 
-    stfx = [0.4e7, 0.5e7, 0.6e7, 0.7e7]
-    stfy = [0.8e7, 0.9e7, 1.0e7, 1.1e7]
-    freq = [400, 800, 1200, 1600]
-    bearing0 = BearingElement(15, kxx=stfx, kyy=stfy, cxx=2e3, frequency=freq)
-    bearing1 = BearingElement(35, kxx=stfx, kyy=stfy, cxx=2e3, frequency=freq)
+    stfx = 1e6
+    stfy = 0.8e6
+    bearing0 = BearingElement(15, kxx=stfx, kyy=stfy, cxx=1000)
+    bearing1 = BearingElement(35, kxx=stfx, kyy=stfy, cxx=1000)
 
     rotor = Rotor(shaft_elem, [disk0, disk1], [bearing0, bearing1])
 
@@ -256,23 +253,33 @@ def test_unbalance_forces(report0, report1, report2):
 
 
 def test_report_mode_shape(report0, report1, report2):
-    n1, n2 = report0.mode_shape(mode=0)
+    _ = report0.mode_shape(mode=0)
+    n1 = report0.node_min
+    n2 = report0.node_max
     nodes = [int(node) for sub_nodes in [n1, n2] for node in sub_nodes]
     assert nodes == [26]
 
-    n1, n2 = report1.mode_shape(mode=0)
+    _ = report1.mode_shape(mode=0)
+    n1 = report1.node_min
+    n2 = report1.node_max
     nodes = [int(node) for sub_nodes in [n1, n2] for node in sub_nodes]
     assert nodes == [0]
 
[…]
