### 2020-03-04 828e23e
Add example to docstring from report_example()
