### 2020-03-04 4112683
Fix typo in report_example()
