### 2020-03-05 183b440
Improvements to stochastic decorator
* Change the method name to "random_var()";
* Add input argument ("is_random") to the user to choose which arguments from the selected class/method should be marked as random;
* Add docstring with example;
* Change the name of some variables to make it clearer;
