### 2020-03-05 ross-rotordynamics/ross#458 @srmnitc
Binder examples error
This is part of my review of the software paper at openjournals/joss-reviews#2120

While trying out the examples on binder, evaluating the first cell in the notebook `tutorial.ipynb`, the following error is generated-

```
FileNotFoundError                         Traceback (most recent call last)
/srv/conda/envs/notebook/lib/python3.7/site-packages/pint/registry.py in load_definitions(self, file, is_resource)
    523                 else:
--> 524                     with open(file, encoding="utf-8") as fp:
    525                         return self.load_definitions(fp, is_resource)

FileNotFoundError: [Errno 2] No such file or directory: '/srv/conda/envs/notebook/lib/python3.7/site-packages/ross/new_units.txt'
```
