### 2020-03-05 ross-rotordynamics/ross#458 @raphaeltimbo
Hi there @srmnitc !

This error is related to a new feature that we have in the master branch but is not present in the v0.2.0 (which is the released version submitted to JOSS).

We have already noticed some issues regarding the current released version (v0.2.0) and parts of the documentation that are tracking the master branch.
Is it ok if we make a new release (v0.3.0) and update the JOSS issue information below?

![image](https://user-images.githubusercontent.com/18506378/75988554-883b0800-5ed0-11ea-87f5-0631f73dccad.png)
