### 2020-03-05 4a76812
Add imports to `.random_var()` docstring example
