### 2020-03-05 c74bb64
Add `# doctest: +ELLIPSIS` to random_var() example
