### 2020-03-05 ross-rotordynamics/ross#458 @srmnitc
@raphaeltimbo Thanks for the quick reply! From what I have seen, I think its fine to change the version. The journal would provide you an option in the end to change version number. I was able to do this for my own publication. I think getting the examples working on binder is a great step! Thanks!
