### 2020-03-05 7546bb2
Change fluid_flow_test to avoid gca()
When running the tests (fluid_flow_graphics.py + test_fluid_flow.py)
our code can sometimes run the function plt.gca(), which is supposed
to get the current matplotlib axis in case the user does not provide
an axes. The problem with this is that, when testing, a function
could get a matplotlib axes with a polar projection (resulting from
another test) and this could result in error on our function.
