+++ ross/tests/test_fluid_flow.py
@@ -52,7 +52,7 @@ def fluid_flow_short_eccentricity():
         visc,
         rho,
         eccentricity=eccentricity,
-        immediately_calculate_pressure_matrix_numerically=False
+        immediately_calculate_pressure_matrix_numerically=False,
     )
 
 
@@ -84,7 +84,7 @@ def fluid_flow_short_friswell(set_load=True):
             visc,
             rho,
             load=load,
-            immediately_calculate_pressure_matrix_numerically=False
+            immediately_calculate_pressure_matrix_numerically=False,
         )
     else:
         return flow.FluidFlow(
@@ -100,7 +100,7 @@ def fluid_flow_short_friswell(set_load=True):
             visc,
             rho,
             eccentricity=eccentricity,
-            immediately_calculate_pressure_matrix_numerically=False
+            immediately_calculate_pressure_matrix_numerically=False,
         )
 
 
@@ -131,9 +131,7 @@ def test_stiffness_matrix():
     Taken from example 5.5.1, page 181 (Dynamics of rotating machine, FRISSWELL)
     """
     bearing = fluid_flow_short_friswell()
-    kxx, kxy, kyx, kyy = calculate_stiffness_matrix(
-        bearing, force_type='short'
-    )
+    kxx, kxy, kyx, kyy = calculate_stiffness_matrix(bearing, force_type="short")
     assert math.isclose(kxx / 10 ** 6, 12.81, rel_tol=0.01)
     assert math.isclose(kxy / 10 ** 6, 16.39, rel_tol=0.01)
     assert math.isclose(kyx / 10 ** 6, -25.06, rel_tol=0.01)
@@ -149,10 +147,8 @@ def test_stiffness_matrix_numerical(fluid_flow_short_eccentricity):
     """
     bearing = fluid_flow_short_eccentricity
     bearing.calculate_pressure_matrix_numerical()
-    kxx, kxy, kyx, kyy = calculate_stiffness_matrix(bearing, force_type='numerical')
-    k_xx, k_xy, k_yx, k_yy = calculate_stiffness_matrix(
-        bearing, force_type='short'
-    )
+    kxx, kxy, kyx, kyy = calculate_stiffness_matrix(bearing, force_type="numerical")
+    k_xx, k_xy, k_yx, k_yy = calculate_stiffness_matrix(bearing, force_type="short")
     assert_allclose(kxx, k_xx, rtol=0.33)
     assert_allclose(kxy, k_xy, rtol=0.21)
     assert_allclose(kyx, k_yx, rtol=0.17)
@@ -166,9 +162,7 @@ def test_damping_matrix():
     Taken from example 5.5.1, page 181 (Dynamics of rotating machine, FRISSWELL)
     """
     bearing = fluid_flow_short_friswell()
-    cxx, cxy, cyx, cyy = calculate_damping_matrix(
-        bearing, force_type='short'
-    )
+    cxx, cxy, cyx, cyy = calculate_damping_matrix(bearing, force_type="short")
     assert math.isclose(cxx / 10 ** 3, 232.9, rel_tol=0.01)
     assert math.isclose(cxy / 10 ** 3, -81.92, rel_tol=0.01)
     assert math.isclose(cyx / 10 ** 3, -81.92, rel_tol=0.01)
@@ -203,7 +197,7 @@ def fluid_flow_short_numerical():
         rho,
         attitude_angle=attitude_angle,
         eccentricity=eccentricity,
-        immediately_calculate_pressure_matrix_numerically=False
+        immediately_calculate_pressure_matrix_numerically=False,
     )
 
 
@@ -234,7 +228,7 @@ def fluid_flow_long_numerical():
         visc,
         rho,
         eccentricity=eccentricity,
-        immediately_calculate_pressure_matrix_numerically=False
[…]
