### 2020-03-06 ross-rotordynamics/ross#460 @nnadeau
Add Statement of Need
Please add a Statement of Need to your README/documentation. It's preferable that this is one of the first things a reader would see in the README.

cc https://github.com/openjournals/joss-reviews/issues/2120
