### 2020-03-06 ross-rotordynamics/ross#461 @nnadeau
Use a standard/common approach to declaring dependencies
Currently, the dependencies are defined in `setup.py`:

```python
REQUIRED = [
    "numpy",
    "scipy",
    "matplotlib",
    "toml",
    "pandas>=0.23",
    "bokeh",
    "coverage",
    "xlrd",
    "pint",
]
```

In the Python ecosystem, dependencies are typically defined in a `requirements.txt` file or using a dependency management framework (e.g., poetry, pipenv).

It would be preferable that repo be formatted such that there is a separation of concerns between the `setup.py` and the declaration of dependencies.

If this repo is intended as a library (i.e., integrated into other packages), then a `requirements.txt` with minimum versions declared is preferred.

If this repo is intended as an end-use application (i.e., not integrated into other packages), then using `poetry` or `pipenv` is preferred.

cc https://github.com/openjournals/joss-reviews/issues/2120
