### 2020-03-06 ross-rotordynamics/ross#462 @nnadeau
pytest is not listed as a dep
Can't test using the testing instructions: https://ross-rotordynamics.github.io/ross-website/contributing.html#tests

cc https://github.com/openjournals/joss-reviews/issues/2120
