### 2020-03-06 ross-rotordynamics/ross#463 @nnadeau
Lint codebase using flake8
It's important in OSS projects to have the codebase conform to a style guide. In Python, `flake8` is one of the de facto standard. The following issues should be fixed and `flake8` should be added to the testing suite.

```
ross/results.py:8:80: E501 line too long (88 > 79 characters)
ross/results.py:9:80: E501 line too long (85 > 79 characters)
ross/results.py:69:80: E501 line too long (80 > 79 characters)
ross/results.py:73:80: E501 line too long (97 > 79 characters)
ross/results.py:182:40: E203 whitespace before ':'
ross/results.py:315:80: E501 line too long (81 > 79 characters)
ross/results.py:456:13: F401 'mpl_toolkits.mplot3d.Axes3D' imported but unused
ross/results.py:463:80: E501 line too long (83 > 79 characters)
ross/results.py:497:16: W605 invalid escape sequence '\o'
ross/results.py:499:21: W605 invalid escape sequence '\_'
ross/results.py:701:80: E501 line too long (81 > 79 characters)
ross/results.py:709:80: E501 line too long (86 > 79 characters)
ross/results.py:712:80: E501 line too long (87 > 79 characters)
ross/results.py:715:80: E501 line too long (84 > 79 characters)
ross/results.py:718:80: E501 line too long (83 > 79 characters)
ross/results.py:773:80: E501 line too long (86 > 79 characters)
ross/results.py:836:80: E501 line too long (87 > 79 characters)
ross/results.py:895:80: E501 line too long (84 > 79 characters)
ross/results.py:896:80: E501 line too long (83 > 79 characters)
ross/results.py:897:80: E501 line too long (86 > 79 characters)
ross/results.py:988:80: E501 line too long (88 > 79 characters)
ross/results.py:1026:32: W605 invalid escape sequence '\o'
ross/results.py:1061:41: W605 invalid escape sequence '\m'
ross/results.py:1168:80: E501 line too long (80 > 79 characters)
ross/results.py:1407:40: W605 invalid escape sequence '\m'
ross/results.py:1576:35: W291 trailing whitespace
ross/results.py:1700:80: E501 line too long (93 > 79 characters)
ross/results.py:1743:80: E501 line too long (87 > 79 characters)
ross/results.py:1922:37: E131 continuation line unaligned for hanging indent
ross/results.py:2071:32: E203 whitespace before ':'
ross/results.py:2071:47: E203 whitespace before ':'
ross/results.py:2078:80: E501 line too long (80 > 79 characters)
ross/results.py:2146:80: E501 line too long (81 > 79 characters)
ross/results.py:2426:80: E501 line too long (84 > 79 characters)
ross/results.py:2427:80: E501 line too long (80 > 79 characters)
ross/results.py:2447:80: E501 line too long (87 > 79 characters)
ross/results.py:2527:80: E501 line too long (86 > 79 characters)
ross/results.py:2574:80: E501 line too long (84 > 79 characters)
ross/results.py:2668:13: F401 'mpl_toolkits.mplot3d.Axes3D' imported but unused
ross/api_report.py:7:1: F401 'pathlib.Path' imported but unused
ross/api_report.py:8:1: F401 'ross.rotor_assembly.rotor_example' imported but unused
ross/api_report.py:15:1: F401 'bokeh.layouts.gridplot' imported but unused
ross/api_report.py:97:80: E501 line too long (81 > 79 characters)
ross/api_report.py:108:80: E501 line too long (81 > 79 characters)
ross/api_report.py:113:80: E501 line too long (81 > 79 characters)
ross/api_report.py:116:80: E501 line too long (81 > 79 characters)
ross/api_report.py:121:80: E501 line too long (81 > 79 characters)
ross/api_report.py:124:80: E501 line too long (81 > 79 characters)
ross/api_report.py:187:80: E501 line too long (81 > 79 characters)
ross/api_report.py:329:80: E501 line too long (80 > 79 characters)
ross/api_report.py:354:80: E501 line too long (89 > 79 characters)
ross/api_report.py:363:80: E501 line too long (87 > 79 characters)
ross/api_report.py:371:80: E501 line too long (81 > 79 characters)
ross/api_report.py:475:17: F841 local variable 'SM_ref' is assigned to but never used
ross/api_report.py:699:80: E501 line too long (87 > 79 characters)
ross/api_report.py:782:80: E501 line too long (85 > 79 characters)
ross/api_report.py:822:80: E501 line too long (86 > 79 characters)
ross/api_report.py:860:80: E501 line too long (82 > 79 characters)
ross/api_report.py:894:80: E501 line too long (96 > 79 characters)
ross/api_report.py:984:80: E501 line too long (82 > 79 characters)
ross/api_report.py:1057:1: W293 blank line contains whitespace
ross/api_report.py:1081:80: E501 line too long (91 > 79 characters)
ross/api_report.py:1082:80: E501 line too long (83 > 79 characters)
ross/shaft_element.py:10:1: F401 'ross' imported but unused
ross/shaft_element.py:12:1: F401 'ross.materials.steel' imported but unused
ross/shaft_element.py:231:80: E501 line too long (87 > 79 characters)
ross/shaft_element.py:246:80: E501 line too long (82 > 79 characters)
ross/shaft_element.py:251:80: E501 line too long (81 > 79 characters)
ross/shaft_element.py:258:80: E501 line too long (92 > 79 characters)
ross/shaft_element.py:259:80: E501 line too long (91 > 79 characters)
ross/shaft_element.py:276:25: E128 continuation line under-indented for visual indent
ross/shaft_element.py:276:80: E501 line too long (99 > 79 characters)
ross/shaft_element.py:277:25: E128 continuation line under-indented for visual indent
ross/shaft_element.py:277:80: E501 line too long (100 > 79 characters)
ross/shaft_element.py:284:80: E501 line too long (102 > 79 characters)
ross/shaft_element.py:289:80: E501 line too long (122 > 79 characters)
ross/shaft_element.py:293:80: E501 line too long (96 > 79 characters)
ross/shaft_element.py:395:80: E501 line too long (80 > 79 characters)
ross/shaft_element.py:414:80: E501 line too long (99 > 79 characters)
ross/shaft_element.py:416:80: E501 line too long (100 > 79 characters)
ross/shaft_element.py:417:80: E501 line too long (103 > 79 characters)
ross/shaft_element.py:419:80: E501 line too long (101 > 79 characters)
ross/shaft_element.py:440:80: E501 line too long (88 > 79 characters)
ross/shaft_element.py:533:80: E501 line too long (80 > 79 characters)
ross/shaft_element.py:566:1: W293 blank line contains whitespace
ross/shaft_element.py:669:18: E201 whitespace after '['
ross/shaft_element.py:669:80: E501 line too long (85 > 79 characters)
ross/shaft_element.py:670:18: E201 whitespace after '['
ross/shaft_element.py:670:80: E501 line too long (85 > 79 characters)
ross/shaft_element.py:671:18: E201 whitespace after '['
ross/shaft_element.py:671:80: E501 line too long (85 > 79 characters)
ross/shaft_element.py:672:18: E201 whitespace after '['
ross/shaft_element.py:672:80: E501 line too long (85 > 79 characters)
ross/shaft_element.py:673:18: E201 whitespace after '['
ross/shaft_element.py:673:80: E501 line too long (85 > 79 characters)
ross/shaft_element.py:674:18: E201 whitespace after '['
ross/shaft_element.py:674:80: E501 line too long (85 > 79 characters)
ross/shaft_element.py:675:18: E201 whitespace after '['
ross/shaft_element.py:675:80: E501 line too long (85 > 79 characters)
ross/shaft_element.py:676:80: E501 line too long (85 > 79 characters)
ross/shaft_element.py:721:22: E201 whitespace after '['
ross/shaft_element.py:721:80: E501 line too long (95 > 79 characters)
ross/shaft_element.py:722:22: E201 whitespace after '['
ross/shaft_element.py:722:80: E501 line too long (95 > 79 characters)
ross/shaft_element.py:723:22: E201 whitespace after '['
ross/shaft_element.py:723:80: E501 line too long (95 > 79 characters)
ross/shaft_element.py:724:80: E501 line too long (95 > 79 characters)
ross/shaft_element.py:725:22: E201 whitespace after '['
ross/shaft_element.py:725:80: E501 line too long (95 > 79 characters)
ross/shaft_element.py:726:22: E201 whitespace after '['
ross/shaft_element.py:726:80: E501 line too long (95 > 79 characters)
ross/shaft_element.py:727:22: E201 whitespace after '['
ross/shaft_element.py:727:80: E501 line too long (95 > 79 characters)
ross/shaft_element.py:728:80: E501 line too long (95 > 79 characters)
ross/shaft_element.py:811:14: E201 whitespace after '['
ross/shaft_element.py:812:14: E201 whitespace after '['
ross/shaft_element.py:813:14: E201 whitespace after '['
ross/shaft_element.py:815:14: E201 whitespace after '['
ross/shaft_element.py:816:14: E201 whitespace after '['
ross/shaft_element.py:817:14: E201 whitespace after '['
ross/shaft_element.py:822:14: E201 whitespace after '['
ross/shaft_element.py:823:14: E201 whitespace after '['
ross/shaft_element.py:824:14: E201 whitespace after '['
ross/shaft_element.py:826:14: E201 whitespace after '['
ross/shaft_element.py:827:14: E201 whitespace after '['
ross/shaft_element.py:828:14: E201 whitespace after '['
ross/shaft_element.py:832:80: E501 line too long (92 > 79 characters)
ross/shaft_element.py:869:1: W293 blank line contains whitespace
ross/shaft_element.py:891:13: F841 local variable 'a1' is assigned to but never used
ross/shaft_element.py:893:13: F841 local variable 'b1' is assigned to but never used
ross/shaft_element.py:938:22: E201 whitespace after '['
ross/shaft_element.py:938:80: E501 line too long (89 > 79 characters)
ross/shaft_element.py:939:22: E201 whitespace after '['
ross/shaft_element.py:939:22: E221 multiple spaces before operator
ross/shaft_element.py:939:80: E501 line too long (89 > 79 characters)
ross/shaft_element.py:940:22: E201 whitespace after '['
ross/shaft_element.py:940:80: E501 line too long (89 > 79 characters)
ross/shaft_element.py:941:22: E201 whitespace after '['
ross/shaft_element.py:941:80: E501 line too long (89 > 79 characters)
ross/shaft_element.py:942:22: E201 whitespace after '['
ross/shaft_element.py:942:80: E501 line too long (89 > 79 characters)
ross/shaft_element.py:943:22: E201 whitespace after '['
ross/shaft_element.py:943:80: E501 line too long (89 > 79 characters)
ross/shaft_element.py:944:22: E201 whitespace after '['
ross/shaft_element.py:944:80: E501 line too long (89 > 79 characters)
ross/shaft_element.py:945:22: E201 whitespace after '['
ross/shaft_element.py:945:80: E501 line too long (89 > 79 characters)
ross/shaft_element.py:1165:80: E501 line too long (80 > 79 characters)
ross/materials.py:53:80: E501 line too long (82 > 79 characters)
ross/materials.py:65:80: E501 line too long (80 > 79 characters)
ross/materials.py:104:80: E501 line too long (86 > 79 characters)
ross/materials.py:105:80: E501 line too long (88 > 79 characters)
ross/materials.py:130:80: E501 line too long (96 > 79 characters)
ross/materials.py:195:80: E501 line too long (88 > 79 characters)
ross/utils.py:26:80: E501 line too long (97 > 79 characters)
ross/utils.py:29:80: E501 line too long (93 > 79 characters)
ross/utils.py:31:80: E501 line too long (105 > 79 characters)
ross/utils.py:39:80: E501 line too long (92 > 79 characters)
ross/utils.py:40:80: E501 line too long (104 > 79 characters)
ross/utils.py:87:80: E501 line too long (88 > 79 characters)
ross/utils.py:131:80: E501 line too long (81 > 79 characters)
ross/utils.py:138:80: E501 line too long (85 > 79 characters)
ross/utils.py:159:80: E501 line too long (96 > 79 characters)
ross/utils.py:160:80: E501 line too long (100 > 79 characters)
ross/utils.py:220:80: E501 line too long (86 > 79 characters)
ross/utils.py:237:80: E501 line too long (88 > 79 characters)
ross/utils.py:307:1: W293 blank line contains whitespace
ross/utils.py:312:1: W293 blank line contains whitespace
ross/utils.py:400:80: E501 line too long (83 > 79 characters)
ross/utils.py:401:80: E501 line too long (83 > 79 characters)
ross/utils.py:438:1: W391 blank line at end of file
ross/bearing_seal_element.py:4:1: F401 'matplotlib.patches as mpatches' imported but unused
ross/bearing_seal_element.py:50:13: E722 do not use bare 'except'
ross/bearing_seal_element.py:59:17: E722 do not use bare 'except'
ross/bearing_seal_element.py:68:80: E501 line too long (84 > 79 characters)
ross/bearing_seal_element.py:168:80: E501 line too long (98 > 79 characters)
ross/bearing_seal_element.py:381:80: E501 line too long (88 > 79 characters)
ross/bearing_seal_element.py:382:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:386:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:390:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:394:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:398:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:402:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:406:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:410:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:463:80: E501 line too long (80 > 79 characters)
ross/bearing_seal_element.py:498:80: E501 line too long (80 > 79 characters)
ross/bearing_seal_element.py:601:80: E501 line too long (84 > 79 characters)
ross/bearing_seal_element.py:609:80: E501 line too long (82 > 79 characters)
ross/bearing_seal_element.py:628:80: E501 line too long (84 > 79 characters)
ross/bearing_seal_element.py:629:80: E501 line too long (85 > 79 characters)
ross/bearing_seal_element.py:637:80: E501 line too long (90 > 79 characters)
ross/bearing_seal_element.py:657:80: E501 line too long (80 > 79 characters)
ross/bearing_seal_element.py:703:80: E501 line too long (84 > 79 characters)
ross/bearing_seal_element.py:711:80: E501 line too long (82 > 79 characters)
ross/bearing_seal_element.py:730:80: E501 line too long (84 > 79 characters)
ross/bearing_seal_element.py:731:80: E501 line too long (85 > 79 characters)
ross/bearing_seal_element.py:739:80: E501 line too long (90 > 79 characters)
ross/bearing_seal_element.py:749:80: E501 line too long (90 > 79 characters)
ross/bearing_seal_element.py:767:80: E501 line too long (103 > 79 characters)
ross/bearing_seal_element.py:790:80: E501 line too long (118 > 79 characters)
ross/bearing_seal_element.py:791:80: E501 line too long (119 > 79 characters)
ross/bearing_seal_element.py:801:80: E501 line too long (101 > 79 characters)
ross/bearing_seal_element.py:812:80: E501 line too long (103 > 79 characters)
ross/bearing_seal_element.py:863:80: E501 line too long (81 > 79 characters)
ross/bearing_seal_element.py:889:80: E501 line too long (88 > 79 characters)
ross/bearing_seal_element.py:919:80: E501 line too long (87 > 79 characters)
ross/bearing_seal_element.py:921:80: E501 line too long (101 > 79 characters)
ross/bearing_seal_element.py:1015:80: E501 line too long (91 > 79 characters)
ross/bearing_seal_element.py:1059:71: W291 trailing whitespace
ross/bearing_seal_element.py:1098:80: E501 line too long (85 > 79 characters)
ross/bearing_seal_element.py:1145:71: W291 trailing whitespace
ross/bearing_seal_element.py:1177:80: E501 line too long (85 > 79 characters)
ross/bearing_seal_element.py:1252:80: E501 line too long (82 > 79 characters)
ross/__init__.py:3:1: F403 'from .bearing_seal_element import *' used; unable to detect undefined names
ross/__init__.py:3:1: F401 '.bearing_seal_element.*' imported but unused
ross/__init__.py:4:1: F403 'from .disk_element import *' used; unable to detect undefined names
ross/__init__.py:4:1: F401 '.disk_element.*' imported but unused
ross/__init__.py:5:1: F403 'from .materials import *' used; unable to detect undefined names
ross/__init__.py:5:1: F401 '.materials.*' imported but unused
ross/__init__.py:6:1: F403 'from .rotor_assembly import *' used; unable to detect undefined names
ross/__init__.py:6:1: F401 '.rotor_assembly.*' imported but unused
ross/__init__.py:7:1: F403 'from .shaft_element import *' used; unable to detect undefined names
ross/__init__.py:7:1: F401 '.shaft_element.*' imported but unused
ross/__init__.py:8:1: F401 '.utils.visualize_matrix' imported but unused
ross/__init__.py:9:1: F403 'from .point_mass import *' used; unable to detect undefined names
ross/__init__.py:9:1: F401 '.point_mass.*' imported but unused
ross/point_mass.py:8:1: F401 'toml' imported but unused
ross/disk_element.py:57:5: E303 too many blank lines (2)
ross/disk_element.py:338:80: E501 line too long (88 > 79 characters)
ross/disk_element.py:341:80: E501 line too long (89 > 79 characters)
ross/disk_element.py:519:80: E501 line too long (96 > 79 characters)
ross/disk_element.py:520:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:21:1: F401 'bokeh.plotting.show' imported but unused
ross/rotor_assembly.py:25:1: F401 'ross' imported but unused
ross/rotor_assembly.py:63:80: E501 line too long (83 > 79 characters)
ross/rotor_assembly.py:176:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:233:9: F841 local variable 'columns' is assigned to but never used
ross/rotor_assembly.py:272:80: E501 line too long (87 > 79 characters)
ross/rotor_assembly.py:298:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:354:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:385:80: E501 line too long (87 > 79 characters)
ross/rotor_assembly.py:438:80: E501 line too long (87 > 79 characters)
ross/rotor_assembly.py:439:80: E501 line too long (87 > 79 characters)
ross/rotor_assembly.py:442:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:446:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:447:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:450:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:451:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:486:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:491:80: E501 line too long (84 > 79 characters)
ross/rotor_assembly.py:505:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:506:80: E501 line too long (84 > 79 characters)
ross/rotor_assembly.py:533:80: E501 line too long (83 > 79 characters)
ross/rotor_assembly.py:577:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:626:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:691:80: E501 line too long (85 > 79 characters)
ross/rotor_assembly.py:697:80: E501 line too long (85 > 79 characters)
ross/rotor_assembly.py:853:9: E741 ambiguous variable name 'I'
ross/rotor_assembly.py:858:80: E501 line too long (127 > 79 characters)
ross/rotor_assembly.py:895:47: E203 whitespace before ':'
ross/rotor_assembly.py:994:9: E741 ambiguous variable name 'I'
ross/rotor_assembly.py:1013:80: E501 line too long (121 > 79 characters)
ross/rotor_assembly.py:1064:80: E501 line too long (83 > 79 characters)
ross/rotor_assembly.py:1111:80: E501 line too long (88 > 79 characters)
ross/rotor_assembly.py:1150:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:1163:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:1205:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:1253:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:1263:80: E501 line too long (105 > 79 characters)
ross/rotor_assembly.py:1273:80: E501 line too long (82 > 79 characters)
ross/rotor_assembly.py:1389:80: E501 line too long (84 > 79 characters)
ross/rotor_assembly.py:1394:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:1396:80: E501 line too long (82 > 79 characters)
ross/rotor_assembly.py:1476:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:1501:80: E501 line too long (84 > 79 characters)
ross/rotor_assembly.py:1508:80: E501 line too long (80 > 79 characters)
ross/rotor_assembly.py:1510:80: E501 line too long (82 > 79 characters)
ross/rotor_assembly.py:1561:80: E501 line too long (88 > 79 characters)
ross/rotor_assembly.py:1565:80: E501 line too long (88 > 79 characters)
ross/rotor_assembly.py:1609:80: E501 line too long (87 > 79 characters)
ross/rotor_assembly.py:1671:80: E501 line too long (81 > 79 characters)
ross/rotor_assembly.py:1719:80: E501 line too long (93 > 79 characters)
ross/rotor_assembly.py:1720:80: E501 line too long (93 > 79 characters)
ross/rotor_assembly.py:1746:80: E501 line too long (85 > 79 characters)
ross/rotor_assembly.py:1828:80: E501 line too long (87 > 79 characters)
ross/rotor_assembly.py:1877:80: E501 line too long (86 > 79 characters)
ross/rotor_assembly.py:1878:80: E501 line too long (84 > 79 characters)
[…]
