### 2020-03-06 ross-rotordynamics/ross#464 @nnadeau
`black` isn't listed as a dep
If `black` is the formatter, it must be listed as a `dev` dep

cc https://github.com/openjournals/joss-reviews/issues/2120
