### 2020-03-06 ross-rotordynamics/ross#465 @nnadeau
Test/check linting/formatting in CI
`black` and other formatting and linting tools should be used in appveyor and travis to check that PRs conform to the guidelines

cc https://github.com/openjournals/joss-reviews/issues/2120
