### 2020-03-06 ross-rotordynamics/ross#466 @nnadeau
Use `isort` to enforce import order
Style guide enforcement is important in OSS projects. `isort` is a de facto standard tool. The following errors should be fixed and `isort` should be added to linting and CI.

```
ERROR: /home/nicholas/git/ross/setup.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/results.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/api_report.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/shaft_element.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/materials.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/utils.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/bearing_seal_element.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/__init__.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/point_mass.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/disk_element.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/rotor_assembly.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/units.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/element.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_materials.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_disk_element.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_examples.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_fluid_flow.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_api_report.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_bearing_seal_element.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_shaft_element.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_units.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_point_mass.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/tests/test_rotor_assembly.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/fluid_flow/fluid_flow.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/fluid_flow/fluid_flow_coefficients.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/ross/fluid_flow/fluid_flow_graphics.py Imports are incorrectly sorted.
ERROR: /home/nicholas/git/ross/docs/conf.py Imports are incorrectly sorted.

```

cc https://github.com/openjournals/joss-reviews/issues/2120
