### 2020-03-06 ross-rotordynamics/ross#462 @nnadeau
just noticed that pytest is listed in `setup.py` under:

```python
EXTRAS = {
    "dev": [
        "pytest",
        "pytest-cov",
        "coverage",
        "codecov",
        "sphinx",
        "sphinx_bootstrap_theme",
        "nbsphinx",
        "numpydoc",
        "sphinxcontrib-bibtex",
    ]
}
```

this is why https://github.com/ross-rotordynamics/ross/issues/461 is important
