### 2020-03-06 9d1cf94
Add new_units.txt to package data
The new_units.txt file needs to be copied during installation.
