+++ setup.py
@@ -138,6 +138,7 @@ setup(
     # },
     install_requires=REQUIRED,
     extras_require=EXTRAS,
+    package_data={"": ["new_units.txt"]},
     include_package_data=True,
     license="MIT",
     classifiers=[
