### 2020-03-06 fa947a8
Update README.md to improve statement of need
This improves the explanation on the use of ross and is related to
#460.
