### 2020-03-06 8eba406
Fix binder link
We have modified the binder link so that our binder will track the
last released version (0.3).
