### 2020-03-06 ac1f0a6
Initial commit for disk_element.py
