+++ ross/disk_element.py
@@ -7,7 +7,7 @@ from ross.utils import read_table_file
 
 from ross.element import Element
 
-__all__ = ["DiskElement"]
+__all__ = ["DiskElement", "DiskElement6DoF"]
 bokeh_colors = bp.RdGy[11]
 
 
@@ -551,3 +551,325 @@ def disk_example():
     """
     disk = DiskElement(0, 32.589_727_65, 0.178_089_28, 0.329_563_62)
     return disk
+
+
+class DiskElement6DoF(Element):
+    """A disk element for 6 DoFs.
+
+     This class will create a disk element with 6 DoF from input data of inertia and mass.
+
+     Parameters
+     ----------
+     n: int
+         Node in which the disk will be inserted.
+     m : float
+         Mass of the disk element.
+     Id : float
+         Diametral moment of inertia.
+     Ip : float
+         Polar moment of inertia
+     tag : str, optional
+         A tag to name the element
+         Default is None
+
+     Examples
+     --------
+     >>> disk = DiskElement(n=0, m=32, Id=0.2, Ip=0.3)
+     >>> disk.Ip
+     0.3
+     """
+
+    def __eq__(self, other):
+        """This function allows disk elements to be compared.
+        Parameters
+        ----------
+        other: object
+            The second object to be compared with.
+
+        Returns
+        -------
+        bool
+            True if the comparison is true; False otherwise.
+        Examples
+        --------
+        >>> disk1 = disk_example()
+        >>> disk2 = disk_example()
+        >>> disk1 == disk2
+        True
+        """
+
+    def __repr__(self):
+        """This function returns a string representation of a disk element.
+        Parameters
+        ----------
+
+        Returns
+        -------
+        A string representation of a disk object.
+        Examples
+        --------
+        >>> disk = disk_example()
+        >>> disk # doctest: +ELLIPSIS
+        DiskElement(Id=0.17809, Ip=0.32956...
+        """
+
+    def __hash__(self):
+        return hash(self.tag)
+
+    def save(self, file_name):
+        """Saves a disk element in a toml format.
+
+        It works as an auxiliary function of the save function in the Rotor
+        class.
+
+        Parameters
+        ----------
+        file_name: string
+            The name of the file the disk element will be saved in.
+
+        Returns
+        -------
+        None
+
+        Examples
+        --------
+        >>> disk = disk_example()
+        >>> disk.save('DiskElement6DoF.toml')
+        """
+
+    @staticmethod
+    def load(file_name="DiskElement6DoF"):
+        """Loads a list of 6DoFs disk elements saved in a toml format.
+
+        Parameters
+        ----------
+        file_name: str
+            The name of the file of the disk element to be loaded.
+
+        Returns
+        -------
+        disk_elements: list
+            A list of disk elements.
+
+        Examples
+        --------
+        >>> disk1 = disk_example()
+        >>> disk1.save('DiskElement6DoF.toml')
+        >>> list_of_disks = DiskElement6DoF.load('DiskElement6DoF.toml')
+        >>> disk1 == list_of_disks[0]
+        True
+        """
+
+    def dof_mapping(self):
+        """6DoFs degrees of freedom mapping.
+
+        Returns a dictionary with a mapping between degree of freedom and its
+        index.
+
+        Returns
+        -------
+        dof_mapping: dict
+            A dictionary containing the degrees of freedom and their indexes.
+
+        Examples
+        --------
+        The numbering of the degrees of freedom for each node. 
+        
+        Being the following their ordering for a node:
+
+        x_0,u_0 - horizontal translation, 
+        y_0,v_0 - vertical translation, 
+        z_0,w_0 - axial translation, 
+        theta_0 - rotation around horizontal, 
+        psi_0   - rotation around vertical, 
+        phi_0   - torsion around axial,
+        """
+
[…]
