### 2020-03-06 d287d0e
Repeating similar ROSS commands for the 6DoF disk element
