### 2020-03-06 ross-rotordynamics/ross#458 @raphaeltimbo
@srmnitc , we have just released the v0.3.0.
We have also updated our docs and our binder link to track the 0.3 branch.
I have made some tests and things seem to be working. Let us know if you have any problems.
Thanks!
