### 2020-03-06 ross-rotordynamics/ross#460 @raphaeltimbo
Hi there @nnadeau ! Thank you for your comments.
We have just released v0.3.0 and I have updated de README file to give a better explanation on the use of our software. Let us know if that is ok to you.
