### 2020-03-06 3e87deb
small changes in generate_pdf()
