### 2020-03-06 ross-rotordynamics/ross#461 @raphaeltimbo
The repo is intended as a library. We actually had a `requirements.txt` file but we have deleted because we were afraid of having duplicated information that could go out of sync.
But I undestand how not having the `requirements.txt` could be a problem in some cases.
Doing some research I have found the following in [this blog post](https://caremad.io/posts/2013/07/setup-vs-requirement/):

>The “Library” and “Application” distinction is all well and good, but whenever you’re developing a Library, in a way it becomes your application. You want a specific set of dependencies that you want to fetch from a specific location and you know that you should have abstract dependencies in your setup.py and concrete dependencies in your requirements.txt but you don’t want to need to maintain two separate lists which will inevitably go out of sync. As it turns out pip requirements file have a construct to handle just such a case. Given a directory with a setup.py inside of it you can write a requirements file that looks like:
```
--index-url https://pypi.python.org/simple/
-e .
```
> Now your pip install -r requirements.txt will work just as before. It will first install the library located at the file path . and then move on to its abstract dependencies, combining them with its --index-url option and turning them into concrete dependencies and installing them.

Do you think this could be a solution in our case?
