### 2020-03-06 248a2a0
Introducing disk matrices for the 6DoF model
