### 2020-03-06 41b8c9e
Repeating ROSS plot and patch functions for the 6DoF disk
