+++ ross/disk_element.py
@@ -873,6 +873,32 @@ class DiskElement6DoF(Element):
         Returns
         -------
         """
+        zpos, ypos = position
+        step = ypos / 5
+
+        #  matplotlib node (x pos), outer diam. (y pos)
+        disk_points_u = [
+            [zpos, ypos],  # upper
+            [zpos + step, ypos * 4],
+            [zpos - step, ypos * 4],
+            [zpos, ypos],
+        ]
+        disk_points_l = [
+            [zpos, -ypos],  # lower
+            [zpos + step, -ypos * 4],
+            [zpos - step, -ypos * 4],
+            [zpos, -ypos],
+        ]
+
+        ax.add_patch(mpatches.Polygon(disk_points_u, facecolor=self.color))
+        ax.add_patch(mpatches.Polygon(disk_points_l, facecolor=self.color))
+
+        ax.add_patch(
+            mpatches.Circle(xy=(zpos, ypos * 4), radius=step, color=self.color)
+        )
+        ax.add_patch(
+            mpatches.Circle(xy=(zpos, -ypos * 4), radius=step, color=self.color)
+        )
 
     def bokeh_patch(self, position, bk_ax):
         """Disk element patch.
@@ -888,6 +914,92 @@ class DiskElement6DoF(Element):
         bk_ax : bokeh plotting axes
             Returns the axes object with the plot.
         """
+        zpos, ypos = position
+        step = ypos / 5
+
+        # bokeh plot - coordinates to plot disks elements
+        z_upper = [zpos, zpos + step, zpos - step]
+        y_upper = [ypos, ypos * 4, ypos * 4]
+
+        z_lower = [zpos, zpos + step, zpos - step]
+        y_lower = [-ypos, -ypos * 4, -ypos * 4]
+
+        source = ColumnDataSource(
+            dict(
+                z_l=[z_lower],
+                y_l=[y_lower],
+                z_u=[z_upper],
+                y_u=[y_upper],
+                elnum=[self.n],
+                IP=[self.Ip],
+                ID=[self.Id],
+                mass=[self.m],
+                tag=[self.tag],
+            )
+        )
+        source_c = ColumnDataSource(
+            dict(
+                z_circle=[z_upper[0]],
+                yu_circle=[y_upper[1]],
+                yl_circle=[-y_upper[1]],
+                radius=[step],
+                elnum=[self.n],
+                IP=[self.Ip],
+                ID=[self.Id],
+                mass=[self.m],
+                tag=[self.tag],
+            )
+        )
+
+        bk_ax.patches(
+            xs="z_u",
+            ys="y_u",
+            source=source,
+            alpha=1,
+            line_width=2,
+            color=self.color,
+            legend_label="Disk",
+            name="ub_disk",
+        )
+        bk_ax.patches(
+            xs="z_l",
+            ys="y_l",
+            source=source,
+            alpha=1,
+            line_width=2,
+            color=self.color,
+            name="ub_disk",
+        )
+        bk_ax.circle(
+            x="z_circle",
+            y="yu_circle",
+            radius="radius",
+            source=source_c,
+            fill_alpha=1,
+            color=self.color,
+            name="uc_disk",
+        )
+        bk_ax.circle(
+            x="z_circle",
+            y="yl_circle",
+            radius="radius",
+            source=source_c,
+            fill_alpha=1,
+            color=self.color,
+            name="lc_disk",
+        )
+
+        hover = HoverTool(names=["uc_disk", "lc_disk", "ub_disk", "lb_disk"])
+        hover.tooltips = [
[…]
