+++ .travis.yml
@@ -9,6 +9,8 @@ before_install:
 - sudo apt-get install pandoc
 install:
 - pip install -e .[dev]
+before_script:
+- black --check ross
 script:
 - pytest ross --cov=ross
 after_success:
