+++ ross/fluid_flow/fluid_flow_coefficients.py
@@ -31,48 +31,112 @@ def calculate_oil_film_force(fluid_flow_object, force_type=None):
     >>> calculate_oil_film_force(my_fluid_flow) # doctest: +ELLIPSIS
     (...
     """
-    if force_type != 'numerical' and (force_type == 'short' or fluid_flow_object.bearing_type == 'short_bearing'):
-        radial_force = 0.5 * \
-           fluid_flow_object.viscosity * \
-           (fluid_flow_object.radius_rotor / fluid_flow_object.difference_between_radius) ** 2 * \
-           (fluid_flow_object.length ** 3 / fluid_flow_object.radius_rotor) * \
-           ((2 * fluid_flow_object.eccentricity_ratio ** 2 * fluid_flow_object.omega) /
-            (1 - fluid_flow_object.eccentricity_ratio ** 2) ** 2)
+    if force_type != "numerical" and (
+        force_type == "short" or fluid_flow_object.bearing_type == "short_bearing"
+    ):
+        radial_force = (
+            0.5
+            * fluid_flow_object.viscosity
+            * (
+                fluid_flow_object.radius_rotor
+                / fluid_flow_object.difference_between_radius
+            )
+            ** 2
+            * (fluid_flow_object.length ** 3 / fluid_flow_object.radius_rotor)
+            * (
+                (
+                    2
+                    * fluid_flow_object.eccentricity_ratio ** 2
+                    * fluid_flow_object.omega
+                )
+                / (1 - fluid_flow_object.eccentricity_ratio ** 2) ** 2
+            )
+        )
 
-        tangential_force = 0.5 * fluid_flow_object.viscosity * \
-            (fluid_flow_object.radius_rotor / fluid_flow_object.difference_between_radius) ** 2 * \
-            (fluid_flow_object.length ** 3 / fluid_flow_object.radius_rotor) * (
-             (np.pi * fluid_flow_object.eccentricity_ratio * fluid_flow_object.omega) /
-             (2 * (1 - fluid_flow_object.eccentricity_ratio ** 2) ** (3. / 2)))
-    elif force_type != 'numerical' and (force_type == 'long' or fluid_flow_object.bearing_type == 'long_bearing'):
-        radial_force = 6 * fluid_flow_object.viscosity * \
-            (fluid_flow_object.radius_rotor / fluid_flow_object.difference_between_radius) ** 2 * \
-            fluid_flow_object.radius_rotor * fluid_flow_object.length * \
-            ((2 * fluid_flow_object.eccentricity_ratio ** 2 * fluid_flow_object.omega) /
-             ((2 + fluid_flow_object.eccentricity_ratio ** 2) *
-              (1 - fluid_flow_object.eccentricity_ratio ** 2)))
-        tangential_force = 6 * fluid_flow_object.viscosity * (fluid_flow_object.radius_rotor / fluid_flow_object.difference_between_radius) ** 2 * \
-            fluid_flow_object.radius_rotor * fluid_flow_object.length * \
-            ((np.pi * fluid_flow_object.eccentricity_ratio * fluid_flow_object.omega) /
-             ((2 + fluid_flow_object.eccentricity_ratio ** 2) *
-             (1 - fluid_flow_object.eccentricity_ratio ** 2) ** 0.5))
+        tangential_force = (
+            0.5
+            * fluid_flow_object.viscosity
+            * (
+                fluid_flow_object.radius_rotor
+                / fluid_flow_object.difference_between_radius
+            )
+            ** 2
+            * (fluid_flow_object.length ** 3 / fluid_flow_object.radius_rotor)
+            * (
[…]
