+++ ross/tests/test_point_mass.py
@@ -11,7 +11,7 @@ def test_point_mass():
 
     p = PointMass(n=0, m=10.0, tag="pointmass")
 
-    assert p.tag == 'pointmass'
+    assert p.tag == "pointmass"
     assert_allclose(m0, p.M())
     assert_allclose(np.zeros((2, 2)), p.K())
     assert_allclose(np.zeros((2, 2)), p.C())

[…]
