+++ ross/api_report.py
@@ -1,28 +1,21 @@
+from copy import copy, deepcopy
+from pathlib import Path
+
+import bokeh.palettes as bp
 import numpy as np
 import pandas as pd
-
-from copy import copy
+from bokeh.layouts import gridplot, widgetbox
+from bokeh.models import (ColumnDataSource, DataRange1d, HoverTool, Label,
+                          LinearAxis, Range1d, Span)
+from bokeh.models.widgets import (DataTable, NumberFormatter, Panel,
+                                  TableColumn, Tabs)
+from bokeh.plotting import figure
 from scipy.interpolate import interp1d
 from scipy.signal import argrelextrema
-from pathlib import Path
-from ross.rotor_assembly import Rotor, rotor_example
-from ross.bearing_seal_element import BearingElement
-import ross as rs
 
-from copy import deepcopy
-import bokeh.palettes as bp
-from bokeh.plotting import figure
-from bokeh.layouts import gridplot, widgetbox
-from bokeh.models.widgets import DataTable, NumberFormatter, TableColumn, Panel, Tabs
-from bokeh.models import (
-    ColumnDataSource,
-    HoverTool,
-    Span,
-    Label,
-    LinearAxis,
-    Range1d,
-    DataRange1d,
-)
+import ross as rs
+from ross.bearing_seal_element import BearingElement
+from ross.rotor_assembly import Rotor, rotor_example
 
 # set bokeh palette of colors
 bokeh_colors = bp.RdGy[11]
