+++ ross/shaft_element.py
@@ -10,8 +10,8 @@ from bokeh.models import ColumnDataSource, HoverTool
 import ross
 from ross.element import Element
 from ross.materials import Material, steel
-from ross.utils import read_table_file
 from ross.units import check_units
+from ross.utils import read_table_file
 
 __all__ = ["ShaftElement"]
 bokeh_colors = bp.RdGy[11]
