+++ ross/tests/test_disk_element.py
@@ -1,7 +1,8 @@
 import os
+
 import numpy as np
 import pytest
-from numpy.testing import assert_almost_equal, assert_allclose
+from numpy.testing import assert_allclose, assert_almost_equal
 
 from ross.disk_element import DiskElement
 from ross.materials import steel
