+++ ross/tests/test_bearing_seal_element.py
@@ -1,4 +1,3 @@
-import math
 import os
 
 import numpy as np
