+++ ross/tests/test_api_report.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-from numpy.testing import assert_allclose, assert_almost_equal
+from numpy.testing import assert_allclose
 
 from ross.api_report import Report
 from ross.bearing_seal_element import BearingElement
