+++ .pre-commit-config.yaml
@@ -9,3 +9,4 @@ repos:
     rev: 4.3.21-2
     hooks:
       - id: isort
+      
