### 2020-03-07 f4abc85
Update fluid_flow_graphics.py
