+++ ross/api_report.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import bokeh.palettes as bp
 import numpy as np
 import pandas as pd
-from bokeh.layouts import widgetbox
+from bokeh.layouts import gridplot, widgetbox
 from bokeh.models import (ColumnDataSource, DataRange1d, HoverTool, Label,
                           LinearAxis, Range1d, Span)
 from bokeh.models.widgets import (DataTable, NumberFormatter, Panel,
