### 2020-03-10 ross-rotordynamics/ross#466 @rodrigomoliveira1
Hello @nnadeau. Thanks for your reviews!

We have added black and isort modules as dep on our CI's. Now, every time someone do a contribution the CI's will check for inconsistencies based on black and isort.
