### 2020-03-10 #2120 @meg-simula
@srmnitc @nnadeau @raphaeltimbo :wave: Just checking in - how are matters progressing?
