### 2020-03-10 ross-rotordynamics/ross#471 @srmnitc
pip install does not work
This is part of my review of the software paper at openjournals/joss-reviews#2120

I tried the installation as mentioned in the documentation in a clean environment with python 3.8.
Upon import the module as,
```
import ross as rs
```
the error I get is,
```
  File "~/anaconda3/envs/testross/lib/python3.8/site-packages/pint/registry.py", line 532, in load_definitions
    raise ValueError("While opening {}\n{}".format(file, msg))
ValueError: While opening ~/anaconda3/envs/testross/lib/python3.8/site-packages/ross/new_units.txt
[Errno 2] No such file or directory: '~/anaconda3/envs/testross/lib/python3.8/site-packages/ross/new_units.txt'
```
