### 2020-03-10 ross-rotordynamics/ross#471 @srmnitc
I can confirm that installing from the repo using `pip install git+https://github.com/ross-rotordynamics/ross.git` works fine. Maybe the pypi version is not updated?
