### 2020-03-10 ross-rotordynamics/ross#472 @srmnitc
Warnings with tests
This is part of my review of the software paper at openjournals/joss-reviews#2120

The tests run as described,  the tests raise a fair number of `DeprecationWarning` (18 ) and skips 6 tests (from `test_rotor_assembly.py` and `test_units.py`).  The first set in from `results.py` - a lot of warnings for invalid escape sequence and the rest are from Bokeh. I will it would be great and potentially would remove a lot of confusion is these are prevented.
