### 2020-03-10 ross-rotordynamics/ross#475 @srmnitc
Citations in the paper
This is part of my review of the software paper at openjournals/joss-reviews#2120

About the paper-
- Timoshenko beam theory needs a citation
- Campbell diagram probably needs a citation
