### 2020-03-10 #2120 @srmnitc
@meg-simula @raphaeltimbo I have completed my first round of review. I think the submission already meets most of the review checkpoints and is quite nicely documented with a lot of examples. I only have three main issues - with installation, tests and citations in the paper. I have raised these issues on the repository. The jupyter notebooks are quite helpful, I will also go a bit more into detail in the next one or two days.
