### 2020-03-11 ross-rotordynamics/ross#463 @rodrigomoliveira1
@nnadeau

We're currently using [Black](https://black.readthedocs.io/en/stable/) as code style guide. PR #468 has formated all the files and cleaned some of these warnings presented. Others would persist because Black is not as strict as flake8 (Black defaults to 88 characters per line, instead of 79).
Besides, some of these warnings are on purpose (manly on test files), for better visualization and understanding of the users / devs.

Would it be a okay to keep following Black as code style instead of flake8?
