### 2020-03-12 43bed46
some changes in generate_pdf() (and running black on file)
