+++ ross/api_report.py
@@ -3,29 +3,26 @@ import bokeh
 import os
 import numpy as np
 import pandas as pd
-import toml
+
 from copy import copy
-from pathlib import Path
+
+from jinja2 import Environment, FileSystemLoader
 from scipy.interpolate import interp1d
 from scipy.signal import argrelextrema
-from jinja2 import Environment, FileSystemLoader
 from pathlib import Path
 from ross.rotor_assembly import Rotor, rotor_example
 from ross.bearing_seal_element import BearingElement
 import ross as rs
 
+from copy import deepcopy
 import bokeh.palettes as bp
 from bokeh.plotting import figure
 from bokeh.layouts import gridplot, widgetbox
-from bokeh.models.widgets import DataTable, NumberFormatter, TableColumn, Panel, Tabs
+from bokeh.models.widgets import (
+    DataTable, NumberFormatter, TableColumn, Panel, Tabs
+)
 from bokeh.models import (
-    ColumnDataSource,
-    HoverTool,
-    Span,
-    Label,
-    LinearAxis,
-    Range1d,
-    DataRange1d,
+    ColumnDataSource, HoverTool, Span, Label, LinearAxis, Range1d, DataRange1d,
 )
 
 # set bokeh palette of colors
@@ -267,7 +264,9 @@ class Report:
         html = weasyprint.HTML(string=rendered_string)
         report = os.path.join(output_dir, output_filename)
         html.write_pdf(report, stylesheets=[str(css / "report.css")])
-        print(f"Report of {self.machine_type + ' '  + rotor.tag} exported in {output_dir}")
+        print(
+            f"Report of {self.machine_type + ' '  + rotor.tag} exported in {output_dir}"
+        )
         return html
 
     def static_forces(self):
@@ -439,7 +438,11 @@ class Report:
 
         # returns de nodes where forces will be applied
         node_min, node_max = self.mode_shape(mode)
-        nodes = [int(node) for sub_nodes in [node_min, node_max] for node in sub_nodes]
+        nodes = [
+            int(node)
+            for sub_nodes in [node_min, node_max]
+            for node in sub_nodes
+        ]
 
         force = self.unbalance_forces(mode)
 
@@ -447,7 +450,9 @@ class Report:
         for node in nodes:
             phase.append(np.pi)
 
-        response = self.rotor.unbalance_response(nodes, force, phase, freq_range)
+        response = self.rotor.unbalance_response(
+            nodes, force, phase, freq_range
+        )
         mag = response.magnitude
 
         for node in nodes:
@@ -775,7 +780,9 @@ class Report:
 
         return node_min, node_max
 
-    def stability_level_1(self, D, H, HP, oper_speed, RHO_ratio, RHOs, RHOd, unit="m"):
+    def stability_level_1(
+        self, D, H, HP, oper_speed, RHO_ratio, RHOs, RHOd, unit="m"
+    ):
         """Stability analysis level 1.
 
         This analysis consider a anticipated cross coupling QA based on
@@ -858,7 +865,13 @@ class Report:
             Dc, Hc = D, H
             for i, disk in enumerate(self.rotor.disk_elements):
                 if disk.n in self.disk_nodes:
-                    qi = HP[i] * Bc * C * RHO_ratio[i] / (Dc[i] * Hc[i] * oper_speed)
+                    qi = (
+                        HP[i]
+                        * Bc
+                        * C
+                        * RHO_ratio[i]
+                        / (Dc[i] * Hc[i] * oper_speed)
+                    )
                     Qi = np.linspace(0, 10 * qi, steps)
[…]
