### 2020-03-13 496421b
np.array call correction and exponent call changes
