+++ ross/shaft_element.py
@@ -26,7 +26,7 @@ class ShaftElement(Element):
 
     .. math::
 
-        [x_0, y_0, \alpha_0, \beta_0, x_1, y_1, \alpha_1, \beta_1]^T
+        [x_0, y_0, \alpha_0, \beta_0, x_1, y_1, \alpha_1, \beta_1]**T
     Where :math:`\alpha_0` and :math:`\alpha_1` are the bending on the yz plane
     and :math:`\beta_0` and :math:`\beta_1` are the bending on the xz plane.
 
@@ -247,7 +247,7 @@ class ShaftElement(Element):
         self.beam_cg = L * c1 / (4 * c2)
         self.axial_cg_pos = None
 
-        # Slenderness ratio of beam elements (G*A*L^2) / (E*I)
+        # Slenderness ratio of beam elements (G*A*L**2) / (E*I)
         sld = (self.material.G_s * self.A * self.L ** 2) / (self.material.E * Ie)
         self.slenderness_ratio = sld
 
@@ -1203,7 +1203,7 @@ class ShaftElement6DoF(Element):
     The matrices will be defined considering the following local
     coordinate vector:
 
-    :math:`[u_0, v_0, w_0, \theta_0, \psi_0, \phi_0, u_1, v_1, w_1, \theta_1, \psi_1, \phi_1]^T`
+    :math:`[u_0, v_0, w_0, \theta_0, \psi_0, \phi_0, u_1, v_1, w_1, \theta_1, \psi_1, \phi_1]**T`
     
     Being the following their ordering for an element:
 
@@ -1756,39 +1756,39 @@ class ShaftElement6DoF(Element):
         # fmt: off
 
         # Standard mass matrix
-        M = aux1 * np.array[
+        M = aux1 * np.array([
             [   156,     0, 0,      0,  -22*L, 0,    54,     0, 0,      0,   13*L, 0]
             [     0,   156, 0,   22*L,      0, 0,     0,    54, 0,  -13*L,      0, 0]
             [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
-            [     0,  22*L, 0,  4*L^2,      0, 0,     0,  13*L, 0, -3*L^2,      0, 0]
-            [ -22*L,     0, 0,      0,  4*L^2, 0, -13*L,     0, 0,      0, -3*L^2, 0]
+            [     0,  22*L, 0,  4*L**2,      0, 0,     0,  13*L, 0, -3*L**2,      0, 0]
+            [ -22*L,     0, 0,      0,  4*L**2, 0, -13*L,     0, 0,      0, -3*L**2, 0]
             [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
             [    54,     0, 0,      0,  -13*L, 0,   156,     0, 0,      0,   22*L, 0]
             [     0,    54, 0,   13*L,      0, 0,     0,   156, 0,  -22*L,      0, 0]
             [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
-            [     0, -13*L, 0, -3*L^2,      0, 0,     0, -22*L, 0,  4*L^2,      0, 0]
-            [  13*L,     0, 0,      0, -3*L^2, 0,  22*L,     0, 0,      0,  4*L^2, 0]
+            [     0, -13*L, 0, -3*L**2,      0, 0,     0, -22*L, 0,  4*L**2,      0, 0]
+            [  13*L,     0, 0,      0, -3*L**2, 0,  22*L,     0, 0,      0,  4*L**2, 0]
[…]
