+++ ross/shaft_element.py
@@ -1760,14 +1760,14 @@ class ShaftElement6DoF(Element):
             [   156,     0, 0,      0,  -22*L, 0,    54,     0, 0,      0,   13*L, 0]
             [     0,   156, 0,   22*L,      0, 0,     0,    54, 0,  -13*L,      0, 0]
             [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
-            [     0,  22*L, 0,  4*L**2,      0, 0,     0,  13*L, 0, -3*L**2,      0, 0]
-            [ -22*L,     0, 0,      0,  4*L**2, 0, -13*L,     0, 0,      0, -3*L**2, 0]
+            [     0,  22*L, 0, 4*L**2,      0, 0,     0,  13*L, 0,-3*L**2,      0, 0]
+            [ -22*L,     0, 0,      0, 4*L**2, 0, -13*L,     0, 0,      0,-3*L**2, 0]
             [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
             [    54,     0, 0,      0,  -13*L, 0,   156,     0, 0,      0,   22*L, 0]
             [     0,    54, 0,   13*L,      0, 0,     0,   156, 0,  -22*L,      0, 0]
             [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
-            [     0, -13*L, 0, -3*L**2,      0, 0,     0, -22*L, 0,  4*L**2,      0, 0]
-            [  13*L,     0, 0,      0, -3*L**2, 0,  22*L,     0, 0,      0,  4*L**2, 0]
+            [     0, -13*L, 0,-3*L**2,      0, 0,     0, -22*L, 0, 4*L**2,      0, 0]
+            [  13*L,     0, 0,      0,-3*L**2, 0,  22*L,     0, 0,      0, 4*L**2, 0]
             [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
         ])
 
@@ -1776,14 +1776,14 @@ class ShaftElement6DoF(Element):
             [   36,   0, 0,     0,  -3*L, 0, -36,    0, 0,     0,  -3*L, 0]
             [    0,  36, 0,   3*L,     0, 0,   0,  -36, 0,   3*L,     0, 0]
             [    0,   0, 0,     0,     0, 0,   0,    0, 0,     0,     0, 0]
-            [    0, 3*L, 0, 4*L**2,     0, 0,   0, -3*L, 0,  -L**2,     0, 0]
-            [ -3*L,   0, 0,     0, 4*L**2, 0, 3*L,    0, 0,     0,  -L**2, 0]
+            [    0, 3*L, 0,4*L**2,     0, 0,   0, -3*L, 0, -L**2,     0, 0]
+            [ -3*L,   0, 0,     0,4*L**2, 0, 3*L,    0, 0,     0, -L**2, 0]
             [    0,   0, 0,     0,     0, 0,   0,    0, 0,     0,     0, 0]
             [  -36,   0, 0,     0,   3*L, 0,  36,    0, 0,     0,   3*L, 0]
             [    0, -36, 0,  -3*L,     0, 0,   0,   36, 0,  -3*L,     0, 0]
             [    0,   0, 0,     0,     0, 0,   0,    0, 0,     0,     0, 0]
-            [    0, 3*L, 0,  -L**2,     0, 0,   0, -3*L, 0, 4*L**2,     0, 0]
-            [ -3*L,   0, 0,     0,  -L**2, 0, 3*L,    0, 0,     0, 4*L**2, 0]
+            [    0, 3*L, 0, -L**2,     0, 0,   0, -3*L, 0,4*L**2,     0, 0]
[…]
