### 2020-03-19 6455d86
matrices declaration fix
commas were missing in the end of each bracket
