+++ ross/shaft_element.py
@@ -1757,66 +1757,66 @@ class ShaftElement6DoF(Element):
 
         # Standard mass matrix
         M = aux1 * np.array([
-            [   156,     0, 0,      0,  -22*L, 0,    54,     0, 0,      0,   13*L, 0]
-            [     0,   156, 0,   22*L,      0, 0,     0,    54, 0,  -13*L,      0, 0]
-            [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
-            [     0,  22*L, 0, 4*L**2,      0, 0,     0,  13*L, 0,-3*L**2,      0, 0]
-            [ -22*L,     0, 0,      0, 4*L**2, 0, -13*L,     0, 0,      0,-3*L**2, 0]
-            [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
-            [    54,     0, 0,      0,  -13*L, 0,   156,     0, 0,      0,   22*L, 0]
-            [     0,    54, 0,   13*L,      0, 0,     0,   156, 0,  -22*L,      0, 0]
-            [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
-            [     0, -13*L, 0,-3*L**2,      0, 0,     0, -22*L, 0, 4*L**2,      0, 0]
-            [  13*L,     0, 0,      0,-3*L**2, 0,  22*L,     0, 0,      0, 4*L**2, 0]
-            [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0]
+            [   156,     0, 0,      0,  -22*L, 0,    54,     0, 0,      0,   13*L, 0],
+            [     0,   156, 0,   22*L,      0, 0,     0,    54, 0,  -13*L,      0, 0],
+            [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0],
+            [     0,  22*L, 0, 4*L**2,      0, 0,     0,  13*L, 0,-3*L**2,      0, 0],
+            [ -22*L,     0, 0,      0, 4*L**2, 0, -13*L,     0, 0,      0,-3*L**2, 0],
+            [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0],
+            [    54,     0, 0,      0,  -13*L, 0,   156,     0, 0,      0,   22*L, 0],
+            [     0,    54, 0,   13*L,      0, 0,     0,   156, 0,  -22*L,      0, 0],
+            [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0],
+            [     0, -13*L, 0,-3*L**2,      0, 0,     0, -22*L, 0, 4*L**2,      0, 0],
+            [  13*L,     0, 0,      0,-3*L**2, 0,  22*L,     0, 0,      0, 4*L**2, 0],
+            [     0,     0, 0,      0,      0, 0,     0,     0, 0,      0,      0, 0],
         ])
 
         # Secondary inertias mass matrix
         Ms = self.material.rho * tempI / (30 * L) * np.array([
-            [   36,   0, 0,     0,  -3*L, 0, -36,    0, 0,     0,  -3*L, 0]
-            [    0,  36, 0,   3*L,     0, 0,   0,  -36, 0,   3*L,     0, 0]
[…]
