### 2020-03-20 7451400
Add stochastic disk element file
* This class creates an instance of random disk elements
