+++ ross/stochastic/st_disk_element.py
@@ -1,0 +1,177 @@
+import bokeh.palettes as bp
+import numpy as np
+
+from ross.disk_element import DiskElement
+
+bokeh_colors = bp.RdGy[11]
+
+__all__ = ["ST_DiskElement"]
+
+
+class ST_DiskElement:
+    """Random disk element.
+
+    Creates an object containing a list with random instances of DiskElement.
+
+    Parameters
+    ----------
+    n: int
+        Node in which the disk will be inserted.
+    m : float, list
+        Mass of the disk element.
+        Input a list to make it random.
+    Id : float, list
+        Diametral moment of inertia.
+        Input a list to make it random.
+    Ip : float, list
+        Polar moment of inertia
+        Input a list to make it random.
+    tag : str, optional
+        A tag to name the element
+        Default is None
+    color : str, optional
+        A color to be used when the element is represented.
+        Default is '#b2182b' (Cardinal).
+    is_random : list
+        List of the object attributes to become random.
+        Possibilities:
+            ["m", "Id", "Ip"]
+
+    Attributes
+    ----------
+    elements : list
+        display the list with random disk elements.
+
+    Example
+    -------
+    >>> import numpy as np
+    >>> import ross.stochastic as srs
+    >>> elms = srs.ST_DiskElement(n=1,
+    ...                           m=30.0,
+    ...                           Id=np.random.uniform(0.20, 0.40, 5),
+    ...                           Ip=np.random.uniform(0.15, 0.25, 5),
+    ...                           is_random=["Id", "Ip"],
+    ...                           )
+    >>> len(elms.elements)
+    5
+    """
+
+    def __init__(
+        self, n, m, Id, Ip, tag=None, color=bokeh_colors[9], is_random=None,
+    ):
+        attribute_dict = dict(n=n, m=m, Id=Id, Ip=Ip, tag=tag, color=color)
+        elements = self.random_var(is_random, attribute_dict)
+
+        self.attribute_dict = attribute_dict
+        self.elements = elements
+
+    def random_var(self, is_random, *args):
+        """Generate a list of objects as random attributes.
+
+        This function creates a list of objects with random values for selected
+        attributes from DiskElement.
+
+        Parameters
+        ----------
+        is_random : list
+            List of the object attributes to become stochastic.
+        *args : dict
+            Dictionary instanciating the ShaftElement class.
+            The attributes that are supposed to be stochastic should be
+            set as lists of random variables.
+
+        Returns
+        -------
+        f_list : list
+            List of random objects.
+
+        Example
+        -------
+        """
+        args_dict = args[0]
+        new_args = []
+        for i in range(len(args_dict[is_random[0]])):
+            arg = []
+            for key, value in args_dict.items():
+                if key in is_random:
+                    arg.append(value[i])
+                else:
+                    arg.append(value)
+            new_args.append(arg)
+        f_list = [DiskElement(*arg) for arg in new_args]
+
+        return f_list
+
+    @classmethod
+    def from_geometry(
+        cls, n, material, width, i_d, o_d, tag=None, is_random=None,
+    ):
+        """Random disk element.
+
+        Creates an object containing a list with random instances of
+        DiskElement.from_geometry.
+
+        Parameters
+        ----------
+        n: int
+            Node in which the disk will be inserted.
+        material: ross.Material, list of ross.Material
+            Disk material.
+            Input a list to make it random.
+        width: float, list
+            The disk width.
+            Input a list to make it random.
+        i_d: float, list
+            Inner diameter.
+            Input a list to make it random.
+        o_d: float, list
+            Outer diameter.
+            Input a list to make it random.
[…]
