### 2020-03-20 818e9ad
Add stochastic point mass element file
* This class creates an instance of random point mass elements
