### 2020-03-20 ff749bc
Add stochastic bearing element file
* This class creates an instance of random bearing elements
