+++ ross/stochastic/st_bearing_seal_element.py
@@ -1,0 +1,199 @@
+from ross.bearing_seal_element import BearingElement
+
+__all__ = ["ST_BearingElement"]
+
+
+class ST_BearingElement:
+    """Random bearing element.
+
+    Creates an object containing a list with random instances of
+    BearingElement.
+
+    Considering constant coefficients, use an 1-D array to make it random.
+    Considering varying coefficients to the frequency, use and 2-D array to
+    make it random (*see the Examples below).
+
+    Parameters
+    ----------
+    n: int
+        Node which the bearing will be located in
+    kxx: float, 1-D array, 2-D array
+        Direct stiffness in the x direction.
+    cxx: float, 1-D array, 2-D array
+        Direct damping in the x direction.
+    kyy: float, 1-D array, 2-D array, optional
+        Direct stiffness in the y direction.
+        (defaults to kxx)
+    kxy: float, 1-D array, 2-D array, optional
+        Cross coupled stiffness in the x direction.
+        (defaults to 0)
+    kyx: float, 1-D array, 2-D array, optional
+        Cross coupled stiffness in the y direction.
+        (defaults to 0)
+    cyy: float, 1-D array, 2-D array, optional
+        Direct damping in the y direction.
+        (defaults to cxx)
+    cxy: float, 1-D array, 2-D array, optional
+        Cross coupled damping in the x direction.
+        (defaults to 0)
+    cyx: float, 1-D array, 2-D array, optional
+        Cross coupled damping in the y direction.
+        (defaults to 0)
+    frequency: array, optional
+        Array with the frequencies (rad/s).
+    tag: str, optional
+        A tag to name the element
+        Default is None.
+    n_link: int, optional
+        Node to which the bearing will connect. If None the bearing is
+        connected to ground.
+        Default is None.
+    scale_factor: float, optional
+        The scale factor is used to scale the bearing drawing.
+        Default is 1.
+    is_random : list
+        List of the object attributes to become stochastic.
+        Possibilities:
+            ["kxx", "kxy", "kyx", "kyy", "cxx", "cxy", "cyx", "cyy"]
+
+    Attributes
+    ----------
+    elements_list : list
+        display the list with random bearing elements.
+
+    Example
+    -------
+    >>> import numpy as np
+    >>> import ross.stochastic as srs
+
+    # Uncertanties on constant bearing coefficients
+
+    >>> s = 10
+    >>> kxx = np.random.uniform(1e6, 2e6, s)
+    >>> cxx = np.random.uniform(1e6, 2e6, s)
+    >>> elms = srs.ST_BearingElement(n=1,
+    ...                              kxx=kxx,
+    ...                              cxx=cxx,
+    ...                              is_random = ["kxx", "cxx"],
+    ...                              )
+    >>> len(elms.elements)
+    10
+
+    # Uncertanties on bearing coefficients varying with frequency
+
+    >>> s = 5
+    >>> kxx = [np.random.uniform(1e6, 2e6, s),
+    ...        np.random.uniform(2.3e6, 3.3e6, s)]
+    >>> cxx = [np.random.uniform(1e3, 2e3, s),
+    ...        np.random.uniform(2.1e3, 3.1e3, s)]
+    >>> frequency = np.linspace(500, 800, len(kxx))
+    >>> elms = srs.ST_BearingElement(n=1,
+    ...                              kxx=kxx,
+    ...                              cxx=cxx,
+    ...                              frequency=frequency,
+    ...                              is_random = ["kxx", "cxx"],
+    ...                              )
+    >>> len(elms.elements)
+    5
+    """
+
+    def __init__(
+        self,
+        n,
+        kxx,
+        cxx,
[…]
