### 2020-03-20 5eb77d7
Create __init__.py for stochastic module
