### 2020-03-20 39c4f6b
Add stochastic rotor assembly file
* This class creates an instance of random rotors
