### 2020-03-22 ross-rotordynamics/ross#479 @nnadeau
More recent references
Many of the references cited in [`paper/paper.md`](paper/paper.md) are old. Unless the reference is a seminal work, references should be <5 years old. Are there other open source software frameworks for rotordynamics?

From https://academia.stackexchange.com/a/24050 :

- How can the reviewer judge the contribution if recent work isn't cited?
- Has what's being proposed already been done? In a better way?
- Not acknowledging recent works can be seen as lazy and potentially misleading.

cc https://github.com/openjournals/joss-reviews/issues/2120
