### 2020-03-22 ross-rotordynamics/ross#481 @nnadeau
Missing DOI
https://github.com/ross-rotordynamics/ross/blob/4de0c5bf825a241859fb989bb28ada128157a9e0/paper/paper.bib#L34-L39

cc https://github.com/openjournals/joss-reviews/issues/2120
