### 2020-03-22 ross-rotordynamics/ross#482 @nnadeau
Missing DOIs
All references should have DOIs.

Relates to:
- #480
- #481

cc https://github.com/openjournals/joss-reviews/issues/2120
