### 2020-03-22 ross-rotordynamics/ross#483 @nnadeau
In-text citations should be in chronological order
e.g.,

https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L50

cc https://github.com/openjournals/joss-reviews/issues/2120
