### 2020-03-22 ross-rotordynamics/ross#484 @nnadeau
Better explain citations
re: https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L50

- how do these citations simulate the behavior of mechanical systems?
- what makes them special?
- why cite them?

cc https://github.com/openjournals/joss-reviews/issues/2120
