### 2020-03-22 ross-rotordynamics/ross#485 @nnadeau
MATLAB is uppercase
https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L55

MATLAB is uppercase

cc https://github.com/openjournals/joss-reviews/issues/2120
