### 2020-03-22 ross-rotordynamics/ross#486 @nnadeau
`independent` software?
https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L54-L58

what is `independent` software?

cc https://github.com/openjournals/joss-reviews/issues/2120
