### 2020-03-22 ross-rotordynamics/ross#487 @nnadeau
`Timoshenko beam theory`
https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L60-L63

no citation for `Timoshenko beam theory` ?

cc https://github.com/openjournals/joss-reviews/issues/2120
