### 2020-03-22 ross-rotordynamics/ross#488 @nnadeau
Improve explanation of simulation/modeling assumptions
https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L62-L63

`Disks are assumed to be rigid bodies, thus their strain energy is not taken
into account. And bearings/seals are included as linear stiffness/damping coefficients.`

Why? Does this differ from other approaches?

cc https://github.com/openjournals/joss-reviews/issues/2120
