### 2020-03-22 ross-rotordynamics/ross#489 @nnadeau
Oxford commas
Use Oxford commas and break into smaller phrases/sentences for easier reading:

https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L79-L82

cc https://github.com/openjournals/joss-reviews/issues/2120
