### 2020-03-22 ross-rotordynamics/ross#490 @nnadeau
Better Python packages explanation
https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L84-L85

Why?

cc cc https://github.com/openjournals/joss-reviews/issues/2120
