### 2020-03-22 ross-rotordynamics/ross#491 @nnadeau
Syntax fix
https://github.com/ross-rotordynamics/ross/blob/master/paper/paper.md#L104

`on the [documentation]` -> `in the [documentation]`

cc https://github.com/openjournals/joss-reviews/issues/2120
