### 2020-03-22 ross-rotordynamics/ross#492 @nnadeau
`software is extensible` claim
`The software is extensible and new elements or new types of analysis can be easily included.`

What kind of extensibility? Any examples?

cc https://github.com/openjournals/joss-reviews/issues/2120
