### 2020-03-22 ross-rotordynamics/ross#493 @nnadeau
Other works using ROSS?
From https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

```
Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.
```

Any citations, works, references, projects, etc. that are using ROSS?

cc https://github.com/openjournals/joss-reviews/issues/2120
