### 2020-03-22 ross-rotordynamics/ross#494 @nnadeau
Explain significant contribution
From https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements

```
The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).
```

How does ROSS fulfil the above statement?

cc https://github.com/openjournals/joss-reviews/issues/2120
