### 2020-03-22 ross-rotordynamics/ross#495 @nnadeau
More clear Statement of Need
From https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

```
A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

A clear Statement of Need that illustrates the research purpose of the software.
```

In my opinion, the paper does not fully convey *WHY* this package is needed. Is this just another simulation package?

cc https://github.com/openjournals/joss-reviews/issues/2120
