### 2020-03-22 ross-rotordynamics/ross#461 @nnadeau
@raphaeltimbo

I personally recommend one of the following:

1) A super packaging framework https://python-poetry.org/

2) Reading your `requirements.txt` file into your `setup.py`:

```
with open('requirements.txt') as f:
    requirements = f.read()

setup(
	requires=requirements
)
```
