### 2020-03-22 ross-rotordynamics/ross#461 @nnadeau
e.g., https://www.reddit.com/r/Python/comments/3uzl2a/setuppy_requirementstxt_or_a_combination/cxk3z76?utm_source=share&utm_medium=web2x
