### 2020-03-22 ross-rotordynamics/ross#463 @nnadeau
yes, `black` is preferred for code formatting. `flake8` is best used as a python linter.

you can copy the `.flake8` config file found here: https://github.com/psf/black/blob/master/.flake8
