### 2020-03-23 c6cbae2
Add stochastic results file
Supporting plots:
* Campbell Diagram
* Frequency Response
* Time Response
