+++ ross/stochastic/st_results.py
@@ -1,0 +1,647 @@
+import bokeh.palettes as bp
+import numpy as np
+from bokeh.layouts import gridplot
+from bokeh.plotting import figure
+
+# set bokeh palette of colors
+colors1 = bp.Category10[10]
+colors2 = bp.Category20c[20]
+
+
+class ST_CampbellResults:
+    """Class used to store stochastic results and provide plots for Campbell Diagram.
+
+    It's possible to visualize multiples harmonics in a single plot to check
+    other speeds which also excite a specific natural frequency.
+    Two options for plooting are available: Matplotlib and Bokeh. The user
+    chooses between them using the attribute plot_type. The default is bokeh
+
+    Parameters
+    ----------
+    speed_range : array
+        Array with the speed range in rad/s.
+    wd : array
+        Array with the damped natural frequencies
+    log_dec : array
+        Array with the Logarithmic decrement
+
+    Returns
+    -------
+    ax : matplotlib axes
+        Returns the matplotlib axes object with the plot
+        if plot_type == "matplotlib"
+    bk_ax : bokeh axes
+        Returns the bokeh axes object with the plot
+        if plot_type == "bokeh"
+    """
+
+    def __init__(self, speed_range, wd, log_dec):
+        self.speed_range = speed_range
+        self.wd = wd
+        self.log_dec = log_dec
+
+    def plot_nat_freq(self, percentile=[], conf_interval=[], harmonics=[1], **kwargs):
+        """Plot the damped natural frequencies vs frequency.
+
+        Parameters
+        ----------
+        percentile : list, optional
+            Sequence of percentiles to compute, which must be between
+            0 and 100 inclusive.
+        conf_interval : list, optional
+            Sequence of confidence intervals to compute, which must be between
+            0 and 100 inclusive.
+        harmonics: list, optional
+            List withe the harmonics to be plotted.
+            The default is to plot 1x.
+        kwargs : optional
+            Additional key word arguments can be passed to change
+            the plot (e.g. linestyle='--').
+
+        Returns
+        -------
+        fig : Bokeh figure
+            The bokeh axes object with the plot.
+        """
+        default_values = dict(line_width=3.0, line_alpha=1.0)
+
+        for k, v in default_values.items():
+            kwargs.setdefault(k, v)
+
+        fig = figure(
+            height=900,
+            width=900,
+            tools="pan, box_zoom, wheel_zoom, reset, save",
+            title="Campbell Diagram",
+            y_axis_label="Damped Natural Frequencies",
+            x_axis_label="Rotor Speed",
+        )
+        fig.xaxis.axis_label_text_font_size = "14pt"
+        fig.yaxis.axis_label_text_font_size = "14pt"
+        fig.axis.major_label_text_font_size = "14pt"
+        fig.title.text_font_size = "14pt"
+
+        for j in range(self.wd.shape[0]):
+            fig.line(
+                x=self.speed_range,
+                y=np.mean(self.wd[j], axis=1),
+                line_color=colors1[j],
+                line_alpha=1.0,
+                line_width=3.0,
+                muted_color=colors1[j],
+                muted_alpha=0.1,
+                legend_label="Mean - Mode %s" % (j + 1),
+            )
+            if len(percentile):
+                for i, p in enumerate(percentile):
+                    fig.line(
+                        x=self.speed_range,
+                        y=np.percentile(self.wd[j], p, axis=1),
+                        line_color=colors2[i],
+                        line_alpha=0.6,
+                        line_width=2.5,
+                        muted_color=colors2[i],
+                        muted_alpha=0.1,
+                        legend_label="percentile %s" % (p),
+                    )
+            if len(conf_interval):
+                for i, p in enumerate(conf_interval):
+                    fig.line(
[…]
