### 2020-03-23 2ad59a0
Add methods for rotor analysis
* Add methods for: Campbell Diagram, Frequency Response and Time Response
* Add routine to generate random rotors
