+++ ross/stochastic/st_rotor_assembly.py
@@ -1,7 +1,15 @@
+# fmt: off
+import numpy as np
+
+from ross.rotor_assembly import Rotor
 from ross.stochastic.st_bearing_seal_element import ST_BearingElement
 from ross.stochastic.st_disk_element import ST_DiskElement
 from ross.stochastic.st_point_mass import ST_PointMass
+from ross.stochastic.st_results import (ST_CampbellResults,
+                                        ST_FrequencyResponseResults,
+                                        ST_TimeResponseResults)
 from ross.stochastic.st_shaft_element import ST_ShaftElement
+# fmt: on
 
 __all__ = ["ST_Rotor"]
 
@@ -35,12 +43,44 @@ class ST_Rotor(object):
 
     Attributes
     ----------
+    rotor_list : list, array
+        List with random rotor objects
 
     Returns
     -------
+    Random rotors objects
 
     Examples
     --------
+    >>> # Rotor with 2 shaft elements, 1 random disk element and 2 bearings
+    >>> import numpy as np
+    >>> import ross as rs
+    >>> import ross.stochastic as srs
+    >>> steel = rs.materials.steel
+    >>> le = 0.25
+    >>> i_d = 0
+    >>> o_d = 0.05
+    >>> tim0 = rs.ShaftElement(le, i_d, o_d, material=steel)
+    >>> tim1 = rs.ShaftElement(le, i_d, o_d, material=steel)
+    >>> shaft_elm = [tim0, tim1]
+
+    # Building random disk element
+    >>> size = 5
+    >>> i_d=np.random.uniform(0.05, 0.06, size)
+    >>> o_d=np.random.uniform(0.35, 0.39, size)
+    >>> disk0 = srs.ST_DiskElement.from_geometry(n=1,
+    ...                                          material=steel,
+    ...                                          width=0.07,
+    ...                                          i_d=i_d,
+    ...                                          o_d=o_d,
+    ...                                          is_random=["i_d", "o_d"],
+    ...                                          )
+    >>> stf = 1e6
+    >>> bearing0 = rs.BearingElement(0, kxx=stf, cxx=0)
+    >>> bearing1 = rs.BearingElement(2, kxx=stf, cxx=0)
+    >>> rand_rotor = srs.ST_Rotor(shaft_elm, [disk0], [bearing0, bearing1])
+    >>> len(rand_rotor.rotor_list)
+    5
     """
 
     def __init__(
@@ -57,24 +97,317 @@ class ST_Rotor(object):
         tag=None,
     ):
 
-        it = iter([elm for elm in shaft_elements if isinstance(elm, ST_ShaftElement)])
-        the_len = len(next(it))
-        if not all(len(l) == the_len for l in it):
-            raise ValueError("not all random shaft elements lists have same length.")
+        if disk_elements is None:
+            disk_elements = []
+        if bearing_elements is None:
+            bearing_elements = []
+        if point_mass_elements is None:
+            point_mass_elements = []
 
-        it = iter([elm for elm in disk_elements if isinstance(elm, ST_DiskElement)])
-        the_len = len(next(it))
-        if not all(len(l) == the_len for l in it):
-            raise ValueError("not all random disk elements lists have same length.")
+        # checking for random elements and matching sizes
+        is_random = []
 
-        it = iter(
-            [elm for elm in bearing_elements if isinstance(elm, ST_BearingElement)]
+        if any(isinstance(elm, ST_ShaftElement) for elm in shaft_elements):
+            is_random.append("shaft_elements")
+
+            it = iter(
+                [elm for elm in shaft_elements if isinstance(elm, ST_ShaftElement)]
+            )
+            the_len = len(next(it).elements)
+            if not all(len(l) == the_len for l in it):
+                raise ValueError(
[…]
