### 2020-03-24 751fd10
Add examples to ST_Rotor methods
* Add examples to `.run_campbell()`, `.run_freq_response()`, `.run_time_response()`.
* Fix a bug when instantiating rotors with multiples random elements.
* Add an example to run the doctests.
