+++ ross/stochastic/st_rotor_assembly.py
@@ -12,7 +12,7 @@ from ross.stochastic.st_shaft_element import ST_ShaftElement
 
 # fmt: on
 
-__all__ = ["ST_Rotor"]
+__all__ = ["ST_Rotor", "st_rotor_example"]
 
 
 class ST_Rotor(object):
@@ -115,7 +115,7 @@ class ST_Rotor(object):
                 [elm for elm in shaft_elements if isinstance(elm, ST_ShaftElement)]
             )
             the_len = len(next(it).elements)
-            if not all(len(l) == the_len for l in it):
+            if not all(len(l.elements) == the_len for l in it):
                 raise ValueError(
                     "not all random shaft elements lists have same length."
                 )
@@ -125,7 +125,7 @@ class ST_Rotor(object):
 
             it = iter([elm for elm in disk_elements if isinstance(elm, ST_DiskElement)])
             the_len = len(next(it).elements)
-            if not all(len(l) == the_len for l in it):
+            if not all(len(l.elements) == the_len for l in it):
                 raise ValueError("not all random disk elements lists have same length.")
 
         if any(isinstance(elm, ST_BearingElement) for elm in bearing_elements):
@@ -135,7 +135,7 @@ class ST_Rotor(object):
                 [elm for elm in bearing_elements if isinstance(elm, ST_BearingElement)]
             )
             the_len = len(next(it).elements)
-            if not all(len(l) == the_len for l in it):
+            if not all(len(l.elements) == the_len for l in it):
                 raise ValueError(
                     "not all random bearing elements lists have same length."
                 )
@@ -147,7 +147,7 @@ class ST_Rotor(object):
                 [elm for elm in point_mass_elements if isinstance(elm, ST_PointMass)]
             )
             the_len = len(next(it).elements)
-            if not all(len(l) == the_len for l in it):
+            if not all(len(l.elements) == the_len for l in it):
                 raise ValueError("not all random point mass lists have same length.")
 
         for i, elm in enumerate(shaft_elements):
@@ -261,6 +261,18 @@ class ST_Rotor(object):
 
         Example
         -------
+        >>> import ross.stochastic as srs
+        >>> rotors = srs.st_rotor_example()
+
+        # Running Campbell Diagram and saving the results
+
+        >>> speed_range = np.linspace(0, 500, 31)
+        >>> results = rotors.run_campbell(speed_range)
+
+        # Plotting Campbell Diagram with bokeh
+
+        >>> results.plot(conf_interval=[90]) # doctest: +ELLIPSIS
+        Column...
         """
         CAMP_size = len(speed_range)
         RV_size = len(self.rotor_list)
@@ -306,6 +318,20 @@ class ST_Rotor(object):
 
         Example
         -------
+        >>> import ross.stochastic as srs
+        >>> rotors = srs.st_rotor_example()
+
+        # Running Frequency Response and saving the results
+
+        >>> speed_range = np.linspace(0, 500, 31)
+        >>> inp = 9
+        >>> out = 9
+        >>> results = rotors.run_freq_response(speed_range, inp, out)
+
+        # Plotting Frequency Response with bokeh
+
+        >>> results.plot(conf_interval=[90]) # doctest: +ELLIPSIS
+        Column...
         """
         FRF_size = len(speed_range)
         RV_size = len(self.rotor_list)
@@ -327,6 +353,7 @@ class ST_Rotor(object):
 
         This function will take a rotor object and plot its time response
         given a force and a time.
+        The force parameter can be passed as random.
 
         Parameters
         ----------
@@ -336,7 +363,8 @@ class ST_Rotor(object):
[…]
