### 2020-03-24 f3fb66b
Add ST_ForcedResponseResults class
Class used to store stochastic results and provide plots for Unbalance Response
