### 2020-03-25 ross-rotordynamics/ross#472 @rodrigomoliveira1
Hello @srmnitc

Those `DeprecationWarning`'s about invalid escape sequence happens  while using LaTeX syntax to write greek letters in matplotlib plot. And about Bokeh, they're happening due their recent updates to version 1.4.0 and 2.0.0. I'm going to fix most of them.

About the tests:
- ``test_rotor_assembly.py`: we've skipped some tests that returns matrices with very small values (e.g. 1e-18, 1e-19). As the computational precision does not exceed 16 decimal digits, each time the tests are run, these values appear different (although this does not imply wrong final results). In fact, we have added some explaination on the command line why we're skipping these tests.

- `test_units.py`: the tests we're skipping are related to future features. So as soon we implement them, we'll turn these tests on back again.
