### 2020-03-25 75c4a01
Removing deprecations
* Replace legend by legend_label when plotting with bokeh
