### 2020-03-25 a170523
Generic 6DoF bearing element added
This code adds the 6 DoF bearing with only its coefficients informed by the user. The graphs are ported from the BallBearingElement.
