+++ ross/bearing_seal_element.py
@@ -1225,7 +1225,290 @@ class RollerBearingElement(BearingElement):
 
         self.color = "#77ACA2"
 
+        
+class BearingElement6DoF(BearingElement):
+    """A generalistic 6 DoF bearing element. This class will create a bearing 
+    element based on the user supplied stiffness and damping coefficients. These
+    are determined alternatively, via purposefully built codes.    
 
+    Parameters
+    ----------
+    kxx: float, optional
+        Direct stiffness in the x direction.
+        Default is None
+    kxy: float, optional
+        Cross stiffness between xy directions.
+        Default is None
+    kyx: float, optional
+        Cross stiffness between yx directions.
+        Default is None
+    kyy: float, optional
+        Direct stiffness in the y direction.
+        Default is None
+    kzz: float, optional
+        Direct stiffness in the z direction.
+        Default is None
+    cxx: float, optional
+        Direct damping in the x direction.
+        Default is None
+    cxy: float, optional
+        Cross damping between xy directions.
+        Default is None
+    cyx: float, optional
+        Cross damping between yx directions.
+        Default is None
+    cyy: float, optional
+        Direct damping in the y direction.
+        Default is None
+    czz: float, optional
+        Direct damping in the z direction.
+        Default is None
+    tag : str, optional
+        A tag to name the element
+        Default is None
+    Examples
+    --------
+    >>> n = 0
+    >>> kxx = 1.0e7
+    >>> kyy = 1.5e7
+    >>> kzz = 5.0e5
+    >>> tag = "rollerbearing"
+    >>> bearing = BearingElement6DoF(n=n, kxx=kxx, kxy=kxy, kyx=kyx, kyy=kyy, 
+    ...                              cxx=cxx, cxy=cxy, cyx=cyx, cyy=cyy, tag=tag)
+    >>> bearing.K(0)
+    array([[1.0e7, 0.000, 0.000, 0.000, 0.000, 0.000],
+           [0.000, 1.5e7, 0.000, 0.000, 0.000, 0.000],
+           [0.000, 0.000, 5.0e5, 0.000, 0.000, 0.000],
+           [0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
+           [0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
+           [0.000, 0.000, 0.000, 0.000, 0.000, 0.000]]])
+    """
+
+    def __init__(
+        self, n, kxx, kxy=none, kyx=none, kyy, kzz, cxx, cxy=none, cyx=none, cyy, czz, tag=None
+    ):
+
+        super().__init__(
+            n=n,
+            kxx=kxx,
+            kxy=0.0,
+            kyx=0.0,
+            kyy=kyy,
+            kzz=kzz,
+            cxx=cxx,
+            cxy=0.0,
+            cyx=0.0,
+            cyy=cyy,
+            czz=czz,
+            tag=tag,
+        )
+
+        self.color = "#77ACA2"
+
+    def patch(self, position, mean, ax, **kwargs):
+        """Bearing element patch.
+        Patch that will be used to draw the bearing element.
+
+        Parameters
+        ----------
+        position : tuple
+            Position (z, y) in which the patch will be drawn.
+        mean : float
+            Mean value of shaft element outer diameters
[…]
