### 2020-03-25 5ccac70
Add new_units.txt to pypi package
The new_units.txt file was added to the installation, but it was not
available in the pypi distribution.

Closes #471.
Closes #497.
