### 2020-03-25 dbe62e7
Add requirements.txt file
This adds a requirements.txt file to the package. To avoid duplicated
information we read the requirements.txt file when running the setup.py.
