### 2020-03-26 e2c9601
Add requirements.txt file to MANIFEST.in
