+++ ross/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 from .bearing_seal_element import *
 from .disk_element import *
