### 2020-03-26 215423a
Multiple updates to BearingElement6DoF and minor corrections to ShaftElement6DoF
