+++ ross/shaft_element.py
@@ -1194,7 +1194,7 @@ class ShaftElement(Element):
         return elements
 
 
-class ShaftElement6DoF(Element):
+class ShaftElement6DoF(ShaftElement):
 
     r"""A 6 Degrees of Freedom shaft element.
     =========================================
@@ -1285,7 +1285,7 @@ class ShaftElement6DoF(Element):
     >>> Euler_Bernoulli_Element.phi
     0
     """
-
+    
     @check_units
     def __init__(
         self,
@@ -1305,24 +1305,17 @@ class ShaftElement6DoF(Element):
         beta=0,
         tag=None,
     ):
-
+    
         if idr is None:
             idr = idl
         if odr is None:
             odr = odl
 
-        ## Changing units to only their magnitude
-        # L = L.m
-        # idl = idl.m
-        # odl = odl.m
-        # idr = idr.m
-        # odr = odr.m
-
         if material is None:
             raise AttributeError("Material is not defined.")
 
         if type(material) is str:
-            os.chdir(Path(os.path.dirname(ross.__file__)))
+            #os.chdir(Path(os.path.dirname(ross.__file__)))
             self.material = Material.use_material(material)
         else:
             self.material = material
@@ -1339,7 +1332,6 @@ class ShaftElement6DoF(Element):
             self.n_r = n + 1
 
         self.tag = tag
-        self.shear_method_calc = shear_method_calc
 
         self.L = float(L)
         self.o_d = (float(odl) + float(odr)) / 2
@@ -1399,44 +1391,7 @@ class ShaftElement6DoF(Element):
         )
         return p >= 0.2
 
-    def __eq__(self, other):
-        """
-        Equality method for comparasions
-
-        Parameters
-        ----------
-        other : obj
-            parameter for comparasion
-
-        Returns
-        -------
-        True if other is equal to the reference parameter.
-        False if not.
-
-        Example
-        -------
-        >>> from ross.materials import steel
-        >>> le = 0.25
-        >>> i_d = 0
-        >>> o_d = 0.05
-        >>> shaft1 = ShaftElement(
-        ...        le, i_d, o_d, steel, rotary_inertia=True, shear_effects=True
-        ... )
-        >>> shaft2 = ShaftElement(
-        ...        le, i_d, o_d, steel, rotary_inertia=True, shear_effects=True
-        ... )
-        >>> shaft1 == shaft2
-        True
-        """
-        if self.__dict__ == other.__dict__:
-            return True
-        else:
-            return False
-
-    def __hash__(self):
-        return hash(self.tag)
-
-    def save(self, file_name):
+    def save(self, file_name=Path(os.getcwd())):
         """Save shaft elements to toml file.
 
         Parameters
@@ -1452,13 +1407,16 @@ class ShaftElement6DoF(Element):
         >>> shaft1 = ShaftElement(
         ...     le, i_d, o_d, steel, rotary_inertia=True, shear_effects=True
         ... )
-        >>> shaft1.save('ShaftElement.toml')
+        >>> shaft1.save()
         """
-        data = self.load_data(file_name)
-        data["ShaftElement"][str(self.n)] = {
+        
+        data = self.get_data(Path(file_name) / "ShaftElement6DoF.toml")
+        data["ShaftElement6DoF"][str(self.n)] = {
             "L": self.L,
-            "i_d": self.i_d,
-            "o_d": self.o_d,
+            "idl": self.idl,
+            "odl": self.odl,
+            "idr": self.idr,
+            "odr": self.odr,
             "alpha": self.alpha,
             "beta": self.beta,
             "material": self.material.name,
@@ -1469,10 +1427,10 @@ class ShaftElement6DoF(Element):
[…]
