### 2020-03-26 9bbcb7d
updating the 6DoF bearing model
