+++ ross/bearing_seal_element.py
@@ -419,7 +419,7 @@ class BearingElement(Element):
         return dict(x_0=0, y_0=1)
 
     def dof_global_index(self):
-        """Get the global index for a element specific degree of freedom."""
+        """Get the global index for an element specific degree of freedom."""
         global_index = super().dof_global_index()
 
         if self.n_link is not None:
@@ -1297,6 +1297,11 @@ class BearingElement6DoF(BearingElement):
         cxy=0.0,
         cyx=0.0,
         czz=0.0,
+        frequency=None,
+        tag=None,
+        n_link=None,
+        scale_factor=1,
+        color="#355d7a"
     ):
         super().__init__(
         n=n,
@@ -1308,12 +1313,60 @@ class BearingElement6DoF(BearingElement):
         cyy=cyy,
         cxy=cxy,
         cyx=cyx,
+        frequency=frequency,
+        tag=tag,
+        n_link=n_link,
+        scale_factor=scale_factor,
+        color=color
         )
 
-        self.kzz = kzz
-        self.czz = czz
+        new_args = ["kzz", "czz"]
+
+        args_dict = locals()
+        coefficients = {}
+
+        if kzz is None:
+            args_dict["kzz"] = kxx * 0.6 # NSK manufacturer sugestion for deep groove ball bearings
+        if czz is None:
+            args_dict["czz"] = cxx
+
+        for arg in new_args:
+            if arg[0] == "k":
+                coefficients[arg] = _Stiffness_Coefficient(
+                    coefficient=args_dict[arg], w=None
+                )
+            else:
+                coefficients[arg] = _Damping_Coefficient(
+                    args_dict[arg], None
+                )
+
+        coefficients_len = [len(v.coefficient) for v in coefficients.values()]
+
+        for c in coefficients_len:
+            if c != 1:
+                raise ValueError(
+                    "Arguments (coefficients and frequency)"
+                    " must have the same dimension"
+                )
+
+        for k, v in coefficients.items():
+            setattr(self, k, v)
 
     def __repr__(self):
+        """This function returns a string representation of a bearing element.
+        Parameters
+        ----------
+
+        Returns
+        -------
+        A string representation of a bearing object.
+        Examples
+        --------
+        >>> bearing = bearing_example()
+        >>> bearing # doctest: +ELLIPSIS
+        BearingElement(n=0, n_link=None,
+         kxx=[...
+        """
         return (
             f"{self.__class__.__name__}"
             f"(n={self.n}, n_link={self.n_link},\n"
@@ -1367,6 +1420,146 @@ class BearingElement6DoF(BearingElement):
                 )
             )
         return False
+
+    def save(self, file_name=Path(os.getcwd())):
+        """Saves a bearing element in a toml format. It works as an auxiliary
+        function of the save function in the Rotor class.
+
+        Parameters
+        ----------
+        file_name: string
+            The name of the file the bearing element will be saved in.
+
+        Returns
+        -------
+        None
+
+        Examples
+        --------
+        >>> bearing = bearing_example()
+        >>> bearing.save(Path(os.getcwd()))
+        """
+        data = self.get_data(Path(file_name)/'BearingElement6DoF.toml')
+
+        if type(self.frequency) == np.ndarray:
+            try:
+                self.frequency[0]
+                frequency = list(self.frequency)
+            except IndexError:
+                frequency = None
+
+        data["BearingElement6DoF"][str(self.n)] = {
+            "n": self.n,
+            "kxx": self.kxx.coefficient,
[…]
