+++ BearingElement6DoF.toml
@@ -1,0 +1,12 @@
+[BearingElement6DoF.0]
+n = 0
+kxx = [ 1000000.0,]
+cxx = [ 200.0,]
+kyy = [ 800000.0,]
+kxy = [ 0.0,]
+kyx = [ 0.0,]
+kzz = [ 100000.0,]
+cyy = [ 150.0,]
+cxy = [ 0.0,]
+cyx = [ 0.0,]
+czz = [ 50.0,]
