### 2020-03-26 92ea58e
finishing touches on the bearings for the 6 DoF model
