### 2020-03-26 ross-rotordynamics/ross#471 @raphaeltimbo
Hi there @srmnitc ! We have updated the `MANIFEST.in` file to add the missing file to the pypi package. I have tested and it looks like we solved the problem. Feel free to reopen the issue if you have any problems!
