### 2020-03-26 08449fa
Add pandoc to local documentation requirements
(cherry picked from commit 064e07118054394b7c40e135d7eb4c5b3fe03798)
