### 2020-03-26 56c1f8d
Add instructions to set development environment
This also addresses #462, since it gives better instructions on how to
set the development environment to run the tests.

(cherry picked from commit 7e06d52da163de3eb7a8a85057287318dc193ccc)
