+++ CONTRIBUTING.rst
@@ -1,17 +1,17 @@
 Feedback and Contribution
 -------------------------
-We welcome any contribution via `ROSS' issue tracker <https://github.com/ross-rotordynamics/ross/issues>`_.
+We welcome any contribution via `ross's issue tracker <https://github.com/ross-rotordynamics/ross/issues>`_.
 These include bug reports, problems on the documentation, feedback, enhancement proposals etc.
 The issue tracker can also be used for questions and further information since the project does not use a mailing list.
 
 Version-control system: Git
 ---------------------------
 Git is a version control system (VCS) for tracking changes in code during software development.
-To download the ROSS source code and contribute to its development,
+To download the ross source code and contribute to its development,
 you need Git installed in your machine. Refer to the `Git website
 <https://git-scm.com/>`_ and follow the instructions to download and install it.
-Once you have Git installed, you will be able to follow the instructions in `How to contribute to ROSS using git`_,
-which explains how to download and contribute to ROSS.
+Once you have Git installed, you will be able to follow the instructions in `How to contribute to ross using git`_,
+which explains how to download and contribute to ross.
 
 Code style: Black
 -----------------
@@ -54,38 +54,45 @@ After that you can access your local server (http://0.0.0.0:8000/) and see the g
 
 Integrated development environment: PyCharm
 -------------------------------------------
-The ROSS development team adopted PyCharm as integrated development environment (IDE).
-You don't need PyCharm to run or contribute to ROSS, as you can choose your preferred IDE or
+The ross development team adopted PyCharm as integrated development environment (IDE).
+You don't need PyCharm to run or contribute to ross, as you can choose your preferred IDE or
 even no IDE at all. But in case you want to use PyCharm, go to the `PyCharm website
 <https://www.jetbrains.com/pycharm/>`_ to download and install it.
 
-How to contribute to ROSS using git
+How to contribute to ross using git
 -----------------------------------
 .. _git-configuration:
 
-To use git to contribute to ROSS project, follow the steps below:
+To use git to contribute to ross project, follow the steps below:
 *For Windows users: commands provided here can be executed using Git Bash instead of Git GUI.*
 
-Step 1: Make your copy (fork) of ROSS
+Step 1: Make your copy (fork) of ross
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 Go to https://github.com/ross-rotordynamics/ross
 In the top-right corner of the page, click Fork, to fork it to your GitHub account.
 
-From the command line::
+From the command line:
+
+::
 
     git clone https://github.com/your-user-name/ross.git
     cd ross
     git remote add upstream https://github.com/ross-rotordynamics/ross.git
 
+
 Step 2: Keep in sync with changes in Ross
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Setup your local repository, so it pulls from upstream by default::
+Setup your local repository, so it pulls from upstream by default:
+
+::
 
     git config branch.master.remote upstream
     git config branch.master.merge refs/heads/master
 
-This can also be done by editing the config file inside your ross/.git directory.
-It should look like this::
+This can also be done by editing the config file inside your .git directory.
+It should look like this:
+
+::
 
     [core]
             repositoryformatversion = 0
@@ -125,7 +132,7 @@ Step 4: Make a new feature branch
     git fetch upstream
     git checkout -b my-new-feature upstream/master
 
-Step 5: Testing the code
+Step 4: Testing the code
 ^^^^^^^^^^^^^^^^^^^^^^^^
[…]
