### 2020-03-26 476025c
Add instructions to set development environment
This also addresses #462, since it gives better instructions on how to
set the development environment to run the tests.
