+++ ross/tests/test_shaft_element.py
@@ -1,12 +1,5 @@
 import os
 
-
-############
-import sys
-
-sys.path.append(r"C:\Users\Vergilio\Desktop\ross6dof\ross")
-############
-
 import numpy as np
 import pytest
 from numpy.testing import assert_allclose, assert_almost_equal
@@ -382,7 +375,7 @@ def test_s6_M_matrix(s6_eb):
 
 def test_s6_G_matrix(s6_eb):
     # fmt: off
-    G_mat_6DoF = np.array(
+    G_mat_6DoF = 0.01 * np.array(
         [[           0,  -9.7239e+10,            0,  -8.1032e+08,            0,            0,            0,   9.7239e+10,            0,  -8.1032e+08,            0,            0],
          [  9.7239e+10,            0,            0,            0,  -8.1032e+08,            0,  -9.7239e+10,            0,            0,            0,  -8.1032e+08,            0],
          [           0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
@@ -421,18 +414,18 @@ def test_s6_K_matrix(s6_eb):
 
 def test_s6_Kst_matrix(s6_eb):
     # fmt: off
-    Kst_mat_6DoF = np.array(
-        [[          0,  -9.7239e+10,            0,  -8.1032e+08,            0,            0,            0,   9.7239e+10,            0,  -8.1032e+08,            0,            0],
-         [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
-         [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
-         [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
-         [          0,   8.1032e+08,            0,   1.0804e+08,            0,            0,            0,  -8.1032e+08,            0,  -2.7011e+07,            0,            0],
-         [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
-         [          0,   9.7239e+10,            0,   8.1032e+08,            0,            0,            0,  -9.7239e+10,            0,   8.1032e+08,            0,            0],
-         [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
-         [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
-         [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
-         [          0,   8.1032e+08,            0,  -2.7011e+07,            0,            0,            0,  -8.1032e+08,            0,   1.0804e+08,            0,            0],
-         [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0]])
+    Kst_mat_6DoF = 0.01 * np.array([
+        [          0,  -9.7239e+10,            0,  -8.1032e+08,            0,            0,            0,   9.7239e+10,            0,  -8.1032e+08,            0,            0],
+        [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
+        [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
+        [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
+        [          0,   8.1032e+08,            0,   1.0804e+08,            0,            0,            0,  -8.1032e+08,            0,  -2.7011e+07,            0,            0],
+        [          0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0,            0],
+        [          0,   9.7239e+10,            0,   8.1032e+08,            0,            0,            0,  -9.7239e+10,            0,   8.1032e+08,            0,            0],
[…]
