### 2020-03-27 ca1da24
bearings testing implementation and debug
