+++ ross/bearing_seal_element.py
@@ -1231,7 +1231,7 @@ class RollerBearingElement(BearingElement):
 
         self.color = "#77ACA2"
 
-        
+
 class BearingElement6DoF(BearingElement):
     """A generalistic 6 DoF bearing element. This class will create a bearing 
     element based on the user supplied stiffness and damping coefficients. These
@@ -1282,15 +1282,13 @@ class BearingElement6DoF(BearingElement):
     >>> bearing = BearingElement6DoF(n=n, kxx=kxx, kxy=kxy, kyx=kyx, kyy=kyy, 
     ...                              cxx=cxx, cxy=cxy, cyx=cyx, cyy=cyy, tag=tag)
     >>> bearing.K(0)
-    array([[1.0e7, 0.000, 0.000, 0.000, 0.000, 0.000],
-           [0.000, 1.5e7, 0.000, 0.000, 0.000, 0.000],
-           [0.000, 0.000, 5.0e5, 0.000, 0.000, 0.000],
-           [0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
-           [0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
-           [0.000, 0.000, 0.000, 0.000, 0.000, 0.000]]])
+    array([[1.0e7, 0.000, 0.000],
+           [0.000, 1.5e7, 0.000],
+           [0.000, 0.000, 5.0e5]])
     """
 
-    def __init__(self,
+    def __init__(
+        self,
         n,
         kxx,
         kyy,
@@ -1302,27 +1300,25 @@ class BearingElement6DoF(BearingElement):
         cxy=0.0,
         cyx=0.0,
         czz=0.0,
-        frequency=None,
         tag=None,
-        n_link=None,
         scale_factor=1,
-        color="#355d7a"
+        color="#355d7a",
     ):
         super().__init__(
-        n=n,
-        kxx=kxx,
-        cxx=cxx,
-        kyy=kyy,
-        kxy=kxy,
-        kyx=kyx,
-        cyy=cyy,
-        cxy=cxy,
-        cyx=cyx,
-        frequency=frequency,
-        tag=tag,
-        n_link=n_link,
-        scale_factor=scale_factor,
-        color=color
+            n=n,
+            kxx=kxx,
+            cxx=cxx,
+            kyy=kyy,
+            kxy=kxy,
+            kyx=kyx,
+            cyy=cyy,
+            cxy=cxy,
+            cyx=cyx,
+            frequency=None,
+            tag=tag,
+            n_link=None,
+            scale_factor=scale_factor,
+            color=color,
         )
 
         new_args = ["kzz", "czz"]
@@ -1331,7 +1327,9 @@ class BearingElement6DoF(BearingElement):
         coefficients = {}
 
         if kzz is None:
-            args_dict["kzz"] = kxx * 0.6 # NSK manufacturer sugestion for deep groove ball bearings
+            args_dict["kzz"] = (
+                kxx * 0.6
+            )  # NSK manufacturer sugestion for deep groove ball bearings
         if czz is None:
             args_dict["czz"] = cxx
 
@@ -1341,9 +1339,7 @@ class BearingElement6DoF(BearingElement):
                     coefficient=args_dict[arg], w=None
                 )
             else:
-                coefficients[arg] = _Damping_Coefficient(
-                    args_dict[arg], None
-                )
+                coefficients[arg] = _Damping_Coefficient(args_dict[arg], None)
 
         coefficients_len = [len(v.coefficient) for v in coefficients.values()]
 
@@ -1382,7 +1378,6 @@ class BearingElement6DoF(BearingElement):
[…]
