### 2020-03-27 1132793
Citations in chronological order
Issue #483
