### 2020-03-27 22331fd
Remove term "independent software"
Issue #486
