+++ paper/paper.md
@@ -51,10 +51,10 @@ systems [@childs1993turbomachinery; @gasch2006rotordynamik; @friswell2010dynamic
 These models are used to support research and decision making. To this purpose, we present ROSS,
 an open source library written in Python for rotordynamic analysis.
 
-Concerning rotordynamics softwares, there are some commercial finite element softwares that have a rotordynamic 
-module [@comsol; @ansys], some softwares based on a commercial platform (Matlab) [@madyn; @dynamicsrotating], and others 
-that are independent [@rotorinsa; @trcsoftware]. 
-To use these softwares one needs to buy licenses, and they are not intended to be developed in a collaborative public 
+Concerning rotordynamics softwares, there are some commercial finite element softwares that have a rotordynamic
+module [@comsol; @ansys], some softwares based on a commercial third platform (Matlab) [@madyn; @dynamicsrotating], and others
+that have their own platform [@rotorinsa; @trcsoftware].
+To use these softwares one needs to buy licenses, and they are not intended to be developed in a collaborative public
 manner.
 
 ROSS allows the construction of rotor models and their numerical simulation. Shaft elements, as a default, are 
