+++ paper/paper.md
@@ -52,7 +52,7 @@ These models are used to support research and decision making. To this purpose, 
 an open source library written in Python for rotordynamic analysis.
 
 Concerning rotordynamics softwares, there are some commercial finite element softwares that have a rotordynamic
-module [@comsol; @ansys], some softwares based on a commercial third platform (Matlab) [@madyn; @dynamicsrotating], and others
+module [@comsol; @ansys], some softwares based on a commercial third platform (MATLAB) [@madyn; @dynamicsrotating], and others
 that have their own platform [@rotorinsa; @trcsoftware].
 To use these softwares one needs to buy licenses, and they are not intended to be developed in a collaborative public
 manner.
