### 2020-03-27 b8b694a
disk 6dof test functions and debug
