+++ ross/disk_element.py
@@ -559,7 +559,7 @@ def disk_example():
     return disk
 
 
-class DiskElement6DoF(Element):
+class DiskElement6DoF(DiskElement):
     """A disk element for 6 DoFs.
 
      This class will create a disk element with 6 DoF from input data of inertia and mass.
@@ -580,144 +580,10 @@ class DiskElement6DoF(Element):
 
      Examples
      --------
-     >>> disk = DiskElement(n=0, m=32, Id=0.2, Ip=0.3)
+     >>> disk = DiskElement6DoF(n=0, m=32, Id=0.2, Ip=0.3)
      >>> disk.Ip
      0.3
      """
-
-    def __init__(self, n, m, Id, Ip, tag=None):
-        self.n = int(n)
-        self.n_l = n
-        self.n_r = n
-
-        self.m = m
-        self.Id = Id
-        self.Ip = Ip
-        self.tag = tag
-        self.color = bokeh_colors[9]
-        self.dof_global_index = None
-
-    def __eq__(self, other):
-        """This function allows disk elements to be compared.
-        Parameters
-        ----------
-        other: object
-            The second object to be compared with.
-
-        Returns
-        -------
-        bool
-            True if the comparison is true; False otherwise.
-        Examples
-        --------
-        >>> disk1 = disk_example()
-        >>> disk2 = disk_example()
-        >>> disk1 == disk2
-        True
-        """
-        false_number = 0
-        for i in self.__dict__:
-            try:
-                if np.allclose(self.__dict__[i], other.__dict__[i]):
-                    pass
-                else:
-                    false_number += 1
-
-            except TypeError:
-                if self.__dict__[i] == other.__dict__[i]:
-                    pass
-                else:
-                    false_number += 1
-
-        if false_number == 0:
-            return True
-        else:
-            return False
-
-    def __repr__(self):
-        """This function returns a string representation of a disk element.
-        Parameters
-        ----------
-
-        Returns
-        -------
-        A string representation of a disk object.
-        Examples
-        --------
-        >>> disk = disk_example()
-        >>> disk # doctest: +ELLIPSIS
-        DiskElement(Id=0.17809, Ip=0.32956...
-        """
-        return (
-            f"{self.__class__.__name__}"
-            f"(Id={self.Id:{0}.{5}}, Ip={self.Ip:{0}.{5}}, "
-            f"m={self.m:{0}.{5}}, color={self.color!r}, "
-            f"n={self.n}, tag={self.tag!r})"
-        )
-
-    def __hash__(self):
-        return hash(self.tag)
-
-    def save(self, file_name):
-        """Saves a disk element in a toml format.
-
-        It works as an auxiliary function of the save function in the Rotor
-        class.
-
-        Parameters
-        ----------
-        file_name: string
-            The name of the file the disk element will be saved in.
-
-        Returns
-        -------
-        None
-
-        Examples
-        --------
-        >>> disk = disk_example()
-        >>> disk.save('DiskElement6DoF.toml')
-        """
-        data = self.load_data(file_name)
-        data["DiskElement6DoF"][str(self.n)] = {
-            "n": self.n,
-            "m": self.m,
-            "Id": self.Id,
-            "Ip": self.Ip,
-            "tag": self.tag,
-        }
-        self.dump_data(data, file_name)
-
-    @staticmethod
-    def load(file_name="DiskElement6DoF"):
-        """Loads a list of 6DoFs disk elements saved in a toml format.
-
-        Parameters
-        ----------
-        file_name: str
-            The name of the file of the disk element to be loaded.
-
[…]
