### 2020-03-27 740c750
Add reference to Timoshenko beam theory
Issues #475 and #487
