### 2020-03-27 16276f3
Add reference to Campbell Diagram
Issue #475
