### 2020-03-27 17c399c
Listing equations terms
Issue #489
