### 2020-03-27 6bdcf8d
Simple python packages explaination
Issue #490
