+++ paper/paper.md
@@ -98,13 +98,13 @@ unbalance or misalignment, and then the output data from this model is used to t
 and see how good it is in diagnosing the problem; ROSS is also interesting for model updating.
 Here we could use vibrations measurements from a rotor to update a model generated by ROSS.
 
-We have built the package using main Python packages such as NumPy [@van2011numpy], SciPy [@jones2001scipy] 
-and Bokeh [@bokeh2019].
+We have built the package using main Python packages such as NumPy [@van2011numpy] and SciPy [@jones2001scipy],
+for their mathematical, engineering and array building tools; and Bokeh [@bokeh2019] for creating interactive plots.
 
-Besides the [documentation](https://ross-rotordynamics.github.io/ross-website/), a set of Jupyter Notebooks 
+Besides the [documentation](https://ross-rotordynamics.github.io/ross-website/), a set of Jupyter Notebooks
 is available for the tutorial and some examples. Users can also access these notebooks through a [Binder server](https://mybinder.org/v2/gh/ross-rotordynamics/ross/master).
 
-As an example, Figure 1 shows a centrifugal compressor modeled with ROSS. 
+As an example, Figure 1 shows a centrifugal compressor modeled with ROSS.
 
 ![Centrifugal Compressor modeled with ROSS.](rotor_plot.png)
 
