+++ paper/paper.md
@@ -108,17 +108,17 @@ As an example, Figure 1 shows a centrifugal compressor modeled with ROSS.
 
 ![Centrifugal Compressor modeled with ROSS.](rotor_plot.png)
 
-The shaft elements are in gray, 
-the impellers represented as disks are in blue and the bearings are displayed as springs and dampers. This plot is generated with Bokeh, 
+The shaft elements are in gray,
+the impellers represented as disks are in blue and the bearings are displayed as springs and dampers. This plot is generated with Bokeh,
 and we can use the hover tool to get additional information on each element.
 
-One analysis that can be carried out is the modal analysis. Figure 2 shows the Campbell Diagram generated for this 
+One analysis that can be carried out is the modal analysis. Figure 2 shows the Campbell Diagram generated for this
 compressor; the natural frequencies and the log dec vary with the machine rotation speed.
 
 ![Campbell Diagram for the Centrifugal Compressor.](campbell.png)
 
-The Campbell Diagram is one of the results that can be obtained from the model, other types of analysis can be found 
-on the [documentation](https://ross-rotordynamics.github.io/ross-website/). The software is extensible and new elements 
+The Campbell Diagram is one of the results that can be obtained from the model, other types of analysis can be found
+in the [documentation](https://ross-rotordynamics.github.io/ross-website/). The software is extensible and new elements
 or new types of analysis can be easily included.
 
 # Acknowledgements
