### 2020-03-27 2a93aa5
Use generator to storage elements
