### 2020-03-27 ce3f036
Use generator to store rotors
