+++ ross/stochastic/st_rotor_assembly.py
@@ -56,7 +56,7 @@ class ST_Rotor(object):
 
     Examples
     --------
-    >>> # Rotor with 2 shaft elements, 1 random disk element and 2 bearings
+    # Rotor with 2 shaft elements, 1 random disk element and 2 bearings
     >>> import numpy as np
     >>> import ross as rs
     >>> import ross.stochastic as srs
@@ -83,7 +83,7 @@ class ST_Rotor(object):
     >>> bearing0 = rs.BearingElement(0, kxx=stf, cxx=0)
     >>> bearing1 = rs.BearingElement(2, kxx=stf, cxx=0)
     >>> rand_rotor = srs.ST_Rotor(shaft_elm, [disk0], [bearing0, bearing1])
-    >>> len(rand_rotor.rotor_list)
+    >>> len(list(rand_rotor.__iter__()))
     5
     """
 
@@ -117,8 +117,8 @@ class ST_Rotor(object):
             it = iter(
                 [elm for elm in shaft_elements if isinstance(elm, ST_ShaftElement)]
             )
-            the_len = len(next(it).elements)
-            if not all(len(l.elements) == the_len for l in it):
+            the_len = len(list(next(it).__iter__()))
+            if not all(len(list(l)) == the_len for l in it):
                 raise ValueError(
                     "not all random shaft elements lists have same length."
                 )
@@ -127,8 +127,8 @@ class ST_Rotor(object):
             is_random.append("disk_elements")
 
             it = iter([elm for elm in disk_elements if isinstance(elm, ST_DiskElement)])
-            the_len = len(next(it).elements)
-            if not all(len(l.elements) == the_len for l in it):
+            the_len = len(list(next(it).__iter__()))
+            if not all(len(list(l)) == the_len for l in it):
                 raise ValueError("not all random disk elements lists have same length.")
 
         if any(isinstance(elm, ST_BearingElement) for elm in bearing_elements):
@@ -137,8 +137,8 @@ class ST_Rotor(object):
             it = iter(
                 [elm for elm in bearing_elements if isinstance(elm, ST_BearingElement)]
             )
-            the_len = len(next(it).elements)
-            if not all(len(l.elements) == the_len for l in it):
+            the_len = len(list(next(it).__iter__()))
+            if not all(len(list(l)) == the_len for l in it):
                 raise ValueError(
                     "not all random bearing elements lists have same length."
                 )
@@ -149,25 +149,25 @@ class ST_Rotor(object):
             it = iter(
                 [elm for elm in point_mass_elements if isinstance(elm, ST_PointMass)]
             )
-            the_len = len(next(it).elements)
-            if not all(len(l.elements) == the_len for l in it):
+            the_len = len(list(next(it).__iter__()))
+            if not all(len(list(l)) == the_len for l in it):
                 raise ValueError("not all random point mass lists have same length.")
 
         for i, elm in enumerate(shaft_elements):
             if isinstance(elm, ST_ShaftElement):
-                shaft_elements[i] = elm.elements
+                shaft_elements[i] = list(elm.__iter__())
 
         for i, elm in enumerate(disk_elements):
             if isinstance(elm, ST_DiskElement):
-                disk_elements[i] = elm.elements
+                disk_elements[i] = list(elm.__iter__())
 
         for i, elm in enumerate(bearing_elements):
             if isinstance(elm, ST_BearingElement):
-                bearing_elements[i] = elm.elements
+                bearing_elements[i] = list(elm.__iter__())
 
         for i, elm in enumerate(point_mass_elements):
             if isinstance(elm, ST_PointMass):
-                point_mass_elements[i] = elm.elements
+                point_mass_elements[i] = list(elm.__iter__())
[…]
