### 2020-03-27 22d9fbf
Fix time response example
