### 2020-03-27 43d589e
Fix from_geometry example
