### 2020-03-30 a01f8f4
Update - Remove term "independent software"
Issue #486
