### 2020-03-30 f115983
Update - Add reference to Campbell Diagram
Issue #475
Change term's name to whirl speed map
