+++ paper/paper.md
@@ -55,12 +55,13 @@ Concerning rotordynamics softwares, there are some commercial finite element sof
 module [@comsol; @ansys], some softwares based on a proprietary commercial software (MATLAB) [@madyn; @dynamicsrotating],
 and others developed as standalone softwares [@rotorinsa; @trcsoftware].
 To use these softwares one needs to buy licenses, and they are not intended to be developed in a collaborative public
-manner.
+manner. Currently, there are no softwares being developed using the open source concept, with the code being fully
+available on code hosting platforms, issues tracked online, clear license, possibility of direct contribution, etc.
 
 ROSS allows the construction of rotor models and their numerical simulation. Shaft elements, as a default, are
-modeled with the Timoshenko beam theory [@Hutchinson2001], which considers shear and rotary inertia effects, and
-discretized by means of the Finite Element Method [@friswell2010dynamics]. Disks are assumed to be rigid bodies,
-thus their strain energy is not taken into account. And bearings/seals are included as linear stiffness/damping coefficients.
+modeled with the Timoshenko beam theory [@Hutchinson2001], which considers shear and rotary inertia effects, and discretized by means of
+the Finite Element Method [@friswell2010dynamics]. Disks are assumed to be rigid bodies, thus their strain energy is not taken
+into account. And bearings/seals are included as linear stiffness/damping coefficients.
 
 After defining the element matrices and assembling the global matrices of the system, ROSS draws the rotor geometry,
 runs simulations, and obtains results in the form of graphics. It performs several analyses, such as static analysis,
