### 2020-03-30 ce7b54a
Examples of extensibility
Issue #492
