### 2020-03-30 cd02685
Replace Campbell Diagram with whirl speed map
