### 2020-03-30 c0d376e
Removing deprecations and docsting improvements
* Replace bokeh.widgetbox with bokeh.Column
* Update docstrings
