+++ ross/api_report.py
@@ -1,11 +1,10 @@
 # fmt: off
 from copy import copy, deepcopy
-from pathlib import Path
 
 import bokeh.palettes as bp
 import numpy as np
 import pandas as pd
-from bokeh.layouts import gridplot, widgetbox
+from bokeh.layouts import gridplot, Column
 from bokeh.models import (ColumnDataSource, DataRange1d, HoverTool, Label,
                           LinearAxis, Range1d, Span)
 from bokeh.models.widgets import (DataTable, NumberFormatter, Panel,
@@ -14,18 +13,98 @@ from bokeh.plotting import figure
 from scipy.interpolate import interp1d
 from scipy.signal import argrelextrema
 
-import ross as rs
 from ross.bearing_seal_element import BearingElement, SealElement
+from ross.disk_element import DiskElement
 from ross.materials import steel
-from ross.rotor_assembly import Rotor, rotor_example
+from ross.rotor_assembly import Rotor
+from ross.shaft_element import ShaftElement
 
 # fmt: on
 
 # set bokeh palette of colors
 bokeh_colors = bp.RdGy[11]
 
+__all__ = ["Report", "report_example"]
+
 
 class Report:
+    """Report according to standard analysis.
+
+    - Perform unbalance response
+    - Perform Stability_level1 analysis
+    - Apply Level 1 Screening Criteria
+    - Perform Stability_level2 analysis
+
+    Parameters
+    ----------
+    rotor : object
+        A rotor built from rotor_assembly.
+    speed_range : tuple
+        Tuple with (min, max) for speed range.
+    tripspeed : float
+        Machine trip speed.
+    bearing_stiffness_range : tuple, optional
+        Tuple with (start, end) bearing stiffness range.
+        Argument to calculate the Undamped Critical Speed Map.
+    bearing_clearance_lists : list of lists, optional
+        List with two bearing elements lists:
+            The first bearing list is set for minimum clearance.
+            The second bearing list it set for maximum clearance.
+    machine_type : str
+        Machine type analyzed. Options: compressor, turbine or axial_flow.
+        If other option is given, it will be treated as a compressor
+        Default is compressor
+    speed_units : str
+        String defining the unit for rotor speed.
+        Default is "rpm".
+    tag : str
+        String to name the rotor model
+        Default is the Rotor.tag attribute
+
+    Attributes
+    ----------
+    rotor_type: str
+        Defines if the rotor is between bearings or overhung
+    disk_nodes: list
+        List of disk between bearings or overhung (depending on the
+        rotor type)
+
+    Returns
+    -------
+    A Report object
+
+    Examples
+    --------
+    >>> import ross as rs
+    >>> rotor = rs.rotor_example()
+    >>>
+    >>> # coefficients for minimum clearance
+    >>> stfx = [0.7e7, 0.8e7, 0.9e7, 1.0e7]
+    >>> damp = [2.0e3, 1.9e3, 1.8e3, 1.7e3]
+    >>> freq = [400, 800, 1200, 1600]
+    >>> bearing0 = rs.BearingElement(0, kxx=stfx, cxx=damp, frequency=freq)
+    >>> bearing1 = rs.BearingElement(6, kxx=stfx, cxx=damp, frequency=freq)
+    >>> min_clearance_brg = [bearing0, bearing1]
+    >>>
+    >>> # coefficients for maximum clearance
+    >>> stfx = [0.4e7, 0.5e7, 0.6e7, 0.7e7]
+    >>> damp = [2.8e3, 2.7e3, 2.6e3, 2.5e3]
+    >>> freq = [400, 800, 1200, 1600]
[…]
