### 2020-03-30 ross-rotordynamics/ross#472 @rodrigomoliveira1
Hello @srmnitc,

I believe these 2 PR's should end those `DeprecationWarning`.
