### 2020-03-30 ea4544d
Add new file - st_materials.py
Module to create materials with random parameters
