### 2020-03-30 364da71
Add interface with st_materials to shaft_element
