+++ paper/paper.bib
@@ -52,7 +52,7 @@
   number={4},
   pages={165},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/s40430-019-1652-8},
 }
 
