### 2020-03-30 07f5c41
Class MagneticBearingElement was created
