+++ ross/bearing_seal_element.py
@@ -21,6 +21,7 @@ __all__ = [
     "SealElement",
     "BallBearingElement",
     "RollerBearingElement",
+    "MagneticBearingElement",
 ]
 # fmt: on
 bokeh_colors = bp.RdGy[11]
@@ -1061,7 +1062,7 @@ class BallBearingElement(BearingElement):
     """A bearing element for ball bearings.
     This class will create a bearing element based on some geometric and
     constructive parameters of ball bearings. The main difference is that
-    cross-coupling stiffness and damping are not modeled in this case.    
+    cross-coupling stiffness and damping are not modeled in this case.
 
     Parameters
     ----------
@@ -1147,7 +1148,7 @@ class RollerBearingElement(BearingElement):
     """A bearing element for roller bearings.
     This class will create a bearing element based on some geometric and
     constructive parameters of roller bearings. The main difference is that
-    cross-coupling stiffness and damping are not modeled in this case.    
+    cross-coupling stiffness and damping are not modeled in this case.
 
     Parameters
     ----------
@@ -1229,6 +1230,205 @@ class RollerBearingElement(BearingElement):
         )
 
         self.color = "#77ACA2"
+
+
+class MagneticBearingElement(BearingElement):
+    """Magnetic bearing.
+
+    This class creates a magnetic bearing element.
+    Parameters can be a constant value or speed dependent.
+    For speed dependent parameters, each argument should be passed
+    as an array and the correspondent speed values should also be
+    passed as an array.
+
+     Parameters
+     ----------
+    n: int
+        Node which the bearing will be located in
+    kxx: float, array
+        Direct stiffness in the x direction.
+    cxx: float, array
+        Direct damping in the x direction.
+    kyy: float, array, optional
+        Direct stiffness in the y direction.
+        (defaults to kxx)
+    cyy: float, array, optional
+        Direct damping in the y direction.
+        (defaults to cxx)
+    kxy: float, array, optional
+        Cross coupled stiffness in the x direction.
+        (defaults to 0)
+    cxy: float, array, optional
+        Cross coupled damping in the x direction.
+        (defaults to 0)
+    kyx: float, array, optional
+        Cross coupled stiffness in the y direction.
+        (defaults to 0)
+    cyx: float, array, optional
+        Cross coupled damping in the y direction.
+        (defaults to 0)
+    frequency: array, optional
+        Array with the speeds (rad/s).
+    tag : str, optional
+        A tag to name the element
+        Default is None
+
+    Examples
+    --------
+    >>> # A seal element located in the first rotor node, with these
+    >>> # following stiffness and damping coefficients and speed range from
+    >>> # 0 to 200 rad/s
+    >>> kxx = 1e3
+    >>> kyy = 1e3
+    >>> kxy = 1e1
+    >>> kyx = 1e1
+    >>> cxx = 1e3
+    >>> cyy = 1e3
+    >>> cxy = 1e1
+    >>> cyx = 1e3
+    >>> frequency = np.linspace(0, 100, 4)
+    >>> mb = MagneticBearingElement(n=0,kxx=kxx,kyy=kyy,kxy=kxy,kyx=kyx,cxx=cxx,cyy=cyy,cxy=cxy,cyx=cyx,frequency=frequency)
+    >>> mb.kxx
+    ['1.00e+03', '1.00e+03', '1.00e+03', '1.00e+03']
+    """
+
+    def __init__(
+        self,
+        n,
+        kxx,
+        cxx,
+        kyy=None,
[…]
