### 2020-03-30 d152fbe
Class MagneticBearingElement run with Black and Isort
