+++ paper/paper.md
@@ -55,13 +55,15 @@ Concerning rotordynamics softwares, there are some commercial finite element sof
 module [@comsol; @ansys], some softwares based on a proprietary commercial software (MATLAB) [@madyn; @dynamicsrotating],
 and others developed as standalone softwares [@rotorinsa; @trcsoftware].
 To use these softwares one needs to buy licenses, and they are not intended to be developed in a collaborative public
-manner. Currently, there are no softwares being developed using the open source concept, with the code being fully
-available on code hosting platforms, issues tracked online, clear license, possibility of direct contribution, etc.
+manner. To our knowledge, in the rotordynamic field, ROSS is the first software being developed using the open source 
+concept, with the code being fully available on code hosting platforms, issues tracked online, clear license and 
+possibility of direct contribution.
 
 ROSS allows the construction of rotor models and their numerical simulation. Shaft elements, as a default, are
 modeled with the Timoshenko beam theory [@Hutchinson2001], which considers shear and rotary inertia effects, and discretized by means of
 the Finite Element Method [@friswell2010dynamics]. Disks are assumed to be rigid bodies, thus their strain energy is not taken
-into account. And bearings/seals are included as linear stiffness/damping coefficients.
+into account. And bearings/seals are included as linear stiffness/damping coefficients. ROSS is extensible and new
+elements can be added to the code.
 
 After defining the element matrices and assembling the global matrices of the system, ROSS draws the rotor geometry,
 runs simulations, and obtains results in the form of graphics. It performs several analyses, such as static analysis,
@@ -87,13 +89,10 @@ where:
 - $\omega$ is the rotor whirl speed;
 - $\mathbf{f}$ is the generalized force vector.
 
-ROSS has been used to evaluate the impact of damper seal coefficients uncertainties in a compressor rotordynamics [@timbo2019]. 
-Other projects are also using ROSS to test machine learning algorithms that can diagnose machine failure. Here we can create
-a model with some problem such as unbalance or misalignment and then use the output data from this model to test the learning
-algorithm and see how good it is in diagnosing the problem.
-
 We have built the package using main Python packages such as NumPy [@van2011numpy] and SciPy [@jones2001scipy],
-for their mathematical, engineering and array building tools; and Bokeh [@bokeh2019] for creating interactive plots.
+for their mathematical, engineering and array building tools; and Bokeh [@bokeh2019] for creating interactive plots. Developing the software using Python and its scientific ecosystem enables the user to also make use of this ecosystem,
+making it easier to run rotordynamics analysis. It is also easier to integrate the code into other programs, since we
+only use open source packages and do not depend on proprietary commercial platforms.
 
 Besides the [documentation](https://ross-rotordynamics.github.io/ross-website/), a set of Jupyter Notebooks
 is available for the tutorial and some examples. Users can also access these notebooks through a [Binder server](https://mybinder.org/v2/gh/ross-rotordynamics/ross/master).
@@ -114,10 +113,10 @@ compressor; the natural frequencies and the log dec vary with the machine rotati
 The whirl speed map is one of the results that can be obtained from the model, other types of analysis can be found
 in the [documentation](https://ross-rotordynamics.github.io/ross-website/).
 
-Developing the software using python and its scientific ecosystem enables the user to also make use this ecosystem,
-making it easier to run rotordynamics analysis. It is also easier to integrate the code into other programs, since we
-only use open source packages and do not depend on proprietary commercial platforms. Therefore, ROSS is extensible to new
-elements (e.g. beam, disk or bearing elements with 6 degrees of freedom) and types of analysis (e.g. torsional analysis).
[…]
