### 2020-03-31 efa6a7c
Small modifications to the text
