### 2020-03-31 724baa9
Add whirl speed map (Campbell Diagram)
