### 2020-04-01 297b46e
add `.get_args()` method
* This method has been added to make `_random_var()` cleaner, and to fix a bug when using random variables, from elements and from analysis' methods, at the same time.
* import Iterable from collecions.abc instead of collections
