### 2020-04-01 2d3e54d
Docstring adjustments in st_rotor_assembly.py
