### 2020-04-02 3941ca7
Add more detailed explanation on model assumptions
