+++ paper/paper.bib
@@ -38,7 +38,7 @@
     volume = {68},
     number = {1},
     pages = {87-92},
-    year = {2000},
+    year = {2001},
     month = {08},
     issn = {0021-8936},
     doi = {10.1115/1.1349417},
