### 2020-04-02 ross-rotordynamics/ross#480 @raphaeltimbo
Unfortunately, we could not find the doi for this reference.
