### 2020-04-02 ross-rotordynamics/ross#481 @raphaeltimbo
We could not find a doi for this reference. We have copied the citation information from bokeh’s website (https://bokeh.org/citation/)

![image](https://user-images.githubusercontent.com/18506378/78302806-f1368f80-7511-11ea-8001-d5b29816b73e.png)
