### 2020-04-02 ross-rotordynamics/ross#483 @raphaeltimbo
Based on #484, we have kept only the Friswell, Penny, Garvey & Lees, 2010 reference, so this should not be a problem anymore.
