### 2020-04-02 ross-rotordynamics/ross#484 @raphaeltimbo
They have a similar approach, so we have decided to remove the citations and keep only Friswell, 2010, which we have used also to improve the explanation on the modeling assumptions (#488).
