### 2020-04-02 ross-rotordynamics/ross#485 @raphaeltimbo
This has been fixed on commit 16a0394.
