### 2020-04-02 ross-rotordynamics/ross#486 @raphaeltimbo
The purpose of the ‘independent’ here was to say that these softwares do not depend on other softwares such as COMSOL, ANSYS or MATLAB.

We have modified the text:
![image](https://user-images.githubusercontent.com/18506378/78303011-44a8dd80-7512-11ea-9947-ee1f2566416c.png)
