### 2020-04-02 ross-rotordynamics/ross#487 @raphaeltimbo
We have added the Hutchinson (2001) reference, which explains the Timoshenko beam theory.
