### 2020-04-02 ross-rotordynamics/ross#488 @raphaeltimbo
We have added more explanation to the modeling assumptions:
