### 2020-04-02 ross-rotordynamics/ross#489 @raphaeltimbo
We have tried to improve this on commit 17c399c.
