### 2020-04-02 ross-rotordynamics/ross#490 @raphaeltimbo
We have added a short explanation on why we are using each package.
