### 2020-04-02 ross-rotordynamics/ross#491 @raphaeltimbo
Fixed on commit 762c990.
