### 2020-04-02 ross-rotordynamics/ross#492 @raphaeltimbo
We have added the following to the text:

_"ROSS is extensible and new elements, such as different types of bearings or seals, can be added to the code."_
