### 2020-04-02 ross-rotordynamics/ross#493 @raphaeltimbo
We have added the following to the text:

_“ROSS has been used to evaluate the impact of damper seal coefficients uncertainties in a compressor rotordynamics [@timbo2019].
Other projects are also using ROSS to test machine learning algorithms that can diagnose machine failure. Here we can create a model with some problem such as unbalance or misalignment and then use the output data from this model to test the learning algorithm and check the accuracy of this algorithm in diagnosing the problem.”_
