### 2020-04-02 ross-rotordynamics/ross#494 @raphaeltimbo
We have modified the text and the following two parts address this JOSS requirement:

“To our knowledge, in the rotordynamic field, ROSS is the first software being developed using the open source concept, with the code being fully available on code hosting platforms, issues tracked online, clear license and possibility of direct contribution.”

“Developing the software using Python and its scientific ecosystem enables the user to also make use of this ecosystem, making it easier to run rotordynamics analysis. It is also easier to integrate the code into other programs, since we only use open source packages and do not depend on proprietary commercial platforms.”
