### 2020-04-02 ross-rotordynamics/ross#495 @raphaeltimbo
I will add here some parts of the text that we have added/modified to make this clearer:

_“Computational mechanical models aim to simulate the behavior of such mechanical systems. These models are used to support research and decision making. To this purpose, we present ROSS, an open source library written in Python for rotordynamic analysis.”_

_“To our knowledge, in the rotordynamic field, ROSS is the first software being developed using the open source concept, with the code being fully available on code hosting platforms, issues tracked online, clear license and possibility of direct contribution.”_

_“Developing the software using Python and its scientific ecosystem enables the user to also make use of this ecosystem, making it easier to run rotordynamics analysis. It is also easier to integrate the code into other programs, since we only use open source packages and do not depend on proprietary commercial platforms.”_
