### 2020-04-02 ross-rotordynamics/ross#464 @raphaeltimbo
Black is now listed as a dev dep.
