### 2020-04-02 2fc207d
Add binder link to docs page
