+++ docs/download.rst
@@ -4,3 +4,11 @@ Download notebooks
 Here you can download the tutorial and examples notebooks.
 
 | :download:`Tutorial </examples/tutorial.ipynb>`
+| :download:`Example 1 </examples/example_1.ipynb>`
+| :download:`Example 2 </examples/example_2.ipynb>`
+| :download:`Example 3 </examples/example_3.ipynb>`
+| :download:`Example 4 </examples/example_4.ipynb>`
+| :download:`Example 5 </examples/example_5.ipynb>`
+| :download:`Example 6 </examples/example_6.ipynb>`
+| :download:`Example 7 </examples/example_7.ipynb>`
+| :download:`Example 8 </examples/example_8.ipynb>`
