### 2020-04-03 ross-rotordynamics/ross#475 @srmnitc
Hi @raphaeltimbo how is this issue progressing? It would be last issue I have, once this is fixed, please let me know and I would be happy to go through once more and finalize the review.
