### 2020-04-03 ross-rotordynamics/ross#475 @rodrigomoliveira1
Hi, @srmnitc. We have solved this issue with these last commits.
- Commit [740c750](https://github.com/rodrigomoliveira1/ross/commit/740c7509b314799e8f7c25bbca0ea56a5bf40859) - adds a reference for Timoshenko beam theory.
- Commits [724baa9](https://github.com/ross-rotordynamics/ross/commit/724baa9e1b9f78f050d1fc9120390ebd261504e8) and [f115983](https://github.com/rodrigomoliveira1/ross/commit/f1159831ab243c31c9b1ae805e9baa180bcdd045) - change the map's name to "whirl speed map", which is more general.
