### 2020-04-03 148eec3
Add test file for elements from STOCHASTIC ROSS
