+++ ross/tests/test_stochastic_elements.py
@@ -1,0 +1,181 @@
+"""Tests file.
+
+Tests for:
+    st_shaft_element.py
+    st_disk_element.py
+    st_bearing_seal_element.py
+    st_point_mass.py
+"""
+
+import numpy as np
+import pytest
+from ross.stochastic.st_bearing_seal_element import ST_BearingElement
+from ross.stochastic.st_disk_element import ST_DiskElement
+from ross.stochastic.st_materials import ST_Material
+from ross.stochastic.st_point_mass import ST_PointMass
+from ross.stochastic.st_shaft_element import ST_ShaftElement
+
+
+@pytest.fixture
+def rand_shaft():
+    E = [209e9, 211e9]
+    st_steel = ST_Material(name="Steel", rho=7810, E=E, G_s=81.2e9)
+    elm = ST_ShaftElement(
+        L=[1.0, 1.1],
+        idl=[0.01, 0.02],
+        odl=[0.1, 0.2],
+        material=st_steel,
+        is_random=["L", "idl", "odl", "material"],
+    )
+    return elm
+
+
+@pytest.fixture
+def rand_disk_from_inertia():
+    elm = ST_DiskElement(
+        n=1, m=[30, 40], Id=[0.2, 0.3], Ip=[0.5, 0.7], is_random=["m", "Id", "Ip"],
+    )
+    return elm
+
+
+@pytest.fixture
+def rand_disk_from_geometry():
+    E = [209e9, 211e9]
+    st_steel = ST_Material(name="Steel", rho=7810, E=E, G_s=81.2e9)
+    elm = ST_DiskElement.from_geometry(
+        n=1,
+        material=st_steel,
+        width=[0.07, 0.08],
+        i_d=[0.05, 0.06],
+        o_d=[0.30, 0.40],
+        is_random=["material", "width", "i_d", "o_d"],
+    )
+    return elm
+
+
+@pytest.fixture
+def rand_bearing_constant_coefs():
+    kxx = [1e6, 2e6]
+    cxx = [1e3, 2e3]
+    elm = ST_BearingElement(n=1, kxx=kxx, cxx=cxx, is_random=["kxx", "cxx"],)
+
+    return elm
+
+
+@pytest.fixture
+def rand_bearing_varying_coefs():
+    kxx = [[1e6, 1.1e6], [2e6, 2.1e6]]
+    kxy = [[1.5e6, 1.6e6], [2.5e6, 2.6e6]]
+    kyx = [[1.5e6, 1.6e6], [2.5e6, 2.6e6]]
+    kyy = [[3e6, 3.1e6], [4e6, 4.1e6]]
+    cxx = [[1e3, 1.1e3], [2e3, 2.1e3]]
+    cxy = [[1.5e3, 1.6e3], [2.5e3, 2.6e3]]
+    cyx = [[1.5e3, 1.6e3], [2.5e3, 2.6e3]]
+    cyy = [[3e3, 3.1e3], [4e3, 4.1e3]]
+    frequency = np.array([500, 800])
+    elm = ST_BearingElement(
+        n=1,
+        kxx=kxx,
+        kxy=kxy,
+        kyx=kyx,
+        kyy=kyy,
+        cxx=cxx,
+        cxy=cxy,
+        cyx=cyx,
+        cyy=cyy,
+        frequency=frequency,
+        is_random=["kxx", "kxy", "kyx", "kyy", "cxx", "cxy", "cyx", "cyy"],
+    )
+
+    return elm
+
+
+@pytest.fixture
+def rand_point_mass():
+    mx = [2.0, 2.5]
+    my = [3.0, 3.5]
+    elm = ST_PointMass(n=1, mx=mx, my=my, is_random=["mx", "my"],)
+    return elm
+
+
+def test_st_shaft_element(rand_shaft):
[…]
