### 2020-04-03 ae8bad9
Add test file for materials from STOCHASTIC ROSS
