### 2020-04-03 b437b3e
Add test file for rotor assembly from STOCHASTIC ROSS
