### 2020-04-03 3180095
Fix issue with check_units
* Add a condition to verify if the argument value is None. It allows the user to pass "None" args without getting an error with the code trying to assign a unit to a None value.
