### 2020-04-03 0bc6064
Fix tests for elements and isort issue
