### 2020-04-03 fcca258
Fix tests for st_materials and isort issue
