### 2020-04-04 ross-rotordynamics/ross#495 @nnadeau
@raphaeltimbo :

- Please link the PR/commit that adds these statements
- `ROSS is the first software being developed using the open source concept`, how is that novel compared to `commercial finite element softwares that have a rotordynamic module[s]`?
- While I personally love Python for its general ease of use, I'm not convinced by the statement `Developing the software using Python and its scientific ecosystem enables the user to also make use of this ecosystem, making it easier to run rotordynamics analysis`.
