### 2020-04-04 ross-rotordynamics/ross#494 @nnadeau
@raphaeltimbo please link the PR/commit that adds these statements.

Similar to https://github.com/ross-rotordynamics/ross/issues/495, I'm not convinced by:

```
“To our knowledge, in the rotordynamic field, ROSS is the first software being developed using the open source concept, with the code being fully available on code hosting platforms, issues tracked online, clear license and possibility of direct contribution.”

“Developing the software using Python and its scientific ecosystem enables the user to also make use of this ecosystem, making it easier to run rotordynamics analysis. It is also easier to integrate the code into other programs, since we only use open source packages and do not depend on proprietary commercial platforms.”
```

How is that scientifically novel? Cost is incurred whether upfront (e.g., paying for proprietary software) or in the long run (e.g., free software that is less efficient). What differentiates ROSS from `commercial finite element softwares that have a rotordynamic module` ?
