### 2020-04-04 ross-rotordynamics/ross#493 @nnadeau
@raphaeltimbo please link the PR/commit that adds this statement.
