### 2020-04-04 ross-rotordynamics/ross#492 @nnadeau
@raphaeltimbo please link the PR/commit that adds this statement. Do you have examples of adding new bearings or seals?
