### 2020-04-04 ross-rotordynamics/ross#492 @raphaeltimbo
Hi there @nnadeau !
I had accidentally removed the line when changing the explanation for modeling assumptions.
On #515 I have added the mentioned text and also examples on how to add new bearings.
Let me know if you have more comments.
