### 2020-04-04 ross-rotordynamics/ross#493 @raphaeltimbo
Link to the commit: 5c0845e2b82192fa1395a0df690073b4a3bc25e1
