### 2020-04-04 ross-rotordynamics/ross#495 @raphaeltimbo
> @raphaeltimbo :
>
> * Please link the PR/commit that adds these statements

Link to commit: def64b5462d227cc902cf671942101a760179270

> * `ROSS is the first software being developed using the open source concept`, how is that novel compared to `commercial finite element softwares that have a rotordynamic module[s]`?

As discussed on #494, we believe that having an open source alternative to commercial rotordynamics softwares presents benefits to research in this area.

> * While I personally love Python for its general ease of use, I'm not convinced by the statement `Developing the software using Python and its scientific ecosystem enables the user to also make use of this ecosystem, making it easier to run rotordynamics analysis`.

My point in this statement is that with some of these rotordynamic commercial softwares you are  'locked' on their environment, interacting with the software only through a GUI, which makes it harder (impossible sometimes) to automate analysis, extract results to use in other tools, present results in different ways, run analysis with different settings and so on.
When you have an open source software developed in Python it is much easier to do those tasks since you have easy access to scientific tools, plotting tools, etc., all within the same environment.

Please, let me know if this makes the point clearer and if you want we can add it to the paper.
