### 2020-04-05 ross-rotordynamics/ross#495 @nnadeau
@raphaeltimbo yes, definitely add that to the paper. `[...] 'locked' on their environment, interacting with the software only through a GUI, which makes it harder (impossible sometimes) to automate analysis, [...]` is the key point that satisfies both of my above concerns
