### 2020-04-05 ross-rotordynamics/ross#494 @nnadeau
Same comment as https://github.com/ross-rotordynamics/ross/issues/495#issuecomment-609410632
