### 2020-04-05 ross-rotordynamics/ross#492 @nnadeau
@raphaeltimbo thanks for the examples, but I wouldn't necessary add the python code to the paper (JOSS likes short papers and the code makes it verbose).

I'd recommend removing the python examples but keeping your updated statements:

```
ROSS is extensible and new elements, such as different types of bearings or seals, can be added to the code. As anexample, one can add a class for a tapered roller bearing by inheriting from `BearingElement`:

Other elements that require more customization can be added by inheriting directly from `Element`, in this case it isnecessary to implement the required methods that should return the element's mass, stiffness, damping, and gyroscopicmatrices:
```

just reformulate the above statements to not target "in-paper examples", but rather the examples you've included in the repo
