### 2020-04-05 ross-rotordynamics/ross#490 @nnadeau
Fixed in https://github.com/ross-rotordynamics/ross/pull/504
