### 2020-04-05 ross-rotordynamics/ross#486 @nnadeau
fixed in https://github.com/ross-rotordynamics/ross/pull/504
