### 2020-04-05 ross-rotordynamics/ross#484 @nnadeau
@raphaeltimbo please link the PR/commit that resolves this
