### 2020-04-05 ross-rotordynamics/ross#484 @nnadeau
@raphaeltimbo from what I see (in https://github.com/ross-rotordynamics/ross/pull/504), all the citations are still there?
