### 2020-04-05 ross-rotordynamics/ross#480 @nnadeau
@raphaeltimbo the recommended citation is found here: https://www.scipy.org/citing.html
