### 2020-04-05 #2120 @nnadeau
@raphaeltimbo my checklist is complete and I'm just waiting to resolve the issues listed below. Once those are resolved, we're good to go! :)

https://github.com/ross-rotordynamics/ross/issues/492
https://github.com/ross-rotordynamics/ross/issues/494
https://github.com/ross-rotordynamics/ross/issues/495
https://github.com/ross-rotordynamics/ross/issues/484
https://github.com/ross-rotordynamics/ross/issues/480
https://github.com/ross-rotordynamics/ross/pull/516
https://github.com/ross-rotordynamics/ross/issues/482
