### 2020-04-05 ac9c755
Use recommended scipy citation
Related to #480.
