### 2020-04-05 ross-rotordynamics/ross#480 @raphaeltimbo
Thank you @nnadeau !
I've added the recommended citation on ac9c755.
