### 2020-04-05 ross-rotordynamics/ross#484 @raphaeltimbo
Hi there @nnadeau !
I have removed the citations on 3941ca78fada5882f195243de9241ecb34de1f68.
