### 2020-04-05 7298d94
Remove python code and add BallBearing as example
This remove the Python code from the paper (JOSS does not recommend)
and mentions BallBearingElement as an example already available in the
code.

Related to #492.
