### 2020-04-05 ross-rotordynamics/ross#492 @raphaeltimbo
I have removed the Python code from the paper and added a mention to the `BallBearingElement` that we have already implemented and is available in the repo (commit
7298d94).
