### 2020-04-05 565aee5
Point that differentiate from commercial softwares
Related to #494, #495
