### 2020-04-05 1ec02d4
Remove references not cited
Related to #484.
