+++ paper/paper.bib
@@ -80,44 +80,12 @@ year = {2019},
 url = {https://bokeh.org/},
 }
 
-@book{vance2010machinery,
-  title={Machinery vibration and rotordynamics},
-  author={Vance, John M and Zeidan, Fouad Y and Murphy, Brian G},
-  year={2010},
-  publisher={John Wiley \& Sons},
-  doi={10.1002/9780470903704},
-}
-
-@book{childs1993turbomachinery,
-  title={Turbomachinery rotordynamics: phenomena, modeling, and analysis},
-  author={Childs, Dara},
-  year={1993},
-  publisher={John Wiley \& Sons}
-}
-
-@book{ishida2012linear,
-  title={Linear and nonlinear rotordynamics},
-  author={Ishida, Yukio and Yamamoto, Toshio},
-  year={2012},
-  publisher={Wiley Online Library},
-  doi={10.1002/9783527651894},
-}
-
 @book{lalanne1998rotordynamics,
   title={Rotordynamics prediction in engineering},
   author={Lalanne, Michel and Ferraris, Guy},
   volume={2},
   year={1998},
   publisher={Wiley}
-}
-
-@book{gasch2006rotordynamik,
-  title={Rotordynamik},
-  author={Gasch, Robert and Nordmann, Rainer and Pf{\"u}tzner, Herbert},
-  year={2006},
-  publisher={Springer},
-  url={https://doi.org/10.1007/978-3-662-09786-1},
-  doi={10.1007/978-3-662-09786-1},
 }
 
 @misc{comsol,
