### 2020-04-05 ross-rotordynamics/ross#484 @raphaeltimbo
I have also removed the references from the bib file - 1ec02d4.
