### 2020-04-05 6e28f4d
Avoid repeating 'an example' and minor format
Related to #
