### 2020-04-05 ross-rotordynamics/ross#492 @raphaeltimbo
I have modified the text to avoid repeating 'an example' ( commit 6e28f4d9a6915d355583eca4b25825b210ec3d54).
The text now is:

> ROSS is extensible and new elements, such as different types of bearings or seals, can be added
> to the code. As an example, one can add a class for a tapered roller bearing by inheriting
> from BearingElement. The implementation of the BallBearingElement in our code uses
> this strategy.
> Other elements that require more customization can be added by inheriting directly from
> Element, in this case it is necessary to implement the required methods that should return
> the element’s mass, stiffness, damping, and gyroscopic matrices.
