### 2020-04-05 ross-rotordynamics/ross#484 @nnadeau
fixed in https://github.com/ross-rotordynamics/ross/commit/3941ca78fada5882f195243de9241ecb34de1f68 and https://github.com/ross-rotordynamics/ross/commit/1ec02d473049bd21bce884a9f736d73113a4fdc6
