### 2020-04-05 ross-rotordynamics/ross#495 @raphaeltimbo
@nnadeau , I have added that here: 565aee5
