### 2020-04-05 ross-rotordynamics/ross#480 @nnadeau
fixed in https://github.com/ross-rotordynamics/ross/commit/ac9c7557b5c3d8fcb1530bbf1e9fa23202c1ade0
