### 2020-04-05 ross-rotordynamics/ross#492 @nnadeau
@raphaeltimbo looks good. Nitpick from me: repeated use of `as an example` two sentences in a row

```
ROSS is extensible and new elements, such as different types of bearings or seals, can be added to the code. As an
example, one can add a class for a tapered roller bearing by inheriting from `BearingElement`. As an example we have
the implementation of the `BallBearingElement` in the code.
```
