### 2020-04-05 ross-rotordynamics/ross#492 @nnadeau
i see you caught that already in https://github.com/ross-rotordynamics/ross/commit/6e28f4d9a6915d355583eca4b25825b210ec3d54
