### 2020-04-05 ross-rotordynamics/ross#492 @nnadeau
:+1: fixed in https://github.com/ross-rotordynamics/ross/commit/6e28f4d9a6915d355583eca4b25825b210ec3d54 and https://github.com/ross-rotordynamics/ross/commit/7298d94214d16526fbb3d716ee79b2dc1a8e6bc4
