### 2020-04-05 ross-rotordynamics/ross#495 @nnadeau
`to automate analysis, To our knowledge, in`

typo: period instead of comma
