### 2020-04-05 ross-rotordynamics/ross#495 @raphaeltimbo
> `to automate analysis, To our knowledge, in`
>
> typo: period instead of comma

Fixed that on https://github.com/ross-rotordynamics/ross/pull/517/commits/75c93b42c24ae4eea27411349f5b517fb138b5c3
