### 2020-04-06 0efc8f8
Add test for error message in st_bearing_seal_elements
