### 2020-04-06 8dac97c
Add test for error message in st_materials
