### 2020-04-06 1703d30
Add test for methods from st_rotor_assembly
