+++ ross/tests/test_stochastic_rotor_assembly.py
@@ -3,11 +3,14 @@
 Tests for:
     st_rotor_assembly.py
 """
+import numpy as np
 import pytest
+from numpy.testing import assert_allclose
 
 from ross.bearing_seal_element import BearingElement
 from ross.disk_element import DiskElement
 from ross.materials import steel
+from ross.rotor_assembly import rotor_example
 from ross.shaft_element import ShaftElement
 from ross.stochastic.st_bearing_seal_element import ST_BearingElement
 from ross.stochastic.st_disk_element import ST_DiskElement
@@ -87,3 +90,351 @@ def test_st_point_mass_elements_odd_length():
     with pytest.raises(ValueError) as ex:
         ST_Rotor(shaft_elm, [disk0], [brg0, brg1, sup0, sup1], [pm0, pm1])
     assert "not all random point mass lists have same length." in str(ex.value)
+
+
+###############################################################################
+# testing methods
+###############################################################################
+
+
+@pytest.fixture
+def rotor1():
+    # rotor with 6 shaft elements, 2 disks and 2 random bearings
+    i_d = 0
+    o_d = 0.05
+    n = 6
+    L = [0.25 for _ in range(n)]
+
+    shaft_elem = [ShaftElement(l, i_d, o_d, material=steel) for l in L]
+
+    disk0 = DiskElement.from_geometry(
+        n=2, material=steel, width=0.07, i_d=0.05, o_d=0.28
+    )
+    disk1 = DiskElement.from_geometry(
+        n=4, material=steel, width=0.07, i_d=0.05, o_d=0.28
+    )
+
+    kxx = [1e6, 2e6]
+    cxx = [1e3, 2e3]
+    bearing0 = ST_BearingElement(n=0, kxx=kxx, cxx=cxx, is_random=["kxx", "cxx"])
+    bearing1 = ST_BearingElement(n=6, kxx=kxx, cxx=cxx, is_random=["kxx", "cxx"])
+
+    return ST_Rotor(shaft_elem, [disk0, disk1], [bearing0, bearing1])
+
+
+def test_run_campbell(rotor1):
+    speed_range = np.linspace(0, 300, 11)
+    results = rotor1.run_campbell(speed_range)
+
+    # fmt: off
+    wd = np.array([
+       [[96.37993790, 107.69984343], [96.31442399, 107.59665761],
+        [96.24864330, 107.49310785], [96.18259489, 107.38919370],
+        [96.11627780, 107.28491475], [96.04969111, 107.18027058],
+        [95.98283388, 107.07526081], [95.91570517, 106.96988508],
+        [95.84830408, 106.86414304], [95.78062967, 106.75803436],
+        [95.71268103, 106.65155874]],
+
+       [[96.37993790, 107.69984343], [96.44518597, 107.80266577],
+        [96.51016917, 107.90512511], [96.57488843, 108.00722194],
+        [96.63934474, 108.10895678], [96.70353904, 108.21033015],
+        [96.76747229, 108.31134261], [96.83114548, 108.41199472],
+        [96.89455957, 108.51228706], [96.95771553, 108.61222022],
+        [97.02061433, 108.71179481]],
+
+       [[298.26815963, 366.00322751], [297.24844811, 365.08745483],
+        [296.22737131, 364.16826537], [295.20500098, 363.24569809],
+        [294.18140949, 362.31979308], [293.15666973, 361.39059156],
[…]
