### 2020-04-08 ross-rotordynamics/ross#495 @raphaeltimbo
Hi there @nnadeau !
I forgot to tag you so I don't know if you received a notification for this.
Just wanted to let you know that I have fixed the typo.
