### 2020-04-09 ross-rotordynamics/ross#494 @nnadeau
fixed in https://github.com/ross-rotordynamics/ross/pull/517
