### 2020-04-09 #2120 @nnadeau
@arfon @meg-simula @raphaeltimbo we're good to go! good work!

@raphaeltimbo i patched the DOI issue here: https://github.com/ross-rotordynamics/ross/pull/518
