### 2020-04-09 9a16be0
In st_bearing_seal_element, add support to fluid_flow
