+++ ross/stochastic/st_bearing_seal_element.py
@@ -1,4 +1,16 @@
+"""Bearing element module for STOCHASTIC ROSS.
+
+This module creates an instance of random bearing for stochastic analysis.
+"""
+# fmt: off
+import numpy as np
+
 from ross.bearing_seal_element import BearingElement
+from ross.fluid_flow import fluid_flow as flow
+from ross.fluid_flow.fluid_flow_coefficients import (
+    calculate_damping_matrix, calculate_stiffness_matrix)
+
+# fmt: on
 
 __all__ = ["ST_BearingElement"]
 
@@ -10,7 +22,7 @@ class ST_BearingElement:
     BearingElement.
 
     Considering constant coefficients, use an 1-D array to make it random.
-    Considering varying coefficients to the frequency, use and 2-D array to
+    Considering varying coefficients to the frequency, use a 2-D array to
     make it random (*see the Examples below).
 
     Parameters
@@ -200,3 +212,193 @@ class ST_BearingElement:
         f_list = (BearingElement(*arg) for arg in new_args)
 
         return f_list
+
+    @classmethod
+    def from_fluid_flow(
+        cls,
+        n,
+        nz,
+        ntheta,
+        nradius,
+        length,
+        omega,
+        p_in,
+        p_out,
+        radius_rotor,
+        radius_stator,
+        visc,
+        rho,
+        eccentricity=None,
+        load=None,
+        tag=None,
+        n_link=None,
+        scale_factor=1,
+        is_random=None,
+    ):
+        """Instantiate a bearing using inputs from its fluid flow.
+
+        Parameters
+        ----------
+        n : int
+            The node in which the bearing will be located in the rotor.
+        is_random : list
+            List of the object attributes to become random.
+            Possibilities:
+                ["length", "omega", "p_in", "p_out", "radius_rotor",
+                 "radius_stator", "visc", "rho", "eccentricity", "load"]
+        tag: str, optional
+            A tag to name the element
+            Default is None.
+        n_link: int, optional
+            Node to which the bearing will connect. If None the bearing is
+            connected to ground.
+            Default is None.
+        scale_factor: float, optional
+            The scale factor is used to scale the bearing drawing.
+            Default is 1.
+
+        Grid related
+        ^^^^^^^^^^^^
+        Describes the discretization of the problem
+        nz: int
+            Number of points along the Z direction (direction of flow).
+        ntheta: int
+            Number of points along the direction theta. NOTE: ntheta must be odd.
+        nradius: int
+            Number of points along the direction r.
+        length: float, list
+            Length in the Z direction (m).
+            Input a list to make it random.
+
+        Operation conditions
+        ^^^^^^^^^^^^^^^^^^^^
+        Describes the operation conditions.
+        omega: float, list
+            Rotation of the rotor (rad/s).
+            Input a list to make it random.
+        p_in: float, list
+            Input Pressure (Pa).
+            Input a list to make it random.
+        p_out: float, list
+            Output Pressure (Pa).
+            Input a list to make it random.
+        load: float, list
+            Load applied to the rotor (N).
+            Input a list to make it random.
+
+        Geometric data of the problem
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+        Describes the geometric data of the problem.
+        radius_rotor: float, list
+            Rotor radius (m).
+            Input a list to make it random.
+        radius_stator: float, list
+            Stator Radius (m).
+            Input a list to make it random.
+        eccentricity: float, list
[…]
