### 2020-04-09 #2120 @raphaeltimbo
> **Submitting author:** @raphaeltimbo ([Raphael Silva](http://orcid.org/0000-0001-7689-5486))
> **Repository:** https://github.com/ross-rotordynamics/ross
> **Version:** v0.2.0
> **Editor:** @meg-simula
> **Reviewer:** @srmnitc, @nnadeau
> **Archive:** Pending

@meg-simula, with the review process we had to make some modifications to the package and the final version now is v0.3.2. Can we update that information?
