### 2020-04-09 571a6fd
Minor readability improvements
