+++ paper/paper.md
@@ -1,9 +1,8 @@
 ---
 title: 'ROSS - Rotordynamic Open Source Software'
-
 tags:
   - rotordynamic
-  - dynamics 
+  - dynamics
   - Python
 authors:
  - name: Raphael Timbó
@@ -44,89 +43,75 @@ bibliography: paper.bib
 
 # Summary
 
-There are several critical rotating equipment crucial to the industry, such as compressors,
-pumps and turbines.
-Computational mechanical models aim to simulate the behavior of such mechanical
-systems. These models are used to support research and decision making. To this purpose, we present ROSS,
-an open source library written in Python for rotordynamic analysis.
+There are several categories of critical rotating equipments crucial to industry, such as compressors, pumps, and turbines. Computational mechanical models aim to simulate the behavior of such mechanical systems and these models are used to support research and decision making. To this purpose, we present ROSS, an open source library written in Python for rotordynamic analysis.
 
-Concerning rotordynamics softwares, there are some commercial finite element softwares that have a rotordynamic
-module [@comsol; @ansys], some softwares based on a proprietary commercial software (MATLAB) [@madyn; @dynamicsrotating],
-and others developed as standalone softwares [@rotorinsa; @trcsoftware].
-To use these softwares, one needs to buy licenses, and they are not intended to be developed in a collaborative public
-manner. Also, for some of these commercial softwares, the user is 'locked' on their environment, interacting with the 
-software only through a GUI, which makes it harder (impossible sometimes) to automate analysis. To our knowledge, in 
-the rotordynamic field, ROSS is the first software being developed using the open source concept, with the code being 
-fully available on code hosting platforms, issues tracked online, clear license and possibility of direct contribution.
+Existing tools that have rotordynamic functionalities include commercial finite element software with rotordynamic modules [@comsol; @ansys], packages based on proprietary runtimes (MATLAB) [@madyn; @dynamicsrotating], and some standalone tools [@rotorinsa; @trcsoftware]. To use all of these options however requires the purchase of a license and they are not intended to be developed in an open, collaborative manner. Additionally for some of these commercial packages, the user is 'locked in' to the environment, interacting with the software only through a graphical user interface, which makes it harder (impossible sometimes) to automate analysis.
 
-ROSS allows the construction of rotor models and their numerical simulation. Shaft elements, as a default, are
-modeled with the Timoshenko beam theory [@Hutchinson2001], which considers shear and rotary inertia effects, and discretized by means of
-the Finite Element Method [@friswell2010dynamics]. Disks (impellers, blades, or other equipment attached to the rotor) 
-are assumed to be rigid bodies, thus their strain energy is not taken into account and only the kinetic energy due 
-translation and rotation is calculated. We can obtain the mass and gyroscopic matrices by applying Lagrange’s equations 
-to the total kinetic energy. 
+To our knowledge, ROSS is the first software being developed using the open source paradigm in rotordynamic field, with the code being clearly licensed and fully available on code hosting platforms, issues tracked online, and the possibility of direct contribution by the community.
 
-The mass matrix is given by: 
[…]
