### 2020-04-09 #2120 @arfon
:wave: @raphaeltimbo - I'll take over this paper from here as I'm not sure about @meg-simula's availability given the COVID-19 stuff going on.

I've read your paper and have a few suggested changes to improve the readability in https://github.com/ross-rotordynamics/ross/pull/519 .
