### 2020-04-09 fc2e454
Replace PyFormat styles in st_results.py
It replaces the old format style: `'%s %s' % ('one', 'two')` for the new one: `'{} {}'.format('one', 'two')`.
**See the following [link](https://pyformat.info/)
