+++ ross/stochastic/st_results.py
@@ -1,3 +1,7 @@
+"""STOCHASTIC ROSS plotting module.
+
+This module returns graphs for each type of analyses in st_rotor_assembly.py.
+"""
 import bokeh.palettes as bp
 import numpy as np
 from bokeh.layouts import gridplot
@@ -90,7 +94,7 @@ class ST_CampbellResults:
                 line_width=3.0,
                 muted_color=colors1[j],
                 muted_alpha=0.1,
-                legend_label="Mean - Mode %s" % (j + 1),
+                legend_label="Mean - Mode {}".format(j + 1),
             )
             if len(percentile):
                 for i, p in enumerate(percentile):
@@ -102,7 +106,7 @@ class ST_CampbellResults:
                         line_width=2.5,
                         muted_color=colors2[i],
                         muted_alpha=0.1,
-                        legend_label="percentile %s" % (p),
+                        legend_label="percentile: {}%".format(p),
                     )
             if len(conf_interval):
                 for i, p in enumerate(conf_interval):
@@ -114,7 +118,9 @@ class ST_CampbellResults:
                         line_width=2.5,
                         muted_color=colors1[j],
                         muted_alpha=0.1,
-                        legend_label="Interval %s - Mode %s" % (p, j + 1),
+                        legend_label="confidence interval: {}% - Mode {}".format(
+                            p, j + 1
+                        ),
                     )
                     fig.line(
                         x=self.speed_range,
@@ -124,7 +130,9 @@ class ST_CampbellResults:
                         line_width=2.5,
                         muted_color=colors1[j],
                         muted_alpha=0.1,
-                        legend_label="Interval %s - Mode %s" % (p, j + 1),
+                        legend_label="confidence interval: {}% - Mode {}".format(
+                            p, j + 1
+                        ),
                     )
 
         fig.legend.background_fill_alpha = 0.1
@@ -185,7 +193,7 @@ class ST_CampbellResults:
                 line_width=3.0,
                 muted_color=colors1[j],
                 muted_alpha=0.1,
-                legend_label="Mean - Mode %s" % (j + 1),
+                legend_label="Mean - Mode {}".format(j + 1),
             )
             if len(percentile):
                 for i, p in enumerate(percentile):
@@ -197,7 +205,7 @@ class ST_CampbellResults:
                         line_width=2.5,
                         muted_color=colors2[i],
                         muted_alpha=0.1,
-                        legend_label="percentile %s" % (p),
+                        legend_label="percentile {}".format(p),
                     )
             if len(conf_interval):
                 for i, p in enumerate(conf_interval):
@@ -209,7 +217,9 @@ class ST_CampbellResults:
                         line_width=2.5,
                         muted_color=colors1[j],
                         muted_alpha=0.1,
-                        legend_label="Interval %s - Mode %s" % (p, j + 1),
+                        legend_label="confidence interval: {}% - Mode {}".format(
+                            p, j + 1
+                        ),
                     )
                     fig.line(
                         x=self.speed_range,
@@ -219,7 +229,9 @@ class ST_CampbellResults:
                         line_width=2.5,
                         muted_color=colors1[j],
                         muted_alpha=0.1,
-                        legend_label="Interval %s - Mode %s" % (p, j + 1),
+                        legend_label="confidence interval: {}% - Mode {}".format(
+                            p, j + 1
+                        ),
                     )
 
         fig.legend.background_fill_alpha = 0.01
@@ -248,11 +260,12 @@ class ST_CampbellResults:
         kwargs : optional
             Additional key word arguments can be passed to change
             the plot (e.g. linestyle='--').
+
[…]
