### 2020-04-09 993f2be
Commit to debug Travis
* Some minor changes in docstring 
* Trying to remove duplicated build from Travis.
